% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2_qfeatures.R
\name{addOriginOfValues}
\alias{addOriginOfValues}
\title{Sets the MEC tag in the OriginOfValues}
\usage{
addOriginOfValues(obj, i, namesOrigin = NULL)
}
\arguments{
\item{obj}{An object of class \code{QFeatures}}

\item{i}{The indice of the assay in obj to be used to build the Origin of values. Generally, it is the first one}

\item{namesOrigin}{A list of names which are known to contain method of identification of peptides.Thoses names must figure in
the colnames of the rowData of an assay}
}
\value{
An instance of class \code{DataFrame} Which has the same dimensions as the quantitaive datas (assay) ans contains the origin
of values with tags such as MEC, POV
}
\description{
Sets the MEC tag in the origin parameter
}
\examples{
\dontrun{
utils::data(Exp1_R25_pept, package='DAPARdata2')
origin <- c("Identification_type_C_R1", "Identification_type_C_R2", 
"Identification_type_C_R3", "Identification_type_D_R1", 
"Identification_type_D_R2", "Identification_type_D_R3")
obj <- addOriginOfValues(Exp1_R25_pept, 2, origin)
}

}
\author{
Samuel Wieczorek
}
