% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filterAdjmat.R
\name{specPeptides}
\alias{specPeptides}
\alias{sharedPeptides}
\alias{topnFunctions}
\alias{topnPeptides}
\title{xxx}
\usage{
specPeptides(object, ...)

sharedPeptides(object, ...)

topnFunctions()

topnPeptides(object, ...)
}
\arguments{
\item{object}{xxx}

\item{...}{A `list()` of additional parameters}
}
\description{
xxx
}
\details{
This function builds an intermediate matrix with scores for each peptide
based on 'fun' parameter. Once this matrix is built, one select the 'n' peptides
which have the higher score

- rowMedians xxx
- rowMeans xxx
- rowSums xxx
}
\examples{

}
