% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggIter}
\alias{aggIter}
\title{Method to aggregate peptides into proteins with the iterative approach.}
\usage{
aggIter(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)
}
\arguments{
\item{qPepData}{A matrix of intensities of peptides.}

\item{X}{An adjacency matrix}

\item{conditions}{xxx}

\item{init.method}{The method used to initialize the iterative algotirhm. Default is 'Sum'.}

\item{method}{The method used for xxx. Default value is 'Mean'.}

\item{n}{xxxx}
}
\value{
A matrix of protein intensities.
}
\description{
Method to aggregate peptides into proteins with the iterative approach
}
\author{
Samuel Wieczorek
}
