% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggIterParallel}
\alias{aggIterParallel}
\title{Method to aggregate peptides into proteins with the iterative approach with use of parallelism.}
\usage{
aggIterParallel(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)
}
\arguments{
\item{qPepData}{A matrix of intensities of peptides.}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{conditions}{xxxx}

\item{init.method}{The method used to initialize the iterative algotirhm. Default is 'Sum'.}

\item{method}{The method used for xxx. Default value is 'Mean'.}

\item{n}{xxxx}
}
\value{
A matrix of intensities of proteins
}
\description{
Method to aggregate peptides into proteins with the iterative approach with use of parallelism.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original_log']][1:1000,])[,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original_log']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
conditions <- colData(Exp1_R25_pept)@listData$Condition
aggIterParallel(assay(Exp1_R25_pept,2)[1:1000,], X.list$all, conditions)

}
\author{
Samuel Wieczorek
}
