% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggMean}
\alias{aggMean}
\title{Compute the intensity of proteins as the mean of the intensities
of their peptides.}
\usage{
aggMean(qPepData, X)
}
\arguments{
\item{qPepData}{A matrix of intensities of peptides.}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins as the mean of the 
intensities of their peptides.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]])$all)
aggMean(assay(obj[[2]]), X)

}
\author{
Alexia Dorffer
}
