% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_with_matAdj}
\alias{aggregate_with_matAdj}
\title{Aggreagate quantitative features.}
\usage{
aggregate_with_matAdj(qPepData, X, FUN, ...)
}
\arguments{
\item{qPepData}{xxxx.}

\item{X}{An adjacency matrix.}

\item{FUN}{A `function` to be applied to the subsets of `x`.}

\item{...}{Additional arguments passed to `FUN`.}
}
\value{
A new `matrix` of dimensions xxxx
}
\description{
This function takes a matrix of quantitative features `x` and a
factor (of length equal to `nrow(x)`) defining subsets, and
applies a user-defined function to aggregate each subset into a
vector of quantitative values. This function is the same as 'aggregate_by_vector' from the package QFeatures
and it is used with DAPAR aggregation functions

User-defined functions must thus return a matrix of dimensions equal
to `X`. Examples thereof are

- [Dapar2:aggSum()] to use the sum of each column (default);

- [Dapar2:aggMean()] to use the sum of each column;

- [Dapar2:aggIter()] to use the mean of each column;

- [Dapar2:aggIterParallel()] same as previous function but use parallelism.

- [Dapar2::aggTopn] to use the sum of each column;
}
\author{
Samuel Wieczorek
}
\concept{Quantitative feature aggregation}
