% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggSum}
\alias{aggSum}
\title{Compute the intensity of proteins with the sum of the intensities
of their peptides.}
\usage{
aggSum(qPepData, X)
}
\arguments{
\item{qPepData}{A matrix of intensities of peptides.}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}
}
\value{
A matrix of intensities of proteins.
}
\description{
This function computes the intensity of proteins based on the sum of the 
intensities of their peptides.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original']][1:1000,])[1:1000,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
aggSum(assay(Exp1_R25_pept[['original']])[1:1000,], X$all)

}
\author{
Alexia Dorffer, Samuel Wieczorek
}
