% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggTopn}
\alias{aggTopn}
\title{Compute the intensity of proteins as the sum of the intensities of their n best peptides.}
\usage{
aggTopn(qPepData, X, method = "Mean", n = 10)
}
\arguments{
\item{qPepData}{A matrix of intensities of peptides}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{method}{Default is 'Mean'}

\item{n}{The maximum number of peptides used to aggregate a protein.}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins as the sum of the 
intensities of their n best peptides.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original_log']][1:1000,])[,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original_log']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
aggTopn(assay(Exp1_R25_pept,2)[1:1000,], X.list$all, n=3)

}
\author{
Alexia Dorffer, Samuel Wieczorek
}
