% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_all_plots.R
\name{all_plots}
\alias{all_plots}
\alias{mod_all_plots_ui}
\alias{mod_all_plots_server}
\title{xxx}
\usage{
mod_all_plots_ui(id)

mod_all_plots_server(id, dataIn)
}
\arguments{
\item{id}{xxxx}

\item{dataIn}{xxxx}
}
\description{
xxxxxx
}
