% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioAnalysis.R
\name{barplotEnrichGO_HC}
\alias{barplotEnrichGO_HC}
\title{A barplot that shows the result of a GO enrichment, using the package \code{highcharter}}
\usage{
barplotEnrichGO_HC(ego, maxRes = 5, title = NULL)
}
\arguments{
\item{ego}{The result of the GO enrichment, provides either by the function
\code{enrichGO} in the package \code{DAPAR} or the function \code{enrichGO} 
of the package \code{\link{clusterProfiler}}}

\item{maxRes}{The maximum number of categories to display in the plot}

\item{title}{The title of the plot}
}
\value{
A barplot
}
\description{
A barplot of GO enrichment analysis
}
\examples{
\donttest{
library(clusterProfiler)
library(highcharter)
library(DAPAR2)
utils::data(Exp1_R25_prot, package='DAPARdata2')
data <- rowData(Exp1_R25_prot[[2]])[['Protein_IDs']]
univ <- univ_AnnotDbPkg("org.Sc.sgd.db")
ego <- enrich_GO(data, idFrom="UNIPROT", orgdb="org.Sc.sgd.db", ont="MF", pval=0.05, universe = univ)
barplotEnrichGO_HC(ego)
}
}
\author{
Samuel Wieczorek, Enora Fremy
}
