% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{Base_Palette}
\alias{Base_Palette}
\title{xxx}
\usage{
Base_Palette(palette.name, conditions)
}
\arguments{
\item{palette.name}{The name of the palette from the package `RColorBrewer`}

\item{conditions}{xxxx}
}
\description{
Builds a vector of colors (one per condition) to be used as a basis palette
}
