% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_buildDesign.R
\name{build-design}
\alias{build-design}
\alias{mod_buildDesign_ui}
\alias{mod_buildDesign_server}
\title{xxxx}
\usage{
mod_buildDesign_ui(id)

mod_buildDesign_server(id, quantCols)
}
\arguments{
\item{id}{xxx}

\item{quantCols}{xxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{
library(shiny)
library(highcharter)

################################################################
ui <- tagList(
  mod_buildDesign_ui("designEx"),
  
)

server <- function(input, output, session) {
  design <- mod_buildDesign_server("designEx", c('a', 'a', 'a', 'b', 'b', 'b'))
  
  observeEvent(design()$trigger, {
    print(design()$design)
  })
}
shinyApp(ui = ui, server = server)


}
