% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{buildGraph}
\alias{buildGraph}
\title{Build graph definition}
\usage{
buildGraph(The.CC, X)
}
\arguments{
\item{The.CC}{A cc (a list)}

\item{X}{An adjacency matrix as given by the function BuildListAdjacencyMatrices}
}
\value{
A list of two itens: nodes and edges
}
\description{
Builds a graph as defined by its nodes and edges.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2') 
plist <- rowData(Exp1_R25_pept[['original_log']][1:1000,])[,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original_log']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
ll <- get.pep.prot.cc(X.list$all)
g <- buildGraph(ll[[1]], X.list$all)

}
\author{
Thomas Burger, Samuel Wieczorek
}
