% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{BuildMetacell}
\alias{BuildMetacell}
\title{metacell function which xxx}
\usage{
BuildMetacell(from = NULL, level, qdata = NULL, conds = NULL, df = NULL)
}
\arguments{
\item{from}{xxx}

\item{level}{xxx}

\item{qdata}{A matrix of quantitative data}

\item{conds}{xxx}

\item{df}{A data.frame which contains the type of identification of the
entities. It must have the same dimensions as \code{qData}.}
}
\value{
xxxxx
}
\description{
xxx
}
\examples{
file <- system.file("extdata", "Exp1_R25_pept.txt", package="DaparToolshedData")
data <- read.table(file, header=TRUE, sep="\t",stringsAsFactors = FALSE)
metadataFile <- system.file("extdata", "samples_Exp1_R25.txt",
package="DaparToolshedData")
metadata <- read.table(metadataFile, header=TRUE, sep="\t", as.is=TRUE,
stringsAsFactors = FALSE)
conds <- metadata$Condition
qdata <- data[,56:61]
df <- data[ , 43:48]
df <- BuildMetacell('maxquant', 'peptide', qdata, conds, df)
df <- BuildMetacell('proline', 'peptide', qdata, conds, df)
}
\author{
Samuel Wieczorek
}
