% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\name{BuildqMetadata}
\alias{BuildqMetadata}
\title{xxxx}
\usage{
BuildqMetadata(from = NULL, level, qdata = NULL, conds = NULL, df = NULL)
}
\arguments{
\item{from}{xxx}

\item{level}{xxx}

\item{qdata}{A matrix of quantitative data}

\item{conds}{xxx}

\item{df}{A list of integer xxxxxxx}
}
\value{
xxxxx
}
\description{
xxxxxx
}
\examples{
file <- system.file("extdata", "Exp1_R25_pept.txt", package="DaparToolshedData")
data <- read.table(file, header=TRUE, sep="\t",stringsAsFactors = FALSE)
metadataFile <- system.file("extdata", "samples_Exp1_R25.txt", 
package="DaparToolshedData")
metadata <- read.table(metadataFile, header=TRUE, sep="\t", as.is=TRUE, 
stringsAsFactors = FALSE)
conds <- metadata$Condition
qdata <- data[,56:61]
df <- data[ , 43:48]
df <- BuildqMetadata('maxquant', 'peptide', qdata, conds, df)
df <- BuildqMetadata('proline', 'peptide', qdata, conds, df)

}
\author{
Samuel Wieczorek
}
