% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{check.conditions}
\alias{check.conditions}
\title{Check if the design of conditions is valid}
\usage{
check.conditions(conds)
}
\arguments{
\item{conds}{A vector of conditions of the samples.}
}
\value{
A list
}
\description{
Check if the design of conditions is valid
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
check.conditions(colData(Exp1_R25_pept)$Condition)

}
\author{
Samuel Wieczorek
}
