% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{checkClusterability}
\alias{checkClusterability}
\title{Prepare the data for subsequent clustering}
\usage{
checkClusterability(standards)
}
\arguments{
\item{standards}{a dataframe containing the standardized mean intensities}
}
\value{
a list of 2 elements:
\itemize{
\item dip_test: the result of the clusterability of the data
\item gap_cluster: the gap statistic obtained with the function \code{\link[cluster:clusGap]{cluster::clusGap()}}.
}
}
\description{
The first step is to standardize the data (with the \code{Mfuzz}
package). Then the function checks that these data are clusterizable or not
(use of \code{\link[diptest:dip.test]{diptest::dip.test()}} to determine whether the distribution is unimodal or
multimodal). Finally, it determines the "optimal" k by the Gap statistic
approach.
}
\examples{
checkClusterability(standardized_means)

}
\author{
Hélène Borges
}
