% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{CheckDesign}
\alias{CheckDesign}
\title{Check the validity of the experimental design}
\usage{
CheckDesign(sampleTab)
}
\arguments{
\item{sampleTab}{xxxx}
}
\value{
A boolean
}
\description{
Check the validity of the experimental design
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
CheckDesign(colData(Exp1_R25_pept))

}
\author{
Thomas Burger, Samuel Wieczorek
}
