% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_compare_normalization_plots.R
\name{compareNormalizationD_HC}
\alias{compareNormalizationD_HC}
\title{Builds a plot from a numeric matrix.}
\usage{
compareNormalizationD_HC(
  qDataBefore,
  qDataAfter,
  condsForLegend = NULL,
  indData2Show = NULL,
  palette = NULL
)
}
\arguments{
\item{qDataBefore}{A numeric matrix that contains quantitative data before normalization.}

\item{qDataAfter}{A numeric matrix that contains quantitative data after normalization.}

\item{condsForLegend}{A vector of the conditions (one condition per sample).}

\item{indData2Show}{A vector of the indices of the columns to show in 
the plot. The indices are those of indices of 
the columns int the data.frame qDataBefore.}

\item{palette}{xxx}
}
\value{
A plot
}
\description{
Builds a plot from a numeric matrix.
}
\examples{
library(highcharter)
library(Features)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
conds <- colData(obj)[["Condition"]]
objBefore <- obj[['original_log']]
qDataBefore <- assay(objBefore)
objAfter <- normalizeD(object[['original_log']], method='SumByColumns', conds=conds, type='overall')
qDataAfter <- assay(objAfter)
compareNormalizationD_HC(qDataBefore, qDataAfter, condsForLegend=conds)

}
\author{
Samuel Wieczorek, Enora Fremy
}
