% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ttest.R
\name{compute.group.t.tests}
\alias{compute.group.t.tests}
\title{xxxxxx}
\usage{
compute.group.t.tests(
  qData,
  sampleTab,
  X,
  X.spec,
  logFC = NULL,
  contrast = "OnevsOne",
  type = "Student"
)
}
\arguments{
\item{qData}{A matrix of quantitative data, without any missing values.}

\item{sampleTab}{xxxx}

\item{X}{An adjacency matrix which contains all peptides (specific and shared).}

\item{X.spec}{An adjacency matrix which contains only specific peptides.}

\item{logFC}{A vector (or list of vectors) xxxx}

\item{contrast}{Indicates if the test consists of the comparison of each 
biological condition versus 
each of the other ones (contrast=1; 
for example H0:"C1=C2" vs H1:"C1!=C2", etc.) 
or each condition versus all others (contrast=2; e.g.  H0:"C1=(C2+C3)/2" vs
H1:"C1!=(C2+C3)/2", etc. if there are three conditions).}

\item{type}{xxxxx}
}
\value{
A DataFrame which contains the
}
\description{
xxxxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata')
obj <- Exp1_R25_pept[1:1000,]
obj <- addAssay(obj, Features::filterNA(obj[[2]],  pNA = 0), name='filtered')
obj <- addListAdjacencyMatrices(obj, 3)
qData <- assay(obj[['filtered']])
X.all <- GetAdjMat(obj, 3, 'all')
X.onlySpec <- GetAdjMat(obj, 3, 'onlySpec')
sTab <- colData(obj)
gttest <- compute.group.t.tests(qData, sTab, X.all, X.onlySpec)

}
\author{
Thomas Burger, Samuel Wieczorek
}
