% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.t.tests.R
\name{compute_t_tests}
\alias{compute_t_tests}
\title{xxxxxx}
\usage{
compute_t_tests(obj, sTab, contrast = "OnevsOne", type = "Student")
}
\arguments{
\item{obj}{xxxxx}

\item{sTab}{A data.frame which contains the samples data.}

\item{contrast}{Indicates if the test consists of the comparison of each 
biological condition versus 
each of the other ones (contrast=1; 
for example H0:"C1=C2" vs H1:"C1!=C2", etc.) 
or each condition versus all others (contrast=2; e.g.  H0:"C1=(C2+C3)/2" vs
H1:"C1!=(C2+C3)/2", etc. if there are three conditions).}

\item{type}{A string which is the type of statistical test to proceed. Available values are 'Student' and 'Welch'.
Default value is 'Student'.}
}
\value{
A list of two items : logFC and P_Value; both are dataframe. The first one contains
the logFC values of all the comparisons (one column for one comparison), the second one contains
the pvalue of all the comparisons (one column for one comparison). The names of the columns for those two dataframes
are identical and correspond to the description of the comparison.
}
\description{
xxx
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
Exp1_R25_pept <- addAssay(Exp1_R25_pept, QFeatures::filterNA(Exp1_R25_pept[[2]], pNA = 0), name='filtered')
sTab <- as.data.frame(colData(Exp1_R25_pept))
ttest <- compute_t_tests(Exp1_R25_pept[['filtered']], sTab ,"OnevsOne")

}
\author{
Florence Combes, Samuel Wieczorek
}
