% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{ComputeAdjacencyMatrices}
\alias{ComputeAdjacencyMatrices}
\title{Function matrix of appartenance group}
\usage{
ComputeAdjacencyMatrices(obj.se)
}
\arguments{
\item{obj}{An object of class 'QFeatures'}

\item{i}{The indice of the dataset (class 'SumarizedExperiment') in the list of 'obj' on which to apply the aggregation.}
}
\value{
AN object of class 'QFeatures'
}
\description{
All the matrices have the same dimensions. The first matrix correspond to all the peptides, the second one contains 
only the peptides shared with several proteins and the last matrix contains only the specific peptides. These matrices are useful 
is one aggregate with only a certain type of peptides or all of them. The list of matrices is stored in the slot 'xxx' of the metadata of the argument
obj (class 'SummarizedExperiment')
}
\details{
Method to create a list of three binary matrices with proteins in columns and peptides in lines.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
ll.x <- ComputeAdjacencyMatrices(obj[[2]])

}
\author{
Samuel Wieczorek
}
