% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectedComps.R
\name{ComputeConnectedComposants}
\alias{ComputeConnectedComposants}
\title{Computes the connex components of a graph, represented by an adjacency matrix.}
\usage{
ComputeConnectedComposants(obj.se)
}
\arguments{
\item{obj.se}{xxxx}
}
\value{
A list of items containing the connex components derived from the matrices given in arguments
}
\description{
This function computes the connex components of a graph, represented by an adjacency matrix.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
cc <- ComputeConnectedComposants(Exp1_R25_pept[[2]])

}
\author{
Samuel Wieczorek
}
