% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectedComps.R, R/mod_graph_pept_prot.R,
%   R/mod_plots_cc.R
\name{ComputeConnectedComposants}
\alias{ComputeConnectedComposants}
\alias{get.pep.prot.cc}
\alias{buildGraph}
\alias{display.CC.visNet}
\alias{plotJitterCC}
\alias{plotJitter_hc}
\alias{mod_graph_pept_prot_ui}
\alias{mod_graph_pept_prot_server}
\alias{mod_plots_cc_ui}
\alias{mod_plots_cc_server}
\title{Computes the connex components of a graph, represented by an adjacency matrix.}
\usage{
ComputeConnectedComposants(obj.se)

get.pep.prot.cc(X)

buildGraph(The.CC, X)

display.CC.visNet(g, obj = NULL, prot.tooltip = NULL, pept.tooltip = NULL)

plotJitterCC(list.of.cc)

plotJitter_hc(df, clickFunction = NULL)

mod_graph_pept_prot_ui(id)

mod_graph_pept_prot_server(id, cc, matAdj, dataIn)

mod_plots_cc_ui(id)

mod_plots_cc_server(input, output, session, cc, matAdj, obj)
}
\arguments{
\item{obj.se}{xxxx}

\item{X}{An adjacency matrix as given by the function BuildListAdjacencyMatrices}

\item{The.CC}{A cc (a list)}

\item{g}{A connex component (a list)}

\item{obj}{xxx}

\item{prot.tooltip}{A vector of strings which are the text for the tooltips of the proteins.}

\item{pept.tooltip}{A vector of strings which are the text for the tooltips of the peptides.}

\item{list.of.cc}{List of cc such as returned by the function get.pep.prot.cc}

\item{df}{A dataframe containing xxx}

\item{clickFunction}{A JavaScript code to show an information when a point is clicked}

\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{cc}{xxx}

\item{matAdj}{xxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{id}{shiny id}
}
\value{
A list of items containing the connex components derived from the matrices given in arguments

A list of Connected Components (CC)

A list of two itens: nodes and edges

A plot

A plot

A plot

NA

NA

NA
}
\description{
This function computes the connex components of a graph, represented by an adjacency matrix.

Method to build the list of connected components of an adjacency matrix

Builds a graph as defined by its nodes and edges.

Displays a connected component as a graph with the package 'visNetwork'.

Jitter plot of CC

Display a jitter plot for connex component with the package highcharter

A shiny Module.

A shiny Module.
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
cc <- ComputeConnectedComposants(Exp1_R25_pept[[2]])

Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData")) 
X <- GetAdjMat(Exp1_R25_pept[[2]])$all
ll <- get.pep.prot.cc(X)

Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData")) 
X <- GetAdjMat(Exp1_R25_pept[[2]])$all
ll <- get.pep.prot.cc(X)
g <- buildGraph(ll[[1]], X)

Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData")) 
X <- GetAdjMat(Exp1_R25_pept[[2]])$all
ll <- get.pep.prot.cc(X)
g <- buildGraph(ll[[1]], X)
display.CC.visNet(g)

Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData")) 
X <- GetAdjMat(Exp1_R25_pept[[2]])$all
ll <- get.pep.prot.cc(X)
plotJitterCC(ll)

df <- data.frame(x=seq_len(10), y=seq_len(10))
plotJitter_hc(df)


xxx

}
\author{
Samuel Wieczorek

Thomas Burger, Samuel Wieczorek

Thomas Burger, Samuel Wieczorek

Thomas Burger, Samuel Wieczorek

Thomas Burger

Thomas Burger, Samuel Wieczorek
}
\keyword{internal}
