% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConvertListToHtml}
\alias{ConvertListToHtml}
\title{Convert a list to unnumbered HTML list}
\usage{
ConvertListToHtml(ll)
}
\arguments{
\item{ll}{A \code{list()} of \code{character()}}
}
\description{
xxx
}
\examples{
ConvertListToHtml(list('foo1', 'foo2', 'foo3'))

}
