% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_corr_matrix.R, R/mod_plots_corr_matrix.R
\name{corrMatrixPlot}
\alias{corrMatrixPlot}
\alias{mod_corrmatrix_plot_ui}
\alias{mod_corrmatrix_plot_server}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
corrMatrixPlot(qData, rate = 0.5, showValues = FALSE)

mod_corrmatrix_plot_ui(id)

mod_corrmatrix_plot_server(
  id,
  qData,
  rate = reactive({     0.5 }),
  showValues = reactive({     FALSE })
)
}
\arguments{
\item{qData}{xxx}

\item{rate}{xxx. Default value is 0.9}

\item{showValues}{Default is FALSE.}

\item{id}{xxx}
}
\value{
A colored correlation matrix
}
\description{
Displays a correlation matrix of the quantitative data of a
numeric matrix.
}
\examples{
library(QFeatures)
data(ft)
corrMatrixPlot(ft[[1]])

}
\author{
Samuel Wieczorek, Enora Fremy
}
