% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_corr_matrix.R
\name{corrMatrixD_HC}
\alias{corrMatrixD_HC}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
corrMatrixD_HC(obj, names = NULL, rate = 0.5, showValues = TRUE)
}
\arguments{
\item{obj}{xxx}

\item{names}{xxxxx}

\item{rate}{The rate parameter to control the exponential law for 
the gradient of colors}

\item{showValues}{xxx}
}
\value{
A colored correlation matrix
}
\description{
Displays a correlation matrix of the quantitative data of a
numeric matrix.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[[2]]
corrMatrixD_HC(obj)

}
\author{
Samuel Wieczorek, Enora Fremy
}
