% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{createFeatures}
\alias{createFeatures}
\title{Creates an object of class \code{MSnSet} from text file}
\usage{
createFeatures(
  data,
  sample,
  indExpData,
  keyId = NULL,
  namesOrigin = NULL,
  logTransform = FALSE,
  forceNA = FALSE,
  typeOfData,
  parentProtId = NULL,
  analysis = "foo",
  processes = NULL,
  pipelineType = NULL
)
}
\arguments{
\item{data}{The name of a tab-separated file that contains the data.}

\item{sample}{A dataframe describing the samples (in lines).}

\item{indExpData}{A vector of string where each element is the name
of a column in designTable that have to be integrated in
the \code{fData()} table of the \code{MSnSet} object.}

\item{keyId}{The indice of the column containing the ID of entities 
(peptides or proteins)}

\item{namesOrigin}{xxxxxxxxxxx}

\item{logTransform}{A boolean value to indicate if the data have to be
log-transformed (Default is FALSE)}

\item{forceNA}{A boolean value to indicate if the 0 and NaN values of
intensity have to be replaced by NA (Default is FALSE)}

\item{typeOfData}{A string that indicates whether the dataset is about}

\item{parentProtId}{For peptide entities, a string which is the name of a column in rowData. It contains the id of parent
proteins and is used to generate adjacency matrix and process to aggregation.}

\item{analysis}{xxx}

\item{processes}{xxxx}

\item{pipelineType}{xxxx}
}
\value{
An instance of class \code{Features}.
}
\description{
Builds an object of class \code{Features} from a 
single tabulated-like file for quantitative and meta-data and a dataframe 
for the samples description.
}
\examples{
data.file <- system.file("extdata", "Exp1_R25_pept.txt", package="DAPARdata2")
data <- read.table(data.file, header=TRUE, sep="\t", as.is=TRUE, stringsAsFactors = FALSE)
sample.file <- system.file("extdata", "samples_Exp1_R25.txt", package="DAPARdata2")
sample <- read.table(sample.file, header=TRUE, sep="\t", as.is=TRUE, stringsAsFactors = FALSE)
indExpData <- 56:61
namesOrigin <- colnames(data)[43:48]
parentId <- 'Protein_group_IDs'
keyid <- 'Sequence'
analysis <- 'test'
ft <- createFeatures(data, sample,indExpData,keyId = keyid, analysis=analysis, namesOrigin = namesOrigin, typeOfData = "peptide", parentProtId = parentId, forceNA=TRUE)

}
\author{
Samuel Wieczorek
}
