% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2_qfeatures.R
\name{createQFeatures}
\alias{createQFeatures}
\title{Creates an object of class \code{QFeatures} from text file}
\usage{
createQFeatures(
  data,
  file,
  sample,
  indQData,
  keyId = "AutoID",
  indQMetadata = NULL,
  logTransform = FALSE,
  forceNA = TRUE,
  typeDataset,
  parentProtId = NULL,
  analysis = "foo",
  processes = NULL,
  typePipeline = NULL,
  software = NULL
)
}
\arguments{
\item{data}{The name of a tab-separated file that contains the data.}

\item{file}{The name of a file xxx}

\item{sample}{A dataframe describing the samples (in lines).}

\item{indQData}{A vector of string where each element is the name
of a column in designTable that have to be integrated in
the \code{rowData()} table of the \code{QFeatures} object.}

\item{keyId}{The indice of the column containing the ID of entities 
(peptides or proteins)}

\item{indQMetadata}{xxxxxxxxxxx}

\item{logTransform}{A boolean value to indicate if the data have to be
log-transformed (Default is FALSE)}

\item{forceNA}{A boolean value to indicate if the 0 and NaN values of
intensity have to be replaced by NA (Default is FALSE)}

\item{typeDataset}{A string that indicates whether the dataset is about}

\item{parentProtId}{For peptide entities, a string which is the name of a column in rowData. It contains the id of parent
proteins and is used to generate adjacency matrix and process to aggregation.}

\item{analysis}{The name of the MS analysis}

\item{processes}{xxxx}

\item{typePipeline}{The type of pipeline used with this dataset. The list of predefined
pipelines in DaparToolshed can be obtained with the function \code{pipelines()}. Default value is NULL}
}
\value{
An instance of class \code{QFeatures}.
}
\description{
Builds an object of class \code{QFeatures} from a 
single tabulated-like file for quantitative and meta-data and a dataframe 
for the samples description.
}
\author{
Samuel Wieczorek
}
