% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_CV_dist.R, R/mod_plots_var_dist.R
\name{CVDist}
\alias{CVDist}
\alias{cv_plots}
\alias{mod_cv_plot_ui}
\alias{mod_cv_plot_server}
\title{mod_var_dist_plot_ui and mod_var_dist_plot_server}
\usage{
CVDist(qData, conds = NULL, pal.name = NULL)

mod_cv_plot_ui(id)

mod_cv_plot_server(id, obj, conds, pal.name = NULL)
}
\arguments{
\item{conds}{A \code{character()} representings the condition for
each sample of the \link{QFeatures} object.}

\item{id}{shiny id}

\item{obj}{xxx}

\item{base_palette}{xxx}
}
\description{
A shiny Module.
}
\examples{

#------------------------------
# plot a single plot
#------------------------------

CVDist(qData, conds)

#------------------------------
# A shiny module
#------------------------------

if(interactive()){
library(QFeatures)
data(ft)
ui <- mod_cv_plot_ui('plot')

server <- function(input, output, session) {
 data(ft)
 conds <- design(ft)$Condition
 
 mod_cv_plot_server('plot',
                    obj = reactive({ft[[1]]}),
                    conds = reactive({conds}),
                    pal.name = reactive({'Dark2'})
                    )
 }
shinyApp(ui=ui, server=server)
}
}
