% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_CV_dist.R
\name{CVDistD_HC}
\alias{CVDistD_HC}
\title{Distribution of CV of entities}
\usage{
CVDistD_HC(qData, conds = NULL, palette = NULL)
}
\arguments{
\item{qData}{A numeric matrix of quantitative data.}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{palette}{A vector of HEX Code colors (one color per condition).}
}
\value{
A density plot
}
\description{
Builds a densityplot of the CV of entities in numeric matrix.
The CV is calculated for each condition present in the dataset
(see the slot \code{'Condition'} in the \code{colData()} DataFrame)
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[[2]])[1:10,]
conds <- colData(Exp1_R25_pept)[["Condition"]]
CVDistD_HC(qData, conds)

}
\author{
Samuel Wieczorek, Enora Fremy
}
