% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed.R
\docType{package}
\name{DaparToolshed}
\alias{DaparToolshed}
\title{DaparToolshed: A package for computating the notorious bar statistic}
\description{
The DaparToolshed package provides three categories of important functions:
foo, bar and baz.
}
\section{DaparToolshed functions}{

The foo functions ...
}

