% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_se_explorer.R
\name{data_explorer}
\alias{data_explorer}
\alias{mod_explorer_ui}
\alias{mod_explorer_server}
\title{Data explorer}
\usage{
mod_explorer_ui(id)

mod_explorer_server(id, object, digits = reactive({
    3
}))
}
\arguments{
\item{id}{xxx}

\item{object}{xxx}

\item{digits}{xxx}
}
\value{
NA
}
\description{
A shiny Module.
}
\examples{
data(ft_na)

ui <- tagList(
  mod_explorer_ui('explore')
)

server <- function(input, output, session) {
  mod_explorer_server(id = 'explore',
                      object = reactive({ft_na})
  )
}
shinyApp(ui=ui, server=server)
}
\author{
Samuel Wieczorek
}
