% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_density_plot.R, R/mod_plots_density.R
\name{densityPlot}
\alias{densityPlot}
\alias{density-plots}
\alias{mod_plots_density_ui}
\alias{mod_plots_density_server}
\title{Shiny module for plots.}
\usage{
densityPlot(qData, conds, legend = NULL, pal = NULL)

mod_plots_density_ui(id)

mod_plots_density_server(id, obj, conds, legend = NULL, base_palette = NULL)
}
\arguments{
\item{qData}{numeric matrix}

\item{conds}{xxx}

\item{legend}{xxx}

\item{id}{shiny id}

\item{obj}{xxx}

\item{base_palette}{xxx}

\item{palette}{A \code{character(1)} which is the name of a palette in
the package \link{RColorBrewer}.}
}
\value{
A density plot

A plot
}
\description{
A shiny Module.
}
\examples{
data(ft)
qData <- assay(ft, 1)
conds <- design(ft)$Condition
densityPlot(qData, conds)

legend <- design(ft)$Sample.name
densityPlot(qData, conds, legend, 'Dark2')

library(QFeatures)
data(ft)
ui <- mod_plots_density_ui('plot')

server <- function(input, output, session) {
 data(ft)
 conds <- design(ft)$Condition
 
 mod_plots_density_server('plot',
                          obj = reactive({ft}),
                          conds = reactive({conds}),
                          pal = reactive({'Dark1'})
                          )
 }
shinyApp(ui=ui, server=server)
}
\author{
Samuel Wieczorek, Enora Fremy
}
