% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_density_plot.R
\name{densityPlotD_HC}
\alias{densityPlotD_HC}
\title{Builds a densityplot from a dataframe}
\usage{
densityPlotD_HC(qData, pData, legend = NULL, palette = NULL)
}
\arguments{
\item{qData}{numeric matrix}

\item{pData}{DataFrame}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{A vector of HEX Code colors (one color per condition).}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[['original']])
pData <- colData(Exp1_R25_pept)
legend <- colData(Exp1_R25_pept)@listData[["Condition"]]
densityPlotD_HC(qData, pData)

}
\author{
Samuel Wieczorek, Enora Fremy
}
