% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_all_plots.R, R/mod_ds_corrmatrix.R,
%   R/mod_ds_density.R, R/mod_ds_explorer.R, R/mod_ds_heatmap.R,
%   R/mod_ds_intensity.R, R/mod_ds_mv.R, R/mod_ds_pca.R, R/mod_ds_variance.R,
%   R/plot_ds_corrmatrix.R, R/plot_ds_variance.R, R/plot_ds_density.R,
%   R/plot_ds_violin.R
\name{listPlotModules}
\alias{listPlotModules}
\alias{mod_all_plots_ui}
\alias{mod_all_plots_server}
\alias{mod_ds_corrmatrix_ui}
\alias{mod_ds_corrmatrix_server}
\alias{mod_ds_density_ui}
\alias{mod_ds_density_server}
\alias{mod_ds_explorer_ui}
\alias{mod_ds_explorer_server}
\alias{mod_ds_heatmap_ui}
\alias{mod_ds_heatmap_server}
\alias{mod_ds_intensity_ui}
\alias{mod_ds_intensity_server}
\alias{mod_ds_mv_ui}
\alias{mod_ds_mv_server}
\alias{mod_ds_pca_ui}
\alias{mod_ds_pca_server}
\alias{mod_ds_variance_ui}
\alias{mod_ds_variance_server}
\alias{corrMatrixPlot}
\alias{CVDist}
\alias{densityPlot}
\alias{violinPlotD}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
listPlotModules()

mod_all_plots_ui(id)

mod_all_plots_server(id, object)

mod_ds_corrmatrix_ui(id)

mod_ds_corrmatrix_server(
  id,
  object,
  rate = reactive({     0.5 }),
  showValues = reactive({     FALSE })
)

mod_ds_density_ui(id)

mod_ds_density_server(id, object, conds, pal.name = reactive({
    NULL
}))

mod_ds_explorer_ui(id)

mod_ds_explorer_server(id, object, digits = reactive({
    3
}))

mod_ds_heatmap_ui(id)

mod_ds_heatmap_server(id, object, conds, width = 900)

mod_ds_intensity_ui(id)

mod_ds_intensity_server(id, object, exp.design, ...)

mod_ds_mv_ui(id)

mod_ds_mv_server(id, object, conds, pal.name)

mod_ds_pca_ui(id)

mod_ds_pca_server(id, object, conds)

mod_ds_variance_ui(id)

mod_ds_variance_server(id, object, conds, pal.name = NULL)

corrMatrixPlot(object, rate = 0.5, showValues = FALSE)

CVDist(object, conds = NULL, pal.name = NULL)

densityPlot(qData, conds, pal.name = NULL)

violinPlotD(object, exp.design, subset = NULL, pal.name)
}
\arguments{
\item{id}{shiny id}

\item{object}{An instance of the class \code{SummarizedExperiment}}

\item{rate}{The rate parameter to control the exponential law for
the gradient of colors}

\item{showValues}{xxx}

\item{conds}{A \code{character()} of condition name for each sample. The
length of 'conds' must be equal to the number of columns of 'qData'.}

\item{pal.name}{xxx}

\item{digits}{xxx}

\item{width}{xxx}

\item{exp.design}{xxx}

\item{...}{xxx}

\item{qData}{numeric matrix}

\item{subset}{A vector of index indicating rows to highlight}

\item{obj}{xxx}

\item{base_palette}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}
}
\value{
NA

NA

NA

A colored correlation matrix

A violinplot
}
\description{
Displays a correlation matrix of the quantitative data of a
numeric matrix.
}
\examples{
library(QFeatures)
data(ft)
corrMatrixPlot(assay(ft, 1))

}
\author{
Samuel Wieczorek, Enora Fremy
}
