% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_corr_matrix.R,
%   R/mod_plots_group_mv.R, R/mod_plots_intensity.R,
%   R/mod_plots_legend_colored_exprs.R, R/mod_plots_pca.R,
%   R/mod_plots_se_explorer.R, R/mod_plots_var_dist.R
\name{mod_plots_corr_matrix_ui}
\alias{mod_plots_corr_matrix_ui}
\alias{mod_plots_corr_matrix_server}
\alias{mod_plots_group_mv_ui}
\alias{mod_plots_group_mv_server}
\alias{mod_plots_intensity_ui}
\alias{mod_plots_intensity_server}
\alias{mod_plots_legend_colored_exprs_ui}
\alias{mod_plots_legend_colored_exprs_server}
\alias{mod_plots_pca_ui}
\alias{mod_plots_pca_server}
\alias{mod_plots_se_explorer_ui}
\alias{mod_plots_se_explorer_server}
\alias{mod_plots_var_dist_ui}
\alias{mod_plots_var_dist_server}
\title{mod_plots_corr_matrix_ui and mod_plots_corr_matrix_server}
\usage{
mod_plots_corr_matrix_ui(id)

mod_plots_corr_matrix_server(id, obj, names = NULL, gradientRate = NULL)

mod_plots_group_mv_ui(id)

mod_plots_group_mv_server(id, obj, conds, base_palette)

mod_plots_intensity_ui(id)

mod_plots_intensity_server(
  id,
  dataIn,
  meta,
  conds,
  base_palette = NULL,
  params = NULL,
  reset = NULL,
  slave = FALSE
)

mod_plots_legend_colored_exprs_ui(id)

mod_plots_legend_colored_exprs_server(id)

mod_plots_pca_ui(id)

mod_plots_pca_server(id, obj, coldata)

mod_plots_se_explorer_ui(id)

mod_plots_se_explorer_server(id, obj, originOfValues = NULL, colData = NULL)

mod_plots_var_dist_ui(id)

mod_plots_var_dist_server(id, obj, conds, base_palette = NULL)
}
\arguments{
\item{id, input, output, session}{Internal parameters for {shiny}.}

\item{obj}{xxx}

\item{names}{xxx}

\item{gradientRate}{xxx. Default value is 0.9}

\item{conds}{xxx}

\item{base_palette}{xxx}

\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\value{
NA

NA

NA

NA

NA

NA

NA

NA

NA

NA

NA
}
\description{
A shiny Module.

A shiny Module.

A shiny Module.

A shiny Module.

A shiny Module.

A shiny Module.

A shiny Module.
}
\author{
Samuel Wieczorek
}
\keyword{internal}
