% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_designExample.R
\name{design-example}
\alias{design-example}
\alias{mod_designExample_ui}
\alias{mod_designExample_server}
\title{xxxx}
\usage{
mod_designExample_ui(id)

mod_designExample_server(id, n)
}
\arguments{
\item{id}{xxx}

\item{n}{xxx}
}
\value{
NA
}
\description{
xxxx
}
\examples{
library(shiny)

################################################################
ui <- tagList(
  mod_designExample_ui("designEx")
)

server <- function(input, output, session) {
  mod_designExample_server("designEx", 3)
}
shinyApp(ui = ui, server = server)
}
