% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\docType{data}
\name{diff_analysis_sam}
\alias{diff_analysis_sam}
\alias{diff_analysis_sam,SummarizedExperiment-method}
\alias{diff_analysis_sam,Features-method}
\title{Check the validity of the experimental design}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
diff_analysis_sam

\S4method{diff_analysis_sam}{SummarizedExperiment}(object, ...)

\S4method{diff_analysis_sam}{Features}(object, i, name = "diffAnaAssay", ...)
}
\arguments{
\item{object}{An object of class `Features` or `SummarizedExperiment`.}

\item{...}{Additional parameters passed to inner functions.}

\item{i}{A numeric vector or a character vector giving the index or the 
name, respectively, of the assay(s) to be processed.}

\item{name}{A `character(1)` naming the new assay name. Default is `ttestAssay`.}
}
\description{
This manual page describes the computation of statistical test using [Features] objects. In the following
functions, if `object` is of class `Features`, and optional assay
index or name `i` can be specified to define the assay (by name of
index) on which to operate.

The following functions are currently available:

- `compute.t.test(xxxxx)` xxxxx.

- `compute.group.t.test(xxxxx)` xxxxx.
  
- `limma.complete.test(object, sampleTab)` uses the package Limma
}
\examples{
library(Features)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept[1:1000,]
object <- addAssay(object, Features::filterNA(object[[2]],  pNA = 0), name='filtered')
sTab <- colData(object)
gttest.se <- t_test_sam(object[[3]], sTab, FUN = compute.t.test)
object <- addAssay(object, gttest.se, name='t-test')
comp <- '25fmol_vs_10fmol'
da.se <- diff_analysis_sam(object[['t-test']], comp)

object <- diff_analysis_sam(object, 't-test', name='diffAna', comp)

}
\keyword{datasets}
