% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{diffAnaComputeFDR}
\alias{diffAnaComputeFDR}
\title{Computes the FDR corresponding to the p-values of the 
differential analysis using}
\usage{
diffAnaComputeFDR(
  logFC,
  pval,
  threshold_PVal = 0,
  threshold_LogFC = 0,
  pi0Method = 1
)
}
\arguments{
\item{logFC}{The result (logFC values) of the differential analysis processed 
by \code{\link{limmaCompleteTest}}}

\item{pval}{The result (p-values) of the differential analysis processed 
by \code{\link{limmaCompleteTest}}}

\item{threshold_PVal}{The threshold on p-pvalue to
distinguish between differential and non-differential data}

\item{threshold_LogFC}{The threshold on log(Fold Change) to
distinguish between differential and non-differential data}

\item{pi0Method}{The parameter pi0.method of the method adjust.p 
in the package \code{cp4p}}
}
\value{
The computed FDR value (floating number)
}
\description{
This function is a wrappper to the function adjust.p from the cp4p package. It returns the FDR corresponding to the p-values of the 
differential analysis. The FDR is computed with the function \code{p.adjust}\{stats\}..
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000]
obj <- addAssay(obj, Features::filterNA(obj[[2]],  pNA = 0), name='filtered')
se <- t.test.sam(obj[[3]], colData(obj), FUN = compute.t.test)
ind_logFC <- grep('_logFC', colnames(metadata(se)$t_test))
logFC <- setNames(as.data.frame(metadata(se)$t_test[,ind_logFC]), colnames(metadata(se)$t_test)[ind_logFC])
ind_pval <- grep('_pval', colnames(metadata(se)$t_test))
pval <- setNames(as.data.frame(metadata(se)$t_test[,ind_pval]), colnames(metadata(se)$t_test)[ind_pval])
diffAnaComputeFDR(logFC[,1], pval[,1])

}
\author{
Samuel Wieczorek
}
