% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R, R/plots_volcano.R
\name{diff_analysis_sam,SummarizedExperiment-method}
\alias{diff_analysis_sam,SummarizedExperiment-method}
\alias{diff_analysis_sam,QFeatures-method}
\alias{histPValue_HC}
\alias{diffAnaComputeFDR}
\alias{Get_AllComparisons}
\alias{diffAnalysis}
\alias{diffAnaVolcanoplot_rCharts}
\title{Plots a histogram of p-values}
\usage{
\S4method{diff_analysis_sam}{SummarizedExperiment}(object, ...)

\S4method{diff_analysis_sam}{QFeatures}(object, i, name = "diffAnaAssay", ...)

histPValue_HC(pval_ll, bins = 80, pi0 = 1)

diffAnaComputeFDR(
  logFC,
  pval,
  threshold_PVal = 0,
  threshold_LogFC = 0,
  pi0Method = 1
)

Get_AllComparisons(obj)

diffAnalysis(obj, comp, th_pval = 0, th_logFC = 0)

diffAnaVolcanoplot_rCharts(
  df,
  threshold_pVal = 1e-60,
  threshold_logFC = 0,
  conditions = NULL,
  clickFunction = NULL,
  palette = NULL,
  swap = FALSE
)
}
\arguments{
\item{object}{An object of class \code{QFeatures} or \code{SummarizedExperiment}.}

\item{...}{Additional parameters passed to inner functions.}

\item{i}{A numeric vector or a character vector giving the index or the
name, respectively, of the assay(s) to be processed.}

\item{name}{A \code{character(1)} naming the new assay name. Default is \code{ttestAssay}.}

\item{pval_ll}{A data frame which contains the p-values for the set of comparisons to show.}

\item{bins}{xxx}

\item{pi0}{xxx}

\item{logFC}{The result (logFC values) of the differential analysis processed
by 'limmaCompleteTest'}

\item{pval}{The result (p-values) of the differential analysis processed
by 'limmaCompleteTest'}

\item{threshold_PVal}{The threshold on p-pvalue to
distinguish between differential and non-differential data}

\item{threshold_LogFC}{The threshold on log(Fold Change) to
distinguish between differential and non-differential data}

\item{pi0Method}{The parameter pi0.method of the method adjust.p
in the package \code{cp4p}}

\item{obj}{An object of class \code{SUmmarizeExperiment}.}

\item{comp}{The name of the comparison (given by the t-tests) on whihc to apply thresholds}

\item{th_pval}{xxx}

\item{th_logFC}{xxx}

\item{df}{A dataframe which contains the following slots :
x : a vector of the log(fold change) values of the differential analysis,
y : a vector of the p-value values returned by the differential analysis.
index : a vector of the rownames of the data.
This dataframe must has been built with the option stringsAsFactors set
to FALSE. There may be additional slots which will be used to show
informations in the tooltip. The name of these slots must begin with the
prefix "tooltip_". It will be automatically removed in the plot.}

\item{threshold_pVal}{A floating number which represents the p-value that
separates differential and non-differential data.}

\item{threshold_logFC}{A floating number which represents the log of the
Fold Change that separates differential and non-differential data.}

\item{conditions}{A list of the names of condition 1 and 2 used for the
differential analysis.}

\item{clickFunction}{A string that contains a JavaScript function used to
show info from slots in df. The variable this.index refers to the slot
named index and allows to retrieve the right row to show in the tooltip.}

\item{palette}{xxx}

\item{swap}{A boolean that indicates if the conditions have been swaped}
}
\value{
A plot

The computed FDR value (floating number)

A list of two slots: logFC and P_Value. Each slot contains a DataFrame

An object of class \code{SummarizedExperiment}

An interactive volcanoplot
}
\description{
This function is a wrappper to the function adjust.p from the cp4p package. It returns the FDR corresponding to the p-values of the
differential analysis. The FDR is computed with the function \code{p.adjust}\{stats\}..

This method returns a list of the statistical tests performed with xxx and formatted as a
list of two DataFrame: one for the logFC data and one for the p-values data.

This method returns a class \code{QFeatures} object with the results
of differential analysis

Plots an interactive volcanoplot after the differential analysis.
Typically, the log of Fold Change is represented on the X-axis and the
log10 of the p-value is drawn on the Y-axis. When the \code{threshold_pVal}
and the \code{threshold_logFC} are set, two lines are drawn respectively on
the y-axis and the X-axis to visually distinguish between differential and
non differential data. With the use of the package Highcharter, a
customizable tooltip appears when the user put the mouse's pointer over
a point of the scatter plot.
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
object <- Exp1_R25_pept[seq_len(1000),]
object <- addAssay(object, QFeatures::filterNA(object[[2]],  pNA = 0), name='filtered')
object <- addListAdjacencyMatrices(object, 3)
sTab <- colData(object)
se <- t_test_sam(object[[3]], sTab, FUN = compute.t.test)
pval <- metadata(se)$t_test[['25fmol_vs_10fmol_pval']]
histPValue_HC(pval)

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000)]
obj <- addAssay(obj, QFeatures::filterNA(obj[[2]],  pNA = 0), name='filtered')
se <- t_test_sam(obj[[3]], colData(obj), FUN = compute.t.test)
ind_logFC <- grep('_logFC', colnames(metadata(se)$t_test))
logFC <- setNames(as.data.frame(metadata(se)$t_test[,ind_logFC]), 
colnames(metadata(se)$t_test)[ind_logFC])
ind_pval <- grep('_pval', colnames(metadata(se)$t_test))
pval <- setNames(as.data.frame(metadata(se)$t_test[,ind_pval]), 
colnames(metadata(se)$t_test)[ind_pval])
diffAnaComputeFDR(logFC[,1], pval[,1])

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
object <- Exp1_R25_pept[seq_len(1000),]
object <- addAssay(object, QFeatures::filterNA(object[[2]],  pNA = 0), name='filtered')
sTab <- colData(object)
gttest.se <- t_test_sam(object[[3]], sTab, FUN = compute.t.test)
object <- addAssay(object, gttest.se, name='t-test')
comp <- '25fmol_vs_10fmol'
object <- diff_analysis_sam(object, 't-test', name='diffAna', comp)

allComp <- Get_AllComparisons(object[['diffAna']])

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000)]
obj <- addAssay(obj, QFeatures::filterNA(obj[[2]],  pNA = 0), name='filtered')
obj <- addAssay(obj, t_test_sam(obj[[3]], colData(obj), FUN = 'compute.t.test'), name='t-test')
comp <- '25fmol_vs_10fmol'
da.se <- diffAnalysis(obj[['t-test']], comp, th_pval=0, th_logFC=0)

\donttest{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DaparToolshedData')
obj <- QFeatures::filterNA(Exp1_R25_pept,i='original') # remove all rows with NAs
qData <- assay(obj[['original']])
pData <- as.data.frame(colData(obj))
data <- limma.complete.test(qData,pData)
df <- data.frame(x=data[,1], y = -log10(data[,2]),index = as.character(rownames(qData)))
colnames(df) <- c("x", "y", "index")
tooltipSlot <- c("Sequence", "Score")
df <- cbind(df,rowData(obj[['original']])[tooltipSlot])
colnames(df) <- gsub(".", "_", colnames(df), fixed=TRUE)
if (ncol(df) > 3){
    colnames(df)[4:ncol(df)] <- 
    paste("tooltip_", colnames(df)[4:ncol(df)], sep="")}
hc_clickFunction <- JS("function(event) {
Shiny.onInputChange('eventPointClicked', [this.index]+'_'+ [this.series.name]);}")
cond <- unique(colData(obj)[['Condition']])
diffAnaVolcanoplot_rCharts(df, 2.5, 1, cond, hc_clickFunction) 
}

}
\author{
Samuel Wieczorek

Samuel Wieczorek

Samuel Wieczorek

Samuel Wieczorek

Samuel Wieczorek, Enora Fremy
}
