% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_func.R
\name{.Build_Topn_Mat}
\alias{.Build_Topn_Mat}
\title{xxxxx}
\usage{
.Build_Topn_Mat(X, qData = NULL, fun = "rowMedians", n = 10)
}
\arguments{
\item{X}{xxx}

\item{qData}{xxx}

\item{fun}{xxx}

\item{n}{xxx}
}
\description{
xxx
}
\details{
This function builds an intermediate matrix with scores for each peptide
based on 'fun' parameter. Once this matrix is built, one select the 'n' peptides
which have the higher score

- rowMedians xxx
- rowMeans xxx
- rowSums xxx
}
\examples{

}
