% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{Example_Palette}
\alias{Example_Palette}
\title{xxx}
\usage{
Example_Palette(conditions, base_palette)
}
\arguments{
\item{conditions}{xxxx}

\item{base_palette}{xxx}
}
\description{
Builds a vector of `#conditions` colors.
}
