% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering_MV.R
\docType{data}
\name{filterFeaturesSam}
\alias{filterFeaturesSam}
\alias{filterFeaturesSam,SummarizedExperiment-method}
\alias{filterFeaturesSam,QFeatures-method}
\title{Check the validity of the experimental design}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
filterFeaturesSam

\S4method{filterFeaturesSam}{SummarizedExperiment}(object, filter, ...)

\S4method{filterFeaturesSam}{QFeatures}(object, i, name = "filterAssay", filter, ...)
}
\arguments{
\item{object}{An object of class `QFeatures`.}

\item{filter}{xxx}

\item{...}{Additional parameters passed to inner functions.}

\item{i}{A numeric vector or a character vector giving the index or the 
name, respectively, of the assay(s) to be processed.}

\item{name}{A `character(1)` naming the new assay name. Defaults
are `ttestAssay`.}
}
\description{
This manual page describes the xxxx
}
\details{
xxx
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_prot, package='DAPARdata2')
object <- Exp1_R25_prot
filter_600 <- VariableFilter(field='Sequence_length', value=as.numeric('600'), condition='>')
object <- filterFeaturesSam(object, i=2, filter=reverse_600)


}
\keyword{datasets}
