% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{formatLimmaResult}
\alias{formatLimmaResult}
\title{xxxx}
\usage{
formatLimmaResult(fit, conds, contrast)
}
\arguments{
\item{fit}{xxxx}

\item{conds}{xxxx}

\item{contrast}{xxxx}
}
\value{
A list of two dataframes : logFC and P_Value. The first one contains
the logFC values of all the comparisons (one column for one comparison), the second one contains
the pvalue of all the comparisons (one column for one comparison). The names of the columns for those two dataframes
are identical and correspond to the description of the comparison.
}
\description{
xxxx
}
\examples{
\dontrun{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept
object <- addAssay(object, QFeatures::filterNA(object[[2]],  pNA = 0), name='filtered')
qData <- assay(object, 'filtered')
sTab <- colData(object)
limma <- limmaCompleteTest(qData, sTab)
}

}
\author{
Samuel Wieczorek
}
