% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{get.pep.prot.cc}
\alias{get.pep.prot.cc}
\title{Build the list of connex composant of an adjacency matrix.}
\usage{
get.pep.prot.cc(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of CC
}
\description{
Method to build the list of connex composant of an adjacency matrix
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2') 
plist <- rowData(Exp1_R25_pept[['original_log']][1:1000,])[,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original_log']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
ll <- get.pep.prot.cc(X.list$all)

}
\author{
Thomas Burger, Samuel Wieczorek
}
