% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{Get_qMetadata,SummarizedExperiment-method}
\alias{Get_qMetadata,SummarizedExperiment-method}
\title{xxxx}
\usage{
\S4method{Get_qMetadata}{SummarizedExperiment}(object, names, ...)
}
\arguments{
\item{obj}{An object of class `SummarizedExperiment`}
}
\description{
xxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:10,]
mc <- Get_qMetadata(obj[[2]])
}
