% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\docType{data}
\name{Get_qMetadata}
\alias{Get_qMetadata}
<<<<<<< HEAD:man/Get_qMetadata.Rd
\title{xxxx}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
Get_qMetadata
=======
\alias{Get_qMetadata,SummarizedExperiment-method}
\alias{Get_qMetadata,QFeatures-method}
\title{xxxx}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
Get_qMetadata

\S4method{Get_qMetadata}{SummarizedExperiment}(object, names, ...)

\S4method{Get_qMetadata}{QFeatures}(object, i, ...)
}
\arguments{
\item{object}{An object of class `QFeatures`}

\item{names}{xxx}

\item{...}{xxx}

\item{i}{xxx}
>>>>>>> bd1ab6a739c5a9402b382daddfe0b803a076915c:man/Get_qMetadata.Rd
}
\description{
xxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:10,]
mc <- Get_qMetadata(obj[[2]])

}
\keyword{datasets}
