% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{Get_qMetadata_names}
\alias{Get_qMetadata_names}
\alias{Get_qMetadata_names,SummarizedExperiment-method}
\alias{Get_qMetadata_names,QFeatures-method}
\title{Get the colnames of quantitative cell metadata}
\usage{
\S4method{Get_qMetadata_names}{SummarizedExperiment}(object, ...)

\S4method{Get_qMetadata_names}{QFeatures}(object, i, ...)
}
\arguments{
\item{object}{An object of class 'QFeatures' the MEC tag in the metacell}

\item{...}{xxx}

\item{i}{xxx}
}
\description{
These names are common to all assays contained in the object. This is why
they are stored in the global metadata. This function is used whenever it is necessary
to (re)detect MEC and POV (new dataset or when post processing protein metacell 
after aggregation)
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
Get_qMetadata_names(Exp1_R25_pept)

}
