% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{GetConnectedComps}
\alias{GetConnectedComps}
\title{Record the adjacency matrices}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
GetConnectedComps
}
\description{
Record the adjacency matrices
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
ll.X <- SetConnectedComps(Exp1_R25_pept, 2)

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
