% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{GetDetailedNbPeptides}
\alias{GetDetailedNbPeptides}
\title{Computes the detailed number of peptides for each protein}
\usage{
GetDetailedNbPeptides(X)
}
\arguments{
\item{X.list}{A list of adjacency matrices}
}
\value{
A data.frame containing the following vectors:
* nTotal: The number of peptides for each protein,
* nShared: the number of shared peptides for each protein,
* nSPec: the number of specific peptides for each protein.

Each row of these three vectors represent a protein. Thus, the length of these vectors is equal 
to the number of proteins.
}
\description{
Method to compute the detailed number of peptides for each protein
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original']][1:1000,])[1:1000,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All', 'Shared', 'Specific'))
n <- GetDetailedNbPeptides(X.list$all)

}
\author{
Samuel Wieczorek
}
