% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{GetDetailedNbPeptidesUsed}
\alias{GetDetailedNbPeptidesUsed}
\title{Computes the detailed number of peptides used for aggregating each protein w.r.t NA values in 
peptide quantitative dataset. Even if a peptide is part of a protein, if its value is NA, it do not be used
for aggreation.}
\usage{
GetDetailedNbPeptidesUsed(X, qPepData)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X.list}{A list of adjacency matrices.}
}
\value{
A list of three items:
* nAll: the number of pall eptides used for aggregation,
* nShared: the number of shared peptides used for aggregation,
* nSpec: the total number of pspecific eptides used for aggregation.
}
\description{
Method to compute the detailed number of quantified peptides used for aggregating each protein
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
qPepData <- assay(Exp1_R25_pept,'original_log')[1:1000,]
plist <- rowData(Exp1_R25_pept[['original']][1:1000,])[1:1000,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All', 'Shared', 'Specific'))
n <- GetDetailedNbPeptidesUsed(X.list$all, qPepData)

}
\seealso{
The function 'BuildListAdjacencyMatrices'.
}
\author{
Samuel Wieczorek
}
