% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{GetIndicesForPrefix}
\alias{GetIndicesForPrefix}
\title{Get the indices of the lines to delete, based on a prefix string}
\usage{
GetIndicesForPrefix(object, col = NULL, prefix = NULL)
}
\arguments{
\item{object}{An object of class \code{xxx}.}

\item{col}{The name of the column that correspond to the data to filter}

\item{prefix}{A character string that is the prefix to find in the data}
}
\value{
A vector of integers.
}
\description{
This function returns the indice of the lines to delete, based on a 
prefix string
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
ind <- getIndicesOfLinesToRemove(Exp1_R25_pept[1:100], "Potential_contaminant", 
prefix="+")

}
\author{
Samuel Wieczorek
}
