% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{GetKeyId}
\alias{GetKeyId}
\alias{GetKeyId,QFeatures-method}
\title{Returns the contains of the slot processing  of an object of 
class \code{MSnSet}}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
GetKeyId

\S4method{GetKeyId}{QFeatures}(object, i, ...)
}
\arguments{
\item{object}{xxx}

\item{i}{xxx}

\item{...}{xxx}
}
\description{
xxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
GetKeyId(Exp1_R25_pept, 2)

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
