% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{GetMetacell,SummarizedExperiment-method}
\alias{GetMetacell,SummarizedExperiment-method}
\title{xxxx}
\usage{
\S4method{GetMetacell}{SummarizedExperiment}(object, names)
}
\arguments{
\item{obj}{An object of class `SummarizedExperiment`}
}
\description{
xxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:10,]
mc <- GetMetacell(obj[[2]])
}
