% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{GetNbPeptidesUsed}
\alias{GetNbPeptidesUsed}
\title{Computes the number of peptides used for aggregating each protein}
\usage{
GetNbPeptidesUsed(qPepData, X)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
A data.frame
}
\description{
Method to compute the number of quantified peptides used for aggregating each protein
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
qPepData <- assay(Exp1_R25_pept,'original_log')[1:1000,]
plist <- rowData(Exp1_R25_pept[['original']][1:1000,])[1:1000,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All', 'Shared', 'Specific'))
n <- GetNbPeptidesUsed(qPepData, X.list$all)

}
\author{
Samuel Wieczorek
}
