% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetProstarVersions}
\alias{GetProstarVersions}
\title{Versions of installed packages of Prostar suite}
\usage{
GetProstarVersions()
}
\value{
A list
}
\description{
Return the versions of Prostar, Dapar2 and DaparData2
}
\examples{
GetProstarVersions()

}
\author{
Samuel Wieczorek
}
