% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{GetSoftAvailables}
\alias{GetSoftAvailables}
\title{The set of available softwares to convert from}
\usage{
GetSoftAvailables()
}
\description{
The set of available softwares to convert from
}
\examples{
GetSoftAvailables()
}
