% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\docType{data}
\name{GetTypeDataset}
\alias{GetTypeDataset}
\alias{GetTypeDataset,SummarizedExperiment-method}
\title{Get the type of dataset}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
GetTypeDataset

\S4method{GetTypeDataset}{SummarizedExperiment}(object, ...)
}
\arguments{
\item{object}{An object of class 'SummarizedExperiment'}
}
\value{
NA

NA
}
\description{
Get the type of dataset
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
GetTypeDataset(Exp1_R25_pept[[2]])

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
