% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{GraphPepProt_hc}
\alias{GraphPepProt_hc}
\title{Function to create a histogram that shows the repartition of
peptides w.r.t. the proteins}
\usage{
GraphPepProt_hc(X, type = "all")
}
\arguments{
\item{X}{An adjacency matrix.}

\item{type}{A string which is the type of matrix (used to build the plot title). Default value is 'all'.}
}
\value{
A histogram
}
\description{
#' This function computes the number of peptides used to aggregate proteins.
#' 
#' @title Compute the number of peptides used to aggregate proteins
#' @param M A "valued" adjacency matrix in which lines and columns correspond 
#' respectively to peptides and proteins.
#' @return A vector of boolean which is the adjacency matrix 
#' but with NA values if they exist in the intensity matrix.
#' @author Alexia Dorffer
#' @examples
#' utils::data(Exp1_R25_pept, package='DAPARdata2')
#' PG <- rowData(Exp1_R25_pept[['original']])[,metadata(Exp1_R25_pept)$parentProtId]
#' names <- names((Exp1_R25_pept[['original']]))
#' X <- BuildAdjacencyMatrix(PG, names, TRUE)
#' CountPep(X)
#' @export
CountPep <- function (M) {
  z <- M
  z[z!=0] <- 1
  return(z)
}
Method to plot the disrtibution (histogram) of peptides w.r.t the proteins with proteins and peptides 
in an adjacency matrix
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original']][1:1000,])[1:1000,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original']])[1:1000]
X <- BuildListAdjacencyMatrices(plist, names, type=c('All', 'Shared', 'Specific'))
GraphPepProt_hc(X$all, 'all')

}
\author{
Alexia Dorffer, Samuel Wieczorek
}
