% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ttest.R
\name{groupttest}
\alias{groupttest}
\title{xxxxxx}
\usage{
groupttest(X, qData1 = NULL, qData2 = NULL)
}
\arguments{
\item{X}{xxxxx.}

\item{qData1}{xxxx.}

\item{qData2}{xxxx.}
}
\value{
NA
}
\description{
xxxxxx
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[seq_len(1000),]
obj <- addAssay(obj, QFeatures::filterNA(obj[[2]],  pNA = 0), name='filtered')
obj <- addListAdjacencyMatrices(obj, 3)
qData <- assay(obj[['filtered']])
X <- GetAdjMat(obj[[3]])$onlySpec
gttest <- groupttest(X, qData[,seq_len(3)], qData[,seq(4,6)])

}
\author{
Thomas Burger, Samuel Wieczorek
}
