% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\name{hc_logFC_DensityPlot}
\alias{hc_logFC_DensityPlot}
\title{Density plots of logFC values}
\usage{
hc_logFC_DensityPlot(df_logFC, threshold_LogFC = 0, palette = NULL)
}
\arguments{
\item{df_logFC}{A dataframe that contains the logFC values for the set of comparisons to show. Each column corresponds to a comparison.}

\item{threshold_LogFC}{The threshold on log(Fold Change) to distinguish between differential and non-differential data}

\item{palette}{xxx}
}
\value{
A highcharts density plot
}
\description{
This function show the density plots of Fold Change (the same as calculated by limma) for a list 
of the comparisons of conditions in a differnetial analysis.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000]
obj <- addAssay(obj, QFeatures::filterNA(obj[[2]],  pNA = 0), name='filtered')
se <- t_test_sam(obj[[3]], colData(obj), FUN = compute.t.test)
ind <- grep('_logFC', colnames(metadata(se)$t_test))
df <- setNames(as.data.frame(metadata(se)$t_test[,ind]), colnames(metadata(se)$t_test)[ind])
pal <- ExtendPalette(2, 'Dark2')
if (length(ind)>0)
hc_logFC_DensityPlot(df, palette = pal)

}
\author{
Samuel Wieczorek
}
