% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_heatmap.R, R/mod_plots_heatmap.R
\name{heatmap_plot}
\alias{heatmap_plot}
\alias{heatmapD}
\alias{mv.heatmap}
\alias{mod_heatmap_plot_ui}
\alias{mod_heatmap_plot_server}
\title{Heatmap of the quantitative proteomics data}
\usage{
heatmapD(qData, conds, distfun = "euclidean", hclustfun = "complete")

mv.heatmap(
  x,
  col = grDevices::heat.colors(100),
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)

mod_heatmap_plot_ui(id)

mod_heatmap_plot_server(id, obj, conds, width = 900)
}
\arguments{
\item{qData}{A \code{matrix} or \code{array} quantitative values}

\item{conds}{xxx}

\item{distfun}{A \code{character(1)} defining the distance used by the clustering
algorithm to compute. Default value is 'euclidean'. See \code{help(heatmap.2)}}

\item{hclustfun}{the clustering algorithm used to build the dendrogram.
Default value is 'complete'. See \code{help(heatmap.2)}}

\item{x}{A \code{matrix} or \code{array} containing the quantitative data.}

\item{col}{Colors used for the image. Defaults to heat colors (heat.colors).}

\item{srtCol}{Angle of column conds, in degrees from horizontal}

\item{labCol}{Character vectors with column conds to use.}

\item{labRow}{Character vectors with row conds to use.}

\item{key}{Logical indicating whether a color-key should be shown.}

\item{key.title}{Main title of the color key. If set to NA no title will
be plotted.}

\item{main}{Main title; default to none.}

\item{ylab}{y-axis title; default to none.}

\item{id}{xxx}

\item{obj}{xxx}

\item{width}{xxx}
}
\value{
A heatmap

NA
}
\description{
This function is a wrapper to 'heatmap.2' of the package \code{gplots}
that displays a numeric matrix
}
\section{generic plot}{


The function \code{\link[=heatmapD]{heatmapD()}}
}

\section{plot for missingvalues}{


The function [] is inspired from the function 'heatmap.2'
that displays a numeric matrix. For more information, please refer to the help
of the heatmap.2 function.
}

\examples{

library(QFeatures)
data(ft)
conds <- design(ft)$Condition

#----------------------------------------------
# Plots a heatmap for generic quantitative data
#----------------------------------------------

heatmapD(assay(ft, 1), conds)

#----------------------------------------------
# Plots a heatmap for missing values visualization
#----------------------------------------------

library(QFeatures)
data(ft)
mv.heatmap(assay(ft, 1))


#----------------------------------------------
# Launches a shiny module
#----------------------------------------------

if(interactive()){
 library(QFeatures)
 library(shiny)
 library(DaparToolshed)
 data(ft)
 ui <- mod_heatmap_plot_ui('plot')

 server <- function(input, output, session) {
  conds <- design(ft)$Condition
 
  mod_heatmap_plot_server('plot',
                      obj = reactive({ft[[1]]}),
                      conds = reactive({conds}),
                      pal.name = reactive({'Dark2'})
                      )
  }
 
 shinyApp(ui=ui, server=server)
}
}
\author{
Samuel Wieczorek
}
