% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_heatmap.R
\name{heatmapD}
\alias{heatmapD}
\title{Heatmap of the quantitative proteomics data}
\usage{
heatmapD(qData, conds, distance = "euclidean", cluster = "complete")
}
\arguments{
\item{qData}{A dataframe of numeric values}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{distance}{The distance used by the clustering algorithm to compute
the dendrogram. See \code{help(heatmap.2)}}

\item{cluster}{the clustering algorithm used to build the dendrogram.
See \code{help(heatmap.2)}}
}
\value{
A heatmap
}
\description{
This function is a wrapper to 'heatmap.2' that displays a numeric matrix
}
\examples{
\donttest{
library(QFeatures)
Exp1_R25_prot <- readRDS(system.file("data", 'Exp1_R25_prot.rda', package="DaparToolshedData"))
conds <- colData(Exp1_R25_prot)[['Condition']]
qData <- assay(Exp1_R25_prot)
heatmapD(qData, conds)
}

}
\author{
Samuel Wieczorek
}
