% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_heatmap.R
\name{heatmapForMissingValues}
\alias{heatmapForMissingValues}
\title{This function is inspired from the function 'heatmap.2' 
that displays a numeric matrix. For more information, please refer to the help 
of the heatmap.2 function.}
\usage{
heatmapForMissingValues(
  x,
  col = grDevices::heat.colors(100),
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)
}
\arguments{
\item{x}{A matrix that contains quantitative data.}

\item{col}{Colors used for the image. Defaults to heat colors (heat.colors).}

\item{srtCol}{Angle of column conds, in degrees from horizontal}

\item{labCol}{Character vectors with column conds to use.}

\item{labRow}{Character vectors with row conds to use.}

\item{key}{Logical indicating whether a color-key should be shown.}

\item{key.title}{Main title of the color key. If set to NA no title will 
be plotted.}

\item{main}{Main title; default to none.}

\item{ylab}{y-axis title; default to none.}
}
\value{
A heatmap
}
\description{
This function is inspired from the function 'heatmap.2' 
that displays a numeric matrix. For more information, please refer to the help 
of the heatmap.2 function.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[[2]])[1:1000,]
heatmapForMissingValues(qData)

}
\author{
Samuel Wieczorek, Enora Fremy
}
