% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\docType{data}
\name{HypoyhesisTest}
\alias{HypoyhesisTest}
\title{Record the adjacency matrices}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
HypoyhesisTest
}
\description{
Record the adjacency matrices
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
Exp1_R25_pept <- addAssay(Exp1_R25_pept, 
QFeatures::filterNA(Exp1_R25_pept[[2]],  pNA = 0), name='filtered')
sTab <- as.data.frame(colData(Exp1_R25_pept))
ttest <- compute_t_tests(Exp1_R25_pept[['filtered']], sTab ,"OnevsOne")
Exp1_R25_pept <- HypoyhesisTest(Exp1_R25_pept, 3, test = 't_test', sample = sTab, contrast='OnevsOne', type='Student')

Exp1_R25_pept <- HypoyhesisTest(Exp1_R25_pept, 3, test = 'limma', sample = sTab, type='anova1way')

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
