% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_qfeatures.R, R/import_2_qfeatures.R
\name{writeQFeaturesToExcel}
\alias{writeQFeaturesToExcel}
\alias{readExcel}
\alias{listSheets}
\alias{createQFeatures}
\title{This function exports a \code{QFeatures} object to a Excel file.}
\usage{
writeQFeaturesToExcel(obj, filename)

readExcel(file, extension, sheet)

listSheets(file)

createQFeatures(
  data = NULL,
  file = NULL,
  sample,
  indQData,
  keyId = "AutoID",
  indQMetadata = NULL,
  logTransform = FALSE,
  forceNA = TRUE,
  typeDataset,
  parentProtId = NULL,
  analysis = "foo",
  processes = NULL,
  typePipeline = NULL,
  software = NULL
)
}
\arguments{
\item{obj}{An object of class \code{QFeatures}.}

\item{filename}{A character string for the name of the Excel file.}

\item{file}{The name of a file xxx}

\item{extension}{The extension of the file}

\item{sheet}{The name of the sheet}

\item{data}{The name of a tab-separated file that contains the data.}

\item{sample}{A dataframe describing the samples (in lines).}

\item{indQData}{A vector of string where each element is the name
of a column in designTable that have to be integrated in
the \code{rowData()} table of the \code{QFeatures} object.}

\item{keyId}{The indice of the column containing the ID of entities
(peptides or proteins)}

\item{indQMetadata}{xxxxxxxxxxx}

\item{logTransform}{A boolean value to indicate if the data have to be
log-transformed (Default is FALSE)}

\item{forceNA}{A boolean value to indicate if the 0 and NaN values of
intensity have to be replaced by NA (Default is FALSE)}

\item{typeDataset}{A string that indicates whether the dataset is about}

\item{parentProtId}{For peptide entities, a string which is the name of a column in rowData. It contains the id of parent
proteins and is used to generate adjacency matrix and process to aggregation.}

\item{analysis}{The name of the MS analysis}

\item{processes}{xxxx}

\item{typePipeline}{The type of pipeline used with this dataset. The list of predefined
pipelines in DaparToolshed can be obtained with the function \code{pipelines()}. Default value is NULL}

\item{software}{xxx}
}
\value{
A Excel file (.xlsx)

A data.frame

A vector

A list of sheets name

An instance of class \code{QFeatures}.
}
\description{
This function exports a \code{QFeatures} data object to a Excel file.
Each of the three data.frames in the \code{QFeatures} object (ie expression data,
colData and rowData are respectively integrated into separate sheets in
the Excel file). The colored cells in the experimental data correspond to the 
original missing values which have been imputed.

Builds an object of class \code{QFeatures} from a
single tabulated-like file for quantitative and meta-data and a dataframe
for the samples description.
}
\examples{
\donttest{
Sys.setenv("R_ZIPCMD"= Sys.which("zip"))
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000)]
writeQFeaturesToExcel(obj, "foo")
}

{
data.file <- system.file("extdata", "Exp1_R25_pept.txt", package="DaparToolshedData")
data <- read.table(data.file, header=TRUE, sep="\t", as.is=TRUE, stringsAsFactors = FALSE)
sample.file <- system.file("extdata", "samples_Exp1_R25.txt", package="DaparToolshedData")
sample <- read.table(sample.file, header=TRUE, sep="\t", as.is=TRUE, stringsAsFactors = FALSE)
ft <- createQFeatures(data = data, sample = sample, indQData = 56:61, keyId = 'Sequence', analysis = 'test',
logTransform = TRUE, indQMetadata = 43:48, typeDataset = 'peptide',
parentProtId = 'Protein_group_IDs', forceNA = TRUE, software = 'maxquant')
}

}
\author{
Samuel Wieczorek

Samuel Wieczorek

Samuel Wieczorek

Samuel Wieczorek
}
