% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2_qfeatures.R
\name{createQFeatures}
\alias{createQFeatures}
\title{Creates an object of class \code{QFeatures} from text file.}
\usage{
createQFeatures(
  data = NULL,
  file = NULL,
  sample,
  indQData,
  keyId = "AutoID",
  indQMetadata = NULL,
  force.na = TRUE,
  typeDataset,
  parentProtId = NULL,
  analysis = "foo",
  processes = NULL,
  typePipeline = NULL,
  software = NULL,
  name = "original"
)
}
\arguments{
\item{data}{The name of a tab-separated file that contains the data.}

\item{file}{`character(1)`. The name of a file xxx}

\item{sample}{A dataframe describing the samples (in lines).}

\item{indQData}{A vector of string where each element is the name
of a column in designTable that have to be integrated in
the \code{rowData()} table of the \code{QFeatures} object.}

\item{keyId}{A `character(1)` or `numeric(1)` which is the indice of the column containing the ID of entities
(peptides or proteins)}

\item{indQMetadata}{xxxxxxxxxxx}

\item{force.na}{A `boolean` that indicates if the '0' and 'NaN' values of
quantitative values  must be replaced by 'NA' (Default is FALSE)}

\item{typeDataset}{A string that indicates whether the dataset is about}

\item{parentProtId}{A `character(1)` For peptide entities, a string which is the name of a column in rowData. It contains the id of parent
proteins and is used to generate adjacency matrix and process to aggregation.}

\item{analysis}{A `character(1)` which is the name of the MS study.}

\item{processes}{A vector of A `character()` which contains the name of processes
which has already been run on the data. Default is 'original'.}

\item{typePipeline}{A `character(1)` The type of pipeline used with this dataset. The list of predefined
pipelines in DaparToolshed can be obtained with the function \code{pipelines()}. Default value is NULL}

\item{software}{A `character(1)`}

\item{name}{A `character(1)` which is the name of the assay in the QFeatures object. 
Default is 'original'}
}
\value{
An instance of class \code{QFeatures}.
}
\description{
Creates an object of class \code{QFeatures} from a
single tabulated-like file for quantitative and meta-data and a dataframe
for the samples description.
}
\examples{
{
data.file <- system.file("extdata", "Exp1_R25_pept.txt", package="DaparToolshedData")
data <- read.table(data.file, header=TRUE, sep="\t", as.is=TRUE, stringsAsFactors = FALSE)
sample.file <- system.file("extdata", "samples_Exp1_R25.txt", package="DaparToolshedData")
sample <- read.table(sample.file, header=TRUE, sep="\t", as.is=TRUE, stringsAsFactors = FALSE)
ft <- createQFeatures(data = data, sample = sample, indQData = 56:61, keyId = 'Sequence', analysis = 'test',
indQMetadata = 43:48, typeDataset = 'peptide',
parentProtId = 'Protein_group_IDs', forceNA = TRUE, software = 'maxquant')
}

}
\author{
Samuel Wieczorek
}
