% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_mv_for_impute.R,
%   R/mod_plots_mv_for_imputation.R
\name{hc_mvTypePlot2}
\alias{hc_mvTypePlot2}
\alias{mvImage}
\alias{mod_plots_mv_for_imputation_ui}
\alias{mod_plots_mv_for_imputation_server}
\title{Distribution of Observed values with respect to intensity values}
\usage{
hc_mvTypePlot2(obj, i, title = NULL, palette = NULL)

mvImage(obj, i)

mod_plots_mv_for_imputation_ui(id)

mod_plots_mv_for_imputation_server(id, obj, ind, title = NULL, palette = NULL)
}
\arguments{
\item{obj}{xxx}

\item{i}{xxx}

\item{title}{The title of the plot}

\item{palette}{The different colors for conditions}

\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\value{
Density plots

A heatmap

NA

NA
}
\description{
This method shows density plots which represents the repartition of
Partial Observed Values for each replicate in the dataset.
The colors correspond to the different conditions (Condition in colData of
the object of class \code{QFeatures}).
The x-axis represent the mean of intensity for one condition and one
entity in the dataset (i. e. a protein)
whereas the y-axis count the number of observed values for this entity
and the considered condition.

Plots a heatmap of the quantitative data. Each column represent one of
the conditions in the object of class \code{MSnSet} and
the color is proportional to the mean of intensity for each line of
the dataset.
The lines have been sorted in order to vizualize easily the different
number of missing values. A white square is plotted for missing values.

A shiny Module.
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
hc_mvTypePlot2(Exp1_R25_pept, 2, title="POV distribution")

pal <- ExtendPalette(length(unique(conds)), 'Dark2')
hc_mvTypePlot2(Exp1_R25_pept, 2, title="POV distribution", palette=pal)

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
mvImage(Exp1_R25_pept, 2)

}
\author{
Samuel Wieczorek, Enora Fremy

Samuel Wieczorek, Thomas Burger
}
\keyword{internal}
