% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{impute_dapar,SummarizedExperiment-method}
\alias{impute_dapar,SummarizedExperiment-method}
\alias{impute_dapar,Features-method}
\title{The function impute_fixed_value is deleted and the dev must use the function impite_with from the package MSCoreUtils}
\usage{
\S4method{impute_dapar}{SummarizedExperiment}(object, method, ...)

\S4method{impute_dapar}{Features}(object, method, ..., i)
}
\arguments{
\item{object}{A `SummarizedExperiment` or `Features` object with
missing values to be imputed.}

\item{method}{`character(1)` defining the imputation method. See
`imputeMethods()` for available ones. See
[MsCoreUtils::impute_matrix()] for details.}

\item{...}{Additional parameters passed to the inner imputation
function. See [MsCoreUtils::impute_matrix()] for details.}

\item{i}{Defines which element of the `Features` instance to
impute. If missing, all assays will be imputed.}
}
\description{
The function impute_fixed_value is deleted and the dev must use the function impite_with from the package MSCoreUtils
}
