% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\docType{data}
\name{impute_dapar}
\alias{impute_dapar}
\alias{impute_dapar,SummarizedExperiment-method}
\alias{impute_dapar,QFeatures-method}
\title{xxxxxx}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
impute_dapar

\S4method{impute_dapar}{SummarizedExperiment}(object, method, ...)

\S4method{impute_dapar}{QFeatures}(object, i, name, method, ...)
}
\arguments{
\item{object}{A `SummarizedExperiment` or `QFeatures` object with
missing values to be imputed.}

\item{method}{`character(1)` defining the imputation method. See
`imputeMethods()` for available ones. See
[MsCoreUtils::impute_matrix()] for details.}

\item{...}{Additional parameters passed to the inner imputation
function. See [MsCoreUtils::impute_matrix()] for details.}

\item{i}{Defines which element of the `QFeatures` instance to
impute. If missing, all assays will be imputed.}

\item{name}{xxx}
}
\value{
NA

NA
}
\description{
xxxxxx
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- impute_dapar(obj, 2,'foo1',  'POV_det_quant', conds=colData(obj)$Condition)

}
\keyword{datasets}
