% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{impute_det_quant}
\alias{impute_det_quant}
\title{xxx}
\usage{
impute_det_quant(x, ...)
}
\arguments{
\item{x}{An instance of class \code{DataFrame}}

\item{...}{Additional arguments}
}
\value{
An imputed instance of class \code{DataFrame}
}
\description{
xxx
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
impute_det_quant(assay(Exp1_R25_pept[[2]]))

}
\author{
Samuel Wieczorek
}
