% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{impute_fixed_value}
\alias{impute_fixed_value}
\title{Missing values imputation from a \code{MSnSet} object}
\usage{
impute_fixed_value(x, value)
}
\arguments{
\item{x}{An assay of a SummarizedExperiment}

\item{value}{A float .}
}
\value{
The object \code{obj} which has been imputed
}
\description{
This method is a wrapper to
objects of class \code{MSnSet} and imputes missing values with a fixed value.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
impute_fixed_value(assay(Exp1_R25_pept[1:1000], 2), 0.001)

}
\author{
Samuel Wieczorek
}
