% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{impute_knn_by_conditions}
\alias{impute_knn_by_conditions}
\title{KNN missing values imputation by conditions}
\usage{
impute_knn_by_conditions(x, conds = NULL, k = 3)
}
\arguments{
\item{k}{the number of neighbors.}

\item{obj}{An object of class \code{MSnSet}.}
}
\value{
The object \code{obj} which has been imputed
}
\description{
This function imputes the missing values condition by condition.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
imp <- impute_knn_by_conditions(assay(obj[[2]]), colData(obj)$Condition, 3)

}
\author{
Samuel Wieczorek
}
