% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{impute_matrix_dapar}
\alias{impute_matrix_dapar}
\title{xxx}
\usage{
impute_matrix_dapar(x, method, ...)
}
\arguments{
\item{x}{xxxx}

\item{method}{xxxx}

\item{...}{xxxxx}
}
\value{
NA
}
\description{
xxx
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[seq_len(1000),]
imp <- impute_matrix_dapar(assay(obj[[2]]), method='knn_by_conds', colData(obj)$Condition, 3)

imp <- impute_matrix_dapar(assay(obj[[2]]), method='pa', colData(obj)$Condition, q.min=0.03)

imp <- impute_matrix_dapar(assay(obj[[2]]), method='det_quant')

imp <- impute_matrix_dapar(assay(obj[[2]]), method='slsa', sampleTab = colData(obj))

}
