% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_PeptideLevel.R
\name{impute_mle_dapar}
\alias{impute_mle_dapar}
\title{Imputation of peptides having no values in a biological condition.}
\usage{
impute_mle_dapar(x, sampleTab)
}
\arguments{
\item{x}{xxxxx}

\item{sampleTab}{xxxx}
}
\value{
The \code{assay(obj)} matrix with imputed values instead of missing values.
}
\description{
Imputation of peptides having no values in a biological condition.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[seq_len(1000),]
sampleTab <- colData(obj)
dat <- impute_mle_dapar(assay(obj[[2]]), colData(obj))

}
\author{
Samuel Wieczorek
}
