% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{impute_pa}
\alias{impute_pa}
\title{Imputation of peptides having no values in a biological condition.}
\usage{
impute_pa(x, conds, q.min = 0.025)
}
\arguments{
\item{q.min}{Same as the function \code{impute.pa} in the package \code{imp4p}}

\item{obj}{An object of class \code{MSnSet}.}
}
\value{
The \code{exprs(obj)} matrix with imputed values instead of missing values.
}
\description{
Imputation of peptides having no values in a biological condition.
}
\examples{
library(Features)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept[1:1000,]
object <- addAssay(object, Features::filterNA(object[[2]],  pNA = 0.2), name='filtered')
imp <- impute_pa(assay(object[['filtered']]), colData(object)$Condition)

}
\author{
Samuel Wieczorek
}
