% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.iter}
\alias{inner.aggregate.iter}
\title{Method to aggregate peptides into proteins with the iterative approach.}
\usage{
inner.aggregate.iter(
  qPepData,
  X,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}

\item{init.method}{The method used to initialize the iterative algotirhm. Default is 'Sum'.}

\item{method}{The method used for xxx. Default value is 'Mean'.}

\item{n}{An integer which is xxx}
}
\value{
A  matrix containing the quantitative aggregated data for proteins.
}
\description{
Method to aggregate peptides into proteins with the iterative approach.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original']][1:1000,])[1:1000,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
X <- BuildListAdjacencyMatrices(PG, names, type=c('All'))
qPepData <- assay(Exp1_R25_pept[['original_log']][1:1000,])
inner.aggregate.iter(qPepData, X$all)
}
\author{
Samuel Wieczorek, Thomas Burger
}
