% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.sum}
\alias{inner.sum}
\title{aggregate peptides into proteins with the sum of the quantitative data per conditions.}
\usage{
inner.sum(qPepData, X)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
A matrix containing the quantitative aggregated data for proteins.
}
\description{
Method to aggregate peptides into proteins with the sum of the quantitative data per conditions.
}
\author{
Samuel Wieczorek
}
