% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_boxplot.R, R/plots_violin.R,
%   R/mod_plots_intensity.R
\name{intensity_plots}
\alias{intensity_plots}
\alias{boxPlotD}
\alias{violinPlotD}
\alias{mod_intensity_plots_ui}
\alias{mod_intensity_plots_server}
\title{Intensity family plots}
\usage{
boxPlotD(object, exp.design, pal.name = NULL, subset = NULL)

violinPlotD(object, exp.design, subset = NULL, pal.name)

mod_intensity_plots_ui(id)

mod_intensity_plots_server(
  id,
  object,
  exp.design,
  meta,
  conds,
  pal.name = NULL,
  params = NULL,
  reset = NULL,
  slave = FALSE
)
}
\arguments{
\item{object}{xxx}

\item{exp.design}{xxx}

\item{pal.name}{xxx}

\item{subset}{A vector of index indicating rows to highlight}

\item{id}{xxx}

\item{meta}{xxx}

\item{conds}{A \code{character()} of the names of conditions
for each sample of the dataset.}

\item{params}{xxx}

\item{reset}{xxx}

\item{slave}{Default is FALSE.}

\item{base_palette}{xxx}
}
\value{
A boxplot

A violinplot

NA

NA
}
\description{
These functions are plotting different views of quantitative data.
Two plots are available:
\itemize{
\item \code{\link[=boxplotD]{boxplotD()}} xxxx,
\item \code{\link[=violinPlotD]{violinPlotD()}} xxxx.
}

A shiny Module.
}
\examples{
library(QFeatures)
data(ft)

#----------------------------------------------
# View violin plot with and without subset view
#----------------------------------------------

violinPlotD(ft[[1]], design(ft))
violinPlotD(ft[[1]], design(ft), subset = c(2, 3), pal.name = 'Dark2')

#----------------------------------------
# xxxx
#----------------------------------------

boxPlotD(ft[[1]], design(ft))
boxPlotD(ft[[1]], design(ft), subset = c(2, 3), pal.name = 'Dark2')






}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy

Samuel Wieczorek, Anais Courtier, Enora Fremy
}
