% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.MV}
\alias{is.MV}
\title{Similar to the function \code{is.na} but focused on the equality with the missing
values in the dataset (type 'POV' and 'MEC')}
\usage{
is.MV(data)
}
\arguments{
\item{data}{A data.frame}
}
\value{
A boolean dataframe
}
\description{
Similar to the function \code{is.na} but focused on the equality with the missing
values in the dataset (type 'POV' and 'MEC')
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept
data <- rowData(Exp1_R25_pept[['original']])[,metadata(Exp1_R25_pept)$OriginOfValues]
is.MV(as.data.frame(data@listData))

}
\author{
Samuel Wieczorek
}
