% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.OfType}
\alias{is.OfType}
\title{Similar to the function \code{is.na} but focused on the equality with the paramter 'type'.}
\usage{
is.OfType(data, type)
}
\arguments{
\item{data}{A data.frame}

\item{type}{The value to search in the dataframe}
}
\value{
A boolean dataframe
}
\description{
Similar to the function \code{is.na} but focused on the equality with the paramter 'type'.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept
data <- rowData(Exp1_R25_pept[['original']])[,metadata(Exp1_R25_pept)$OriginOfValues]
is.OfType(as.data.frame(data@listData), "MEC")

}
\author{
Samuel Wieczorek
}
