% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_agregation.R
\name{inner.aggregate.iter}
\alias{inner.aggregate.iter}
\alias{aggIterParallel}
\alias{aggIterative}
\title{Method to aggregate peptides into proteins with the iterative approach.}
\usage{
inner.aggregate.iter(qData, X, init.method = "sum", iter.method = "mean", ...)

aggIterParallel(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)

aggIterative(
  qData,
  X,
  conditions = NULL,
  init.method = "sum",
  iter.method = "mean",
  n = NULL
)
}
\arguments{
\item{X}{An adjacency matrix}

\item{init.method}{The method used to initialize the iterative algotirhm. Default is 'Sum'.}

\item{qPepData}{A matrix of intensities of peptides.}

\item{conditions}{xxx}

\item{method}{The method used for xxx. Default value is 'Mean'.}

\item{n}{xxxx}
}
\value{
A  matrix containing the quantitative aggregated data for proteins.

A matrix of intensities of proteins

A matrix of protein intensities.
}
\description{
Method to aggregate peptides into proteins with the iterative approach.

Method to aggregate peptides into proteins with the iterative approach with use of parallelism.

Method to aggregate peptides into proteins with the iterative approach
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000),]
obj <- addListAdjacencyMatrices(obj, 2)
X <- GetAdjMat(obj[[2]])$all
qPepData <- assay(obj[[2]])
inner.aggregate.iter(qPepData, X)

library(doParallel)
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000),]
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]])$all)
conditions <- SummarizedExperiment::colData(obj)$Condition
aggIterParallel(assay(obj,2), X, conditions)

ft <- create_ft_example(with.na = TRUE)
X <- adjacencyMatrix(ft[[1]])
conds <- design(ft)$Condition
aggIterative(assay(ft,1), X, conditions=conds)

}
\author{
Samuel Wieczorek, Thomas Burger

Samuel Wieczorek

Samuel Wieczorek
}
