% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepa.R
\name{LH1}
\alias{LH1}
\title{xxxxxx}
\usage{
LH1(X, y1, y2, j)
}
\arguments{
\item{X}{an n.pep*n.prot indicator matrix.}

\item{y1}{n.pep*n.samples matrice giving the observed counts for
each peptide in each sample from the condition 1}

\item{y2}{n.pep*n.samples matrice giving the observed counts for
each peptide in each sample from the condition 2}

\item{j}{the index of the protein being tested, ie which has different
expression in the two conditions under H1}
}
\value{
xxxxxxxxxx..
}
\description{
xxxxxx
}
\author{
Thomas Burger, Laurent Jacob
}
