% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{limmaCompleteTest}
\alias{limmaCompleteTest}
\title{Computes a hierarchical differential analysis}
\usage{
limmaCompleteTest(obj, sampleTab, comp.type = "OnevsOne")
}
\arguments{
\item{obj}{A matrix of quantitative data, without any missing values.}

\item{sampleTab}{A DataFrame of experimental design.}

\item{comp.type}{A string that corresponds to the type of comparison. 
Values are: 'anova1way', 'OnevsOne' and 'OnevsAll'; default is 'OnevsOne'.}
}
\value{
A list of two dataframes : logFC and P_Value. The first one contains
the logFC values of all the comparisons (one column for one comparison), the second one contains
the pvalue of all the comparisons (one column for one comparison). The names of the columns for those two dataframes
are identical and correspond to the description of the comparison.
}
\description{
Computes a hierarchical differential analysis
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
Exp1_R25_pept <- addAssay(Exp1_R25_pept, QFeatures::filterNA(Exp1_R25_pept[[2]],  pNA = 0), name='filtered')
sampleTab <- colData(Exp1_R25_pept)
se <- Exp1_R25_pept[['filtered']]
limma <- limmaCompleteTest(se, sTab, comp.type='anova1way')

}
\author{
Hélène Borges, Thomas Burger, Quentin Giai-Gianetto, Samuel Wieczorek
}
