% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{matAdjStats}
\alias{matAdjStats}
\title{Computes the number of proteins that are only defined by 
specific peptides, shared peptides or a mixture of two.}
\usage{
matAdjStats(X)
}
\arguments{
\item{X}{The adjacency matrix with both specific and shared peptides.}
}
\value{
A list of values:
* nbPeptides: the number of peptides in the matrix,
nbSpecificPeptides: the number of specific peptides in the matrix,
nbSharedPeptides: the number of shared peptides in the matrix,
nbProt: the number of proteins in the matrix,
protOnlyUniquePep: the list of proteins only defined by specific peptides,
protOnlySharedPep: the list of proteins only defined by shared peptides,
protMixPep: the list of proteins defined by both shared and specific peptides.
}
\description{
This function computes few values about the adjacency matrix such as the number of proteins that are only defined by 
specific peptides, shared peptides or a mixture of two.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- GetAdjMat(obj[[2]])$all
matAdjStats(X)

}
\author{
Samuel Wieczorek
}
