% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\name{match.metacell}
\alias{match.metacell}
\title{Similar to the function \code{is.na} but focused on the equality 
with the paramter 'type'.}
\usage{
match.metacell(df, pattern, level)
}
\arguments{
\item{df}{A data.frame}

\item{pattern}{The value to search in the dataframe}

\item{level}{xxx}
}
\value{
A boolean dataframe
}
\description{
Similar to the function \code{is.na} but focused on the equality with 
the paramter 'type'.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:10,]
metadata <- Get_qMetadata(obj)
m <- match.metacell(metadata, pattern="missing", level = 'peptide')

}
\author{
Samuel Wieczorek
}
