% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\name{match.qMetadata}
\alias{match.qMetadata}
\title{Similar to the function \code{is.na} but focused on the equality 
with the paramter 'type'.}
\usage{
match.qMetadata(df, pattern, level)
}
\arguments{
\item{df}{A data.frame}

\item{pattern}{The value to search in the dataframe}

\item{level}{xxx}
}
\value{
A boolean dataframe
}
\description{
Similar to the function \code{is.na} but focused on the equality with 
the paramter 'type'.
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(10),]
metadata <- qMetadata(obj)
m <- match.qMetadata(metadata, "missing", 'peptide')

}
\author{
Samuel Wieczorek
}
