% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{MeanCentering}
\alias{MeanCentering}
\title{Normalisation MeanCentering}
\usage{
MeanCentering(
  qData,
  conds = NULL,
  type = NULL,
  subset.norm = NULL,
  scaling = FALSE
)
}
\arguments{
\item{qData}{A numeric matrix.}
}
\value{
A normalized numeric matrix
}
\description{
Normalisation MeanCentering
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[['original']])
conds <- colData(Exp1_R25_pept)[['Condition']]
normalized <- MeanCentering(qData, conds, type="overall")
}
\author{
Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy
}
