% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{MetacellFilteringScope}
\alias{MetacellFilteringScope}
\title{Quantitative cell metadata scopes for filtering}
\usage{
MetacellFilteringScope()
}
\description{
Quantitative cell metadata scopes for filtering
}
