% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Agregation.R
\name{mod_Agregation_ui}
\alias{mod_Agregation_ui}
\alias{mod_Agregation_server}
\title{Module agregation}
\usage{
mod_Agregation_ui(id)

mod_Agregation_server(
  id,
  dataIn = reactive({     NULL }),
  steps.enabled = reactive({     NULL }),
  remoteReset = reactive({     FALSE }),
  steps.status = reactive({     NULL }),
  current.pos = reactive({     1 }),
  verbose = FALSE
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{steps.enabled}{xxx}

\item{remoteReset}{xxx}

\item{steps.status}{xxx}

\item{current.pos}{xxx}

\item{verbose}{xxx}
}
\description{
Module agregation
}
