% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_all_plots.R
\name{mod_all_plots_server}
\alias{mod_all_plots_server}
\title{all_plots UI Function}
\usage{
mod_all_plots_server(id, dataIn)
}
\arguments{
\item{id}{xxxx}

\item{dataIn}{xxxx}
}
<<<<<<< HEAD
\value{
NA
}
=======
>>>>>>> bd1ab6a739c5a9402b382daddfe0b803a076915c
\description{
A shiny Module.
}
