% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_all_plots.R
\name{mod_all_plots_ui}
\alias{mod_all_plots_ui}
\title{all_plots UI Function}
\usage{
mod_all_plots_ui(id)
}
\arguments{
<<<<<<< HEAD
\item{id, input, output, session}{Internal parameters for {shiny}.}
}
\value{
NA
=======
\item{id}{xxx}
>>>>>>> bd1ab6a739c5a9402b382daddfe0b803a076915c
}
\description{
A shiny Module.
}
