% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert.R
\name{mod_Convert_ui}
\alias{mod_Convert_ui}
\alias{mod_Convert_server}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
mod_Convert_ui(id)

mod_Convert_server(
  id,
  dataIn = NULL,
  steps.enabled = reactive({     NULL }),
  remoteReset = reactive({     FALSE })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{steps.enabled}{xxx}

\item{remoteReset}{xxx}
<<<<<<< HEAD
}
\value{
NA

NA
=======
>>>>>>> bd1ab6a739c5a9402b382daddfe0b803a076915c
}
\description{
A shiny Module.
}
\keyword{internal}
