% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtering_example.R
\name{mod_filterExample}
\alias{mod_filterExample}
\alias{mod_filterExample_ui}
\alias{mod_filterExample_server}
\title{xxx}
\usage{
mod_filterExample_ui(id)

mod_filterExample_server(id, objBefore, objAfter, query)
}
\arguments{
\item{id}{A \code{character(1)}}

\item{objBefore}{An instance of the class \code{SummarizedExperiment}}

\item{objAfter}{xxx}

\item{query}{A \code{character()}}
}
\value{
NA
}
\description{
xxxxx
}
\examples{
if (interactive()) {
    data(ft_na, package='DaparToolshed')
    ui <- mod_filterExample_ui("example")

    server <- function(input, output, session) {
        mod_filterExample_server(
            id = "example",
            objBefore = reactive({ft_na[[1]]}),
            objAfter = reactive({ft_na[[1]][-c(2, 6)]}),
            query = reactive({"query"})
        )
    }

    shinyApp(ui = ui, server = server)
}
}
