% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_filtering_example.R
\name{mod_filtering_example}
\alias{mod_filtering_example}
\alias{mod_filtering_example_ui}
\alias{mod_filtering_example_server}
\title{xxx}
\usage{
mod_filtering_example_ui(id)

mod_filtering_example_server(id, obj, indices, params, txt)
}
\arguments{
\item{id}{xxx}

\item{obj}{xxx}

\item{indices}{xxx}

\item{params}{xxx}

\item{txt}{}
}
\description{
xxxxx
}
\examples{
if(interactive()){



}
}
