% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_format_DT.R
\name{mod_format_DT}
\alias{mod_format_DT}
\alias{mod_format_DT_ui}
\alias{mod_format_DT_server}
\title{mod_format_DT_ui and mod_format_DT_server}
\usage{
mod_format_DT_ui(id)

mod_format_DT_server(id, df, rownames = FALSE, dom = "Bt", style = NULL)
}
\arguments{
\item{id}{internal}

\item{rownames}{xxxx}

\item{dom}{xxx}

\item{style}{xxx}

\item{table2show}{internal}
}
\value{
NA
}
\description{
A shiny Module.
}
