% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_helpPopover.R, R/mod_inputGroup.R
\name{mod_helpPopover}
\alias{mod_helpPopover}
\alias{mod_helpPopover_ui}
\alias{mod_helpPopover_server}
\alias{mod_inputGroup_ui}
\alias{mod_inputGroup_server}
\title{Help popover windows}
\usage{
mod_helpPopover_ui(id)

mod_helpPopover_server(id, title, content)

mod_inputGroup_ui(id)

mod_inputGroup_server(id, df, quantCols)
}
\arguments{
\item{id}{A \code{character(1)} xxx}

\item{title}{A \code{character()} xxx}

\item{content}{A \code{character()} xxx}
}
\value{
NA

NA
}
\description{
A shiny Module.

A shiny Module.
}
\examples{
if (interactive()) {
    ui <- mod_helpPopover_ui("help")

    server <- function(input, output, session) {
        mod_helpPopover_server("help",
            title = "Foo",
            content = "xxx"
        )
    }

    shinyApp(ui = ui, server = server)
}
if (interactive()) {
    ui <- mod_inputGroup_ui("help")

    server <- function(input, output, session) {
        mod_inputGroup_server("help",
            title = "Foo",
            content = "xxx"
        )
    }

    shinyApp(ui = ui, server = server)
}
}
