% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_infos_dataset.R
\name{mod_infos_dataset}
\alias{mod_infos_dataset}
\alias{mod_infos_dataset_ui}
\alias{mod_infos_dataset_server}
\title{mod_infos_dataset_ui and mod_infos_dataset_server}
\usage{
mod_infos_dataset_ui(id)

mod_infos_dataset_server(id, obj)
}
\arguments{
\item{id}{shiny id}

\item{obj}{An instance of the class \code{QFeatures}.}
}
\value{
A shiny app
}
\description{
A shiny Module.
}
\examples{

library(QFeatures)
library(DaparViz)
library(shinyjs)
library(shiny)

data(ft_na, package='DaparViz')

ui <- fluidPage(mod_infos_dataset_ui("mod_info"))

server <- function(input, output, session) {
  mod_infos_dataset_server("mod_info", reactive({ft_na}))
}

shinyApp(ui, server)

}
\keyword{internal}
