% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inputGroup.R
\name{mod_inputGroup}
\alias{mod_inputGroup}
\alias{mod_inputGroup_ui}
\alias{mod_inputGroup_server}
\title{Help popover windows}
\usage{
mod_inputGroup_ui(id)

mod_inputGroup_server(id, df, quantCols)
}
\arguments{
\item{id}{A \code{character(1)} xxx}

\item{df}{xxx}

\item{quanCols}{A vector of}
}
\value{
A shiny app
}
\description{
A shiny Module.
}
\examples{
library(shiny)

    ui <- mod_inputGroup_ui("help")

    server <- function(input, output, session) {
        mod_inputGroup_server("help",
            title = "Foo",
            content = "xxx"
        )
    }

    shinyApp(ui = ui, server = server)
}
