% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_open_dataset.R
\name{mod_open_dataset}
\alias{mod_open_dataset}
\alias{mod_open_dataset_ui}
\alias{mod_open_dataset_server}
\title{mod_open_dataset_ui and mod_open_dataset_server}
\usage{
mod_open_dataset_ui(id)

mod_open_dataset_server(id)
}
\arguments{
\item{id}{xxx}
}
\description{
A shiny Module.
}
\examples{
if (interactive()){
ui <- fluidPage(
tagList(
  mod_open_dataset_ui("qf_file"),
  textOutput('res')
)
)

server <- function(input, output, session) {
  rv <- reactiveValues(
    obj = NULL
  )
  rv$obj <- mod_open_dataset_server("qf_file")
  
  output$res <- renderText({
    rv$obj()
    paste0('Names of the datasets: ', names(rv$obj()))
  })
}

shinyApp(ui, server)
}

}
\keyword{internal}
