% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Protein.R
\name{mod_Protein_ui}
\alias{mod_Protein_ui}
\alias{mod_Protein_server}
\title{Module of Protein pipeline}
\usage{
mod_Protein_ui(id)

mod_Protein_server(
  id,
  nav.mode = "pipeline",
  dataIn = reactive({     NULL }),
  steps.enabled = reactive({     NULL }),
  remoteReset = reactive({     FALSE }),
  status = reactive({     NULL })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{The dataset}

\item{steps.enabled}{A vector of boolen which has the same length of the steps
of the pipeline. xxx}

\item{remoteReset}{It is a remote command to reset the module. A boolen that
indicates is the pipeline has been reseted by a program of higher level
Basically, it is the program which has called this module}

\item{status}{xxx}
}
\value{
NA

NA
}
\description{
This module contains the configuration informations for the corresponding pipeline.
It is called by the nav_pipeline module of the package Magellan
}
\section{Fields}{

\describe{
\item{\code{global}}{xxxx}
}}

\author{
Samuel Wieczorek
}
