% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Protein_Description.R
\name{mod_Protein_Description_ui}
\alias{mod_Protein_Description_ui}
\alias{mod_Protein_Description_server}
\title{Module for the description of pipeline Protein}
\usage{
mod_Protein_Description_ui(id)

mod_Protein_Description_server(
  id,
  dataIn = reactive({     NULL }),
  steps.enabled = reactive({     NULL }),
  remoteReset = reactive({     FALSE })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{The dataset}

\item{steps.enabled}{A vector of boolean which has the same length of the steps
of the pipeline. This information is used to enable/disable the widgets. It is not
a communication variable between the caller and this module, thus there is no
corresponding output variable}

\item{remoteReset}{It is a remote command to reset the module. A boolean that
indicates is the pipeline has been reseted by a program of higher level
Basically, it is the program which has called this module}
}
\value{
NA

NA
}
\description{
This module contains the configuration informations for the corresponding pipeline.
It is called by the nav_pipeline module of the package Magellan
}
\section{Fields}{

\describe{
\item{\code{global}}{xxxx}
}}

\author{
Samuel Wieczorek
}
