% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Protein_Filtering.R
\name{mod_Protein_Filtering_ui}
\alias{mod_Protein_Filtering_ui}
\alias{mod_Protein_Filtering_server}
\title{Module of filtering protein process}
\usage{
mod_Protein_Filtering_ui(id)

mod_Protein_Filtering_server(
  id,
  dataIn = NULL,
  steps.enabled = reactive({     NULL }),
  remoteReset = reactive({     FALSE })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{steps.enabled}{xxx}

\item{remoteReset}{xxx}
}
\value{
NA

NA
}
\description{
#' @title Add an item to the dataset
#' 
#' @param dataset xxx
#' @param name xxx
#' @return the dataset
AddItemToDataset <- function(dataset, name){
  addAssay(dataset,
           dataset[[length(dataset)]],
           name=name)
}
}
\section{Fields}{

\describe{
\item{\code{global}}{xxxx}

\item{\code{config}}{xxxx}
}}

