% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Protein_Normalization.R
\name{mod_Protein_Normalization_server}
\alias{mod_Protein_Normalization_server}
\title{xxx}
\usage{
mod_Protein_Normalization_server(
  id,
  dataIn = NULL,
  steps.enabled = reactive({     NULL }),
  remoteReset = reactive({     FALSE })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{steps.enabled}{xxx}

\item{remoteReset}{xxx}
}
\description{
A shiny Module.
}
\section{Fields}{

\describe{
\item{\code{global}}{xxxx}

\item{\code{config}}{xxxx}
}}

