% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Protein_Normalization.R
\name{mod_Protein_Normalization_ui}
\alias{mod_Protein_Normalization_ui}
\title{#' @export
#' 
#' @return NA
#'
AddItemToDataset <- function(dataset, name){
  addAssay(dataset,
           dataset[[length(dataset)]],
           name=name)
}}
\usage{
mod_Protein_Normalization_ui(id)
}
\description{
#' @export
#' 
#' @return NA
#'
AddItemToDataset <- function(dataset, name){
  addAssay(dataset,
           dataset[[length(dataset)]],
           name=name)
}
}
