% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_mv_histo.R
\name{mvHisto_HC}
\alias{mvHisto_HC}
\title{Histogram of missing values}
\usage{
mvHisto_HC(qData, conds, showValues = FALSE, palette = NULL)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{showValues}{A logical that indicates wether numeric values should be
drawn above the bars.}

\item{palette}{xxx}
}
\value{
A histogram
}
\description{
This method plots a histogram of missing values.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[[2]])
conds <- colData(Exp1_R25_pept)[['Condition']]
mvHisto_HC(qData, conds, showValues=TRUE)

}
\author{
Samuel Wieczorek
}
