% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering_MV.R
\name{MVindicesToZero}
\alias{MVindicesToZero}
\title{Filter lines in the matrix of intensities w.r.t. some criteria}
\usage{
MVindicesToZero(obj, sampleTab, newColName, type, th = 0)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment} containing
quantitative data.}

\item{sampleTab}{xxx}

\item{newColName}{Name of the new column in rowData()}

\item{type}{Method used to choose the lines to delete.
Values are : "None", "EmptyLines", "WholeMatrix", "AllCond", "AtLeastOneCond"}

\item{th}{An integer value of the threshold}
}
\value{
The object of class \code{SummarizedExperiment} with extra column in rowData
indicating 0 for the lines to keep, else 1.
}
\description{
Returns the \code{SummarizedExperiment} object with a extra column in \code{rowData()}.
The extra column indicates feature(s) to delete in 1 and 0 otherwise.
The user chooses the minimum amount of intensities that is acceptable and
the filter tags the lines that do not respect this condition.
The condition may be on the whole line or condition by condition.

The different methods are :
"WholeMatrix": given a threshold \code{th}, only the lines that contain
at least \code{th} values are kept.
"AllCond": given a threshold \code{th}, only the lines which contain
at least \code{th} values for each of the conditions are kept.
"AtLeastOneCond": given a threshold \code{th}, only the lines that contain
at least \code{th} values, and for at least one condition, are kept.
}
\examples{
library(DAPAR2)
library(Features)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept[[2]]
sampleTab <- colData(Exp1_R25_pept)
obj<-MVindicesToZero(obj=object, sampleTab, newColName="LinesKept", type="AllCond", th=2)
head(rowData(obj))

}
\author{
Enora Fremy
}
