% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_mv_PerLinesHistoPerCondition.R
\name{mvPerLinesHistoPerCondition_HC}
\alias{mvPerLinesHistoPerCondition_HC}
\title{Bar plot of missing values per lines and per condition}
\usage{
mvPerLinesHistoPerCondition_HC(
  qData,
  conds,
  indLegend = "auto",
  palette = NULL
)
}
\arguments{
\item{qData}{A dataframe that contains quantitative data.}

\item{conds}{A vector of conditions names.}

\item{indLegend}{The indice of the column names of \code{colData()}}

\item{palette}{A vector of HEX Code colors (one color per unique condition).}
}
\value{
A bar plot
}
\description{
This method plots a bar plot which represents the distribution of the 
number of missing values (NA) per lines (ie proteins) and per conditions.
}
\examples{
library(highcharter)
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- SummarizedExperiment::assay(Exp1_R25_pept[[2]])
conds <- SummarizedExperiment::colData(Exp1_R25_pept)$Condition
mvPerLinesHistoPerCondition_HC(qData, conds)
mvPerLinesHistoPerCondition_HC(qData, conds, palette = c('yellow', 'green'))

}
\author{
Samuel Wieczorek, Enora Fremy
}
