% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nEmptyLines}
\alias{nEmptyLines}
\title{Returns the number of empty lines in the data}
\usage{
nEmptyLines(qData)
}
\arguments{
\item{qData}{A matrix corresponding to the quantitative data.}
}
\value{
A list
}
\description{
Returns the number of empty lines in the data
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
qData <- assay(Exp1_R25_pept,1)
nEmptyLines(qData)

}
\author{
Samuel Wieczorek
}
