% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R,
%   R/plots_compare_normalization_plots.R
\name{normalizeMethods.dapar}
\alias{normalizeMethods.dapar}
\alias{normalizeD,SummarizedExperiment-method}
\alias{normalizeD,QFeatures-method}
\alias{GlobalQuantileAlignment}
\alias{SumByColumns}
\alias{QuantileCentering}
\alias{MeanCentering}
\alias{vsn}
\alias{LOESS}
\alias{compareNormalizationD_HC}
\title{List normalization methods}
\usage{
normalizeMethods.dapar(withTracking = FALSE)

\S4method{normalizeD}{SummarizedExperiment}(object, method, withTracking = FALSE, ...)

\S4method{normalizeD}{QFeatures}(object, i, name = "normalizedAssay", ...)

GlobalQuantileAlignment(qData)

SumByColumns(qData, conds = NULL, type = NULL, subset.norm = NULL)

QuantileCentering(
  qData,
  conds = NULL,
  type = "overall",
  subset.norm = NULL,
  quantile = 0.15
)

MeanCentering(
  qData,
  conds,
  type = "overall",
  subset.norm = NULL,
  scaling = FALSE
)

vsn(qData, conds, type = NULL)

LOESS(qData, conds, type = "overall", span = 0.7)

compareNormalizationD_HC(
  qDataBefore,
  qDataAfter,
  conds,
  keyId = NULL,
  palette = NULL,
  subset.view = NULL,
  n = 100,
  type = "scatter"
)
}
\arguments{
\item{withTracking}{xxx}

\item{object}{An object of class \code{QFeatures}.}

\item{method}{xxxxx. See xxx for methods available.}

\item{...}{Additional parameters passed to inner functions.}

\item{i}{xxxx}

\item{name}{xxxx}

\item{qData}{A numeric matrix.}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{type}{scatter or line}

\item{subset.norm}{A vector of index indicating rows to be used for normalization}

\item{quantile}{A float that corresponds to the quantile used to align the data.}

\item{scaling}{A boolean that indicates if the variance of the data have to
be forced to unit (variance reduction) or not.}

\item{span}{xxx}

\item{qDataBefore}{A numeric matrix containing quantitative data before normalization.}

\item{qDataAfter}{A numeric matrix containing quantitative data after normalization.}

\item{keyId}{xxx}

\item{palette}{xxx}

\item{subset.view}{xxx}

\item{n}{xxx}
}
\value{
A list of methods

NA

A normalized numeric matrix

A normalized numeric matrix

A normalized numeric matrix

A normalized numeric matrix

A normalized numeric matrix

A normalized numeric matrix

A plot
}
\description{
This plot compares the quantitative proteomics data before and after
normalization
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
qData <- assay(Exp1_R25_pept[['original']])
normalized <- GlobalQuantileAlignment(qData)

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
library(QFeatures)
qData <- assay(Exp1_R25_pept[['original_log']])
conds <- colData(Exp1_R25_pept)[["Condition"]]
normalized <- SumByColumns(qData, conds, type="within conditions", subset.norm= seq_len(10))

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[['original_log']]
conds <- colData(Exp1_R25_pept)[['Condition']]
normalized <- QuantileCentering(assay(obj), conds, type="within conditions", subset.norm= seq_len(10))

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
qData <- assay(Exp1_R25_pept[['original_log']])
conds <- colData(Exp1_R25_pept)[['Condition']]
normalized <- MeanCentering(qData, conds, type="overall")

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
qData <- assay(Exp1_R25_pept[['original_log']])
conds <- colData(Exp1_R25_pept)[['Condition']]
normalized <- vsn(qData, conds, type="overall")

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
qData <- assay(Exp1_R25_pept[['original_log']])
conds <- colData(Exp1_R25_pept)[['Condition']]
normalized <- LOESS(qData, conds, type="overall")

library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[1:1000,]
conds <- colData(obj)$Condition
id <- metadata(obj)$keyId
obj <- normalizeD(obj, 2, name='norm', method='SumByColumns', conds=conds, type='overall')
palette <- ExtendPalette(2, "Dark2")
qBefore <- assay(obj, 2)
qAfter <- assay(obj, 3)
compareNormalizationD_HC(qBefore, qAfter, conds=conds, n=100)
compareNormalizationD_HC(qBefore, qAfter, conds=conds, keyId = id, n=100, palette=palette)

}
\author{
Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy

Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy

Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy

Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy

Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy

Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy

Samuel Wieczorek, Enora Fremy
}
