% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\docType{data}
\name{normalizeD}
\alias{normalizeD}
\alias{normalizeD,SummarizedExperiment-method}
\alias{normalizeD,QFeatures-method}
\title{Check the validity of the experimental design.}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
normalizeD

\S4method{normalizeD}{SummarizedExperiment}(object, method, withTracking = FALSE, ...)

\S4method{normalizeD}{QFeatures}(object, i, name = "normalizedAssay", ...)
}
\arguments{
\item{object}{An object of class `QFeatures`.}

\item{method}{xxxxx. See xxx for methods available.}

\item{withTracking}{xxx}

\item{...}{Additional parameters passed to inner functions.}

\item{i}{xxxx}

\item{name}{xxxx}
}
\value{
An instance of class 'SummerizedExperiment' where the quantitative
data in the \code{array()} tab has been normalized.
}
\description{
This manual page describes the computation of statistical test using [QFeatures] objects. In the following
functions, if `object` is of class `QFeatures`, and optional assay
index or name `i` can be specified to define the assay (by name of
index) on which to operate.

The following functions are currently available:

- `GlobalQuantileAlignment` xxx
- `SumByColumns` xxx
- `QuantileCentering` xxx
- `MeanCentering` xxxx
- `LOESS` xxx
- `vsn` xxx
}
\details{
#' Provides several methods to normalize quantitative data from a numeric matrix.
They are organized in six main families : GlobalQuantileAlignement, sumByColumns, QuantileCentering, MeanCentering, LOESS, vsn
For the first family, there is no type.
For the five other families, two type categories are available : 
"Overall" which means that the value for each protein (ie row) is computed over all the samples ;
"within conditions" which means that the value for each protein (ie row) is computed condition by condition.
Normalization can be based only on a subset of protein.
Proteins with NA in this subset are ignored.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept
conds <- colData(object)$Condition
object <- normalizeD(object, i=2, name = "norm", method='SumByColumns', conds=conds, type='overall')

}
\author{
Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy
}
\keyword{datasets}
