% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeMethods.dapar}
\alias{normalizeMethods.dapar}
\title{List normalization methods}
\usage{
normalizeMethods.dapar(withTracking = FALSE)
}
\arguments{
\item{withTracking}{A boolean to indicate methods with tracking feature}
}
\description{
List normalization methods
}
