% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Protein.R, R/mod_Protein_Description.R,
%   R/mod_Protein_Filtering.R, R/mod_Protein_Normalization.R
\name{pipeline-protein}
\alias{pipeline-protein}
\alias{mod_Protein_ui}
\alias{mod_Protein_server}
\alias{mod_Protein_Description_ui}
\alias{mod_Protein_Description_server}
\alias{mod_Protein_Filtering_ui}
\alias{mod_Protein_Filtering_server}
\alias{mod_Protein_Normalization_ui}
\alias{mod_Protein_Normalization_server}
\title{Module of Protein pipeline}
\usage{
mod_Protein_ui(id)

mod_Protein_server(
  id,
  dataIn = reactive({
     NULL
 }),
  steps.enabled = reactive({
     NULL
 }),
  remoteReset = reactive({
     FALSE
 }),
  status = reactive({
     NULL
 })
)

mod_Protein_Description_ui(id)

mod_Protein_Description_server(
  id,
  dataIn = reactive({
     NULL
 }),
  steps.enabled = reactive({
     NULL
 }),
  remoteReset = reactive({
     FALSE
 })
)

mod_Protein_Filtering_ui(id)

mod_Protein_Filtering_server(
  id,
  dataIn = NULL,
  steps.enabled = reactive({
     NULL
 }),
  remoteReset = reactive({
     FALSE
 })
)

mod_Protein_Normalization_ui(id)

mod_Protein_Normalization_server(
  id,
  dataIn = reactive({
     NULL
 }),
  steps.enabled = reactive({
     NULL
 }),
  remoteReset = reactive({
     FALSE
 })
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{The dataset}

\item{steps.enabled}{A vector of boolean which has the same length of the
steps of the pipeline. This information is used to enable/disable the
widgets. It is not a communication variable between the caller and this
module, thus there is no corresponding output variable}

\item{remoteReset}{It is a remote command to reset the module. A boolean that
indicates is the pipeline has been reseted by a program of higher level
Basically, it is the program which has called this module}

\item{status}{xxx}
}
\value{
NA

NA

NA

NA

NA

NA
}
\description{
This module contains the configuration information for the corresponding
pipeline. It is called by the nav_pipeline module of the package MagellanNTK

This module contains the configuration informations for the corresponding
pipeline. It is called by the nav_pipeline module of the package MagellanNTK

This module contains the configuration informations for the corresponding
pipeline. It is called by the nav_pipeline module of the package MagellanNTK
}
\section{Fields}{

\describe{
\item{\code{global}}{xxxx}

\item{\code{global}}{xxxx}

\item{\code{global}}{xxxx}

\item{\code{config}}{xxxx}

\item{\code{config}}{xxxx}
}}

\examples{
if(interactive()){


}
}
\author{
Samuel Wieczorek
}
