% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ds_mv.R
\name{plot_ds_mv}
\alias{plot_ds_mv}
\title{Bar plot of missing values per lines using highcharter.}
\arguments{
\item{qData}{A \code{data.frame} or \code{matrix} that contains the quantitative data.}

\item{conds}{A \code{character()} of condition name for each sample. The
length of 'conds' must be equal to the number of columns of 'qData'.}

\item{pal.name}{A \code{character(1)} which is the name of the palette
(from the package \link{RColorBrewer}) to use.}
}
\value{
A plot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\details{
\itemize{
\item distribution of the missing values per line,
\item a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins) and per conditions,
\item Histogram of missing values.
}
}
\examples{
library(QFeatures)
library(DaparToolshed)
data(ft)

qData <- assay(ft, 1)
conds <- design(ft)$Condition
pal <- 'Dark2'

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvPerLinesHisto(qData)

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvHisto(qData, conds, pal.name = pal)

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvPerLinesHistoPerCondition(qData, conds)
mvPerLinesHistoPerCondition(qData, conds, pal.name = pal)

#----------------------------------------
# Launch a shiny module
#----------------------------------------

if(interactive()){
 library(QFeatures)
 library(shiny)
 library(DaparToolshed)
 data(ft)
 ui <- mod_plot_ds_mv_ui('plot')

 server <- function(input, output, session) {
  conds <- design(ft)$Condition
 
  mod_plot_ds_mv_server('plot',
                      obj = reactive({ft[[1]]}),
                      conds = reactive({conds}),
                      pal.name = reactive({'Dark2'})
                      )
  }
 
 shinyApp(ui=ui, server=server)
}
}
\author{
Samuel Wieczorek, Enora Fremy
}
