% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ds_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{Compute the PCA}
\value{
A list including eigenvalues of obj
}
\description{
xxxx
}
\section{xxx}{







}

\examples{
library(QFeatures)
data(ft)
qData <- assay(ft, 1)
conds <- design(ft)$Condition

res.pca <- wrapper_pca(qData, conds)

plotPCA_Eigen(res.pca)

#-----------------------------------
# Launch the shiny module
#-----------------------------------

if(interactive()){
 library(QFeatures)
 library(shiny)
 library(DaparToolshed)
 library(highcharter)
 data(ft)
 conds <- design(ft)$Condition
 
  
 ui <- mod_plot_ds_pca_ui('plot')

 server <- function(input, output, session) {
  mod_plot_ds_pca_server(id = 'plot',
                       object = reactive({ft[[1]]}),
                       conds = reactive({conds})
                       )
  }
 
 shinyApp(ui, server)
}
}
\author{
Samuel Wieczorek, Enora Fremy
}
