% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{plotJitter_hc}
\alias{plotJitter_hc}
\title{Display a a jitter plot for CC}
\usage{
plotJitter_hc(df, clickFunction = NULL)
}
\arguments{
\item{df}{A dataframe containing xxx}

\item{clickFunction}{A JavaScript code to show an information when a point is clicked}
}
\value{
A plot
}
\description{
Display a jitter plot for connex component with the package highcharter
}
\examples{
\dontrun{
}

}
\author{
Thomas Burger, Samuel Wieczorek
}
