% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_pca.R
\name{plotPCA_Ind}
\alias{plotPCA_Ind}
\title{Plots individuals of PCA}
\usage{
plotPCA_Ind(res.pca, chosen.axes = c(1, 2))
}
\arguments{
\item{res.pca}{Result of FactoMineR::PCA}

\item{chosen.axes}{The dimensions to plot}
}
\value{
A plot
}
\description{
Plots the individuals of PCA
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[['original']])
res.pca <- wrapper.pca(qData)
plotPCA_Ind(res.pca)
}
\author{
Samuel Wieczorek, Enora Fremy
}
