% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_pca.R
\docType{data}
\name{plotPCA_Var}
\alias{plotPCA_Var}
\title{Plots variables of PCA}
\format{
An object of class \code{function} of length 1.
}
\usage{
plotPCA_Var
}
\arguments{
\item{res.pca}{Result of FactoMineR::PCA}

\item{chosen.axes}{The dimensions to plot}
}
\value{
A plot
}
\description{
Plots variables of PCA
}
\author{
Samuel Wieczorek, Enora Fremy
}
\keyword{datasets}
