% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{POV_impute_det_quant}
\alias{POV_impute_det_quant}
\title{Imputation of the POV with the method of deterministic Quantile}
\usage{
POV_impute_det_quant(x, sampleTab, ...)
}
\arguments{
\item{x}{An instance of class \code{DataFrame}}

\item{sampleTab}{xxx}

\item{...}{Additional arguments}
}
\value{
An imputed instance of class \code{DataFrame}
}
\description{
Imputation of the POV with the method of deterministic Quantile
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
POV_impute_det_quant(assay(Exp1_R25_pept[[2]]), colData(Exp1_R25_pept))

}
\author{
Samuel Wieczorek
}
