% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{POV_impute_knn_by_conditions}
\alias{POV_impute_knn_by_conditions}
\title{KNN missing values imputation by conditions}
\usage{
POV_impute_knn_by_conditions(x, conds = NULL, k = 3)
}
\arguments{
\item{x}{xxxx}

\item{conds}{xxx}

\item{k}{the number of neighbors.}
}
\value{
The object \code{obj} which has been imputed
}
\description{
This function imputes the missing values condition by condition.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
imp <- POV_impute_knn_by_conditions(assay(obj[[2]]), colData(obj)$Condition, 3)

}
\author{
Samuel Wieczorek
}
