% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{POV_impute_slsa}
\alias{POV_impute_slsa}
\title{Imputation of peptides having no values in a biological condition.}
\usage{
POV_impute_slsa(x, sampleTab)
}
\arguments{
\item{x}{A DataFrame.}

\item{sampleTab}{xxxxx}
}
\value{
A DataFrame
}
\description{
Imputation of peptides having no values in a biological condition.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[seq_len(1000)]
imp <- POV_impute_slsa(assay(obj[[2]]), colData(obj))

}
\author{
Samuel Wieczorek
}
