% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_QFeatures_Addons.R
\name{QFeatures-accessors}
\alias{QFeatures-accessors}
\alias{qMetadata,QFeatures-method}
\alias{qMetadata,SummarizedExperiment-method}
\alias{qMetadata<-}
\alias{.GetMetadataSlot}
\alias{.GetRowdataSlot}
\alias{typeDataset,QFeatures-method}
\alias{typeDataset,SummarizedExperiment-method}
\alias{typeDataset<-}
\alias{idcol,QFeatures-method}
\alias{idcol,SummarizedExperiment-method}
\alias{idcol<-}
\alias{parentProtId,QFeatures-method}
\alias{parentProtId,SummarizedExperiment-method}
\alias{parentProtId<-}
\alias{analysis,QFeatures-method}
\alias{analysis,SummarizedExperiment-method}
\alias{analysis<-}
\alias{version,QFeatures-method}
\alias{version<-}
\alias{design.qf,QFeatures-method}
\alias{design.qf<-}
\alias{mainAssay}
\alias{params,QFeatures-method}
\alias{params,SummarizedExperiment-method}
\alias{params<-}
\title{Additional accessors for instances of class \code{Qfeatures}.}
\usage{
\S4method{qMetadata}{QFeatures}(object, i, slotName = "qMetadata")

\S4method{qMetadata}{SummarizedExperiment}(object, slotName = "qMetadata")

qMetadata(object, i, slotName = "qMetadata") <- value

.GetMetadataSlot(object, slotName = NULL)

.GetRowdataSlot(object, slotName = NULL)

\S4method{typeDataset}{QFeatures}(object, i, slotName = "typeDataset")

\S4method{typeDataset}{SummarizedExperiment}(object, slotName = "typeDataset")

typeDataset(object, i, slotName = "typeDataset") <- value

\S4method{idcol}{QFeatures}(object, i, slotName = "idcol")

\S4method{idcol}{SummarizedExperiment}(object, slotName = "idcol")

idcol(object, i, slotName = "idcol") <- value

\S4method{parentProtId}{QFeatures}(object, i, slotName = "parentProtId")

\S4method{parentProtId}{SummarizedExperiment}(object, slotName = "parentProtId")

parentProtId(object, i, slotName = "parentProtId") <- value

\S4method{analysis}{QFeatures}(object, i, slotName = "analysis")

\S4method{analysis}{SummarizedExperiment}(object, slotName = "analysis")

analysis(object, i, slotName = "analysis") <- value

\S4method{version}{QFeatures}(object, slotName = "version")

version(object, slotName = "version") <- value

\S4method{design.qf}{QFeatures}(object, slotName = "design")

design.qf(object, slotName = "design") <- value

mainAssay(object)

\S4method{params}{QFeatures}(object, i, slotName = "params")

\S4method{params}{SummarizedExperiment}(object, slotName = "params")

params(object, i, slotName = "params") <- value
}
\arguments{
\item{object}{An instance of class \code{SummarizedExperiment} or \code{QFeatures}.}

\item{i}{The index or name of the assays to extract the quantitative
metadata from. All must have a rowdata variable named as \code{slotName}}

\item{slotName}{xxx}

\item{value}{xxx}
}
\value{
NA

NA

NA

NA

NA

NA
}
\description{
These names are common to all assays contained in the object. This is why
they are stored in the global metadata. This function is used whenever it i
s necessary to (re)detect MEC and POV (new dataset or when post processing
protein qMetadata after aggregation)
}
\details{
Additional slots for rowdata of a \code{SummarizedExperiment} object:
\itemize{
\item qMetadata: xxx
}

Additional slots for Metadata for a \code{QFeatures} object:
\itemize{
\item xxx: xxxx
}

Additional slots for Metadata for a \code{SummarizedExperiment} object:
\itemize{
\item qMetadata: xxxx
\item parentProtId: xxx
\item idcol: xxxx
\item typeDataset: xxx
}
}
\section{Quantitative metadata}{


Default slotName is \code{"qMetadata"}.
The value is an adjacency matrix with row and column names. The
matrix will be coerced to compressed, column-oriented sparse
matrix (class \code{dgCMatrix}) as defined in the \code{Matrix} package,
as generaled by the \code{\link[=sparseMatrix]{sparseMatrix()}} constructor.
}

\examples{
data(ft, package='DaparToolshed')
design.qf(ft)
}
