% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_qfeatures.R
\name{QFeatures-excel}
\alias{QFeatures-excel}
\alias{write2excel,QFeatures-method}
\alias{write2excel,SummarizedExperiment-method}
\title{Exports a \code{QFeatures} object to a Excel file.}
\usage{
\S4method{write2excel}{QFeatures}(object, i = NULL, filename = "newFile", ...)

\S4method{write2excel}{SummarizedExperiment}(object, filename, exp.design, ...)
}
\arguments{
\item{object}{An object of class \code{QFeatures}.}

\item{i}{xxx}

\item{filename}{A character string for the name of the Excel file.}

\item{exp.design}{xxx}
}
\value{
A Excel file.
}
\description{
This function exports an instance of the class \code{QFeatures} to a Excel file.
The resulting file is composed of four sheets:
\itemize{
\item \verb{quantitative data} which contains the content of \code{\link[=assay]{assay()}} object whith a
color code for each cell w.r.t. to cell quantitative metadata.
\item \code{metadata} which is the content of \code{\link[=rowData]{rowData()}} with only one-dimensionnal
data (i.e. the adjacencyMatrix and the qMetadata slots are not part of
the sheet),
\item \verb{exp. design} which is the content of \code{\link[=colData]{colData()}}. Each condition in the table
is colored with a different color,
\item \verb{quantitative metadata} which is the content of \code{\link[=qMetadata]{qMetadata()}}. There is a color
code for the different tags.
}
}
\examples{
\donttest{
data(ft)

#---------------------------------------
# Export the whole dataset
#---------------------------------------

write2excel(ft, filename = 'foo')
write2excel(ft, 1, 'foo')
}

}
\author{
Samuel Wieczorek
}
