% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFeatures_utils.R
\name{QFeatures-utils}
\alias{QFeatures-utils}
\alias{last_assay}
\alias{n_assays_in_qf}
\title{Utility funcitons to dela with QFeatures objects.}
\usage{
last_assay(object)

n_assays_in_qf(object)
}
\arguments{
\item{object}{An instance of the class \code{QFeatures}}
}
\value{
NA
}
\description{
xxxxx
}
\examples{

xxxx

}
