% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R, R/aggregation_func.R
\name{qMetadata,QFeatures-method}
\alias{qMetadata,QFeatures-method}
\alias{qMetadata<-}
\alias{aggregateQmetadata,QFeatures-method}
\alias{aggregateQmetadata,SummarizedExperiment-method}
\alias{aggQmeta}
\alias{FinalizeAggregation,QFeatures-method}
\title{Finalizes the aggregation process}
\usage{
\S4method{qMetadata}{QFeatures}(object, i, qMetaName = "qMetadata")

qMetadata(object, i, qMetaName = "qMetadata") <- value

\S4method{aggregateQmetadata}{QFeatures}(object, i, fcol, name = "newAssay", fun = aggQmeta, ...)

\S4method{aggregateQmetadata}{SummarizedExperiment}(object, fun, conds)

aggQmeta(object, conds)

\S4method{FinalizeAggregation}{QFeatures}(object, from, to, ...)
}
\arguments{
\item{object}{An object of class 'SummarizedExperiment'}

\item{i}{When adding an adjacency matrix to an assay of a
`QFeatures` object, the index or name of the assay the
adjacency matrix will be added to. Ignored when `x` is an
`SummarizedExperiment`.}

\item{qMetaName}{`character(1)` with the variable name containing
the adjacency matrix. Default is `"qMetadata"`.}

\item{value}{An adjacency matrix with row and column names. The
matrix will be coerced to compressed, column-oriented sparse
matrix (class `dgCMatrix`) as defined in the `Matrix` package,
as generaled by the [sparseMatrix()] constructor.}

\item{conds}{xxx}

\item{obj.pep}{A peptide object of class \code{MSnset}}

\item{pepData}{xxxx}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{protData}{xxxxx}

\item{protMetacell}{xxx}
}
\value{
NA

A protein object of class \code{MSnset}
}
\description{
Finalizes the aggregation process
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[1:100]
FinalizeAggregation(obj, 2)

}
\author{
Samuel Wieczorek
}
