% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\docType{data}
\name{qMetadata}
\alias{qMetadata}
\alias{qMetadata,SummarizedExperiment-method}
\alias{qMetadata,QFeatures-method}
\title{Get the quantitative cell metadata}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
qMetadata

\S4method{qMetadata}{SummarizedExperiment}(object, qMetaName = "qMetadata")

\S4method{qMetadata}{QFeatures}(object, i, qMetaName = "qMetadata")
}
\arguments{
\item{object}{An object of class 'QFeatures'}

\item{i}{xxx}
}
\value{
NA

NA
}
\description{
These names are common to all assays contained in the object. This is why
they are stored in the global metadata. This function is used whenever it is necessary
to (re)detect MEC and POV (new dataset or when post processing protein qMetadata 
after aggregation)
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
qMetadata(Exp1_R25_pept)

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
