% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\name{qMetadata_maxquant}
\alias{qMetadata_maxquant}
\title{Sets the quantitative metadata dataframe for maxquant datasets}
\usage{
qMetadata_maxquant(qdata, conds, df, level = NULL)
}
\arguments{
\item{qdata}{A matrix of quantitative data}

\item{conds}{xxx}

\item{df}{A list of integer xxxxxxx}

\item{level}{xxx}
}
\value{
xxxxx
}
\description{
Initial conversion rules for maxquant
|------------|-----------------------|--------|
| Quanti     |     Identification    |    Tag |
|------------|-----------------------|--------|
|  == 0			|       whatever 				|    2.0 |
|  > 0				|       'By MS/MS'			|    1.1 |
|  > 0				|      'By matching'		|    1.2 |
|  > 0				|       unknown col			|    1.0 |
|------------|-----------------------|--------|
}
\examples{
file <- system.file("extdata", "Exp1_R25_pept.txt", package="DaparToolshedData")
data <- read.table(file, header=TRUE, sep="\t",stringsAsFactors = FALSE)
metadataFile <- system.file("extdata", "samples_Exp1_R25.txt", 
package="DaparToolshedData")
metadata <- read.table(metadataFile, header=TRUE, sep="\t", as.is=TRUE, 
stringsAsFactors = FALSE)
conds <- metadata$Condition
qdata <- data[seq_len(10),seq(56, 61)]
df <- data[seq_len(10) , seq(43, 48)]
df2 <- qMetadata_maxquant(qdata, conds, df, 'peptide')

}
\author{
Samuel Wieczorek
}
