% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\name{qMetadata_proline}
\alias{qMetadata_proline}
\title{Sets the qMetadata dataframe}
\usage{
qMetadata_proline(qdata, conds, df, level = NULL)
}
\arguments{
\item{qdata}{A matrix of quantitative data}

\item{conds}{xxx}

\item{df}{A list of integer xxxxxxx}

\item{level}{xxx}
}
\value{
xxxxx
}
\description{
In the quantitative columns, a missing value is identified by no value rather
than a value equal to 0. 
Conversion rules
Initial conversion rules for maxquant
|--------------|-----------------|-----|
| Quanti       |    PSM count    | Tag |
|--------------|-----------------|-----|
|  == 0 | N.A.	|   whatever 			| 2.0 |
|  > 0		  		|    > 0		     	| 1.1 |
|  > 0		  		|    == 0	      	| 1.2 |
|  > 0		  		|   unknown col   | 1.0 |
|--------------|-----------------|-----|
}
\examples{
\donttest{ 
file <- system.file("extdata", "Exp1_R25_pept.txt", package="DaparToolshedData")
data <- read.table(file, header=TRUE, sep="\t",stringsAsFactors = FALSE)
metadataFile <- system.file("extdata", "samples_Exp1_R25.txt", 
package="DaparToolshedData")
metadata <- read.table(metadataFile, header=TRUE, sep="\t", as.is=TRUE, 
stringsAsFactors = FALSE)
conds <- metadata$Condition
qdata <- data[seq_len(100), seq(56, 61)]
df <- data[seq_len(100) , seq(43, 48)]
df <- qMetadata_proline(qdata, conds, df, 'peptide')
}

}
\author{
Samuel Wieczorek
}
