% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filter_qMetadata.R
\name{qMetadataFilteringScope}
\alias{qMetadataFilteringScope}
\title{Quantitative cell metadata scopes for filtering}
\usage{
qMetadataFilteringScope()
}
\value{
NA
}
\description{
Quantitative cell metadata scopes for filtering
}
