% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filter_qMetadata.R
\name{qMetadataOnConditions}
\alias{qMetadataOnConditions}
\title{Search lines which respects request on one or more conditions.}
\usage{
qMetadataOnConditions(object, cmd, mode, pattern, conds, percent, operator, th)
}
\arguments{
\item{conds}{xxx}

\item{percent}{xxx}

\item{th}{The theshold to apply}

\item{mask}{xxx}

\item{type}{Available values are:
* 'AllCond' (the query is valid in all the conditions),
* 'AtLeatOneCond' (the query is valid in at leat one condition.}

\item{op}{String for operator to use. List of operators is available with SymFilteringOperators().}
}
\value{
NA

#' @examples
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(10)]
level <- obj@experimentData@other$typeOfData
pattern <- 'missing'
mask <- match.qMetadata(metadata=qMetadata(obj), pattern=pattern, level=level)
type <- 'AllCond'
conds <- Biobase::pData(obj)$Condition
op <- '>='
th <- 2
percent <- FALSE
ind <- GetIndices_BasedOnConditions(mask, type, conds, percent, op, th)
}
\description{
This function looks for the lines that respect the request in either all conditions
or at least one condition.
}
