% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filter_qMetadata.R
\name{qMetadataWholeLine}
\alias{qMetadataWholeLine}
\title{Search lines which respects query on all their elements.}
\usage{
qMetadataWholeLine(object, cmd, pattern)
}
\arguments{
\item{mask}{xxx}
}
\value{
NA
}
\description{
This function looks for the lines where each element respect the query.
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[20:30]
level <- obj@experimentData@other$typeOfData
pattern <- 'missing POV'
mask <- match.qMetadata(metadata=GetqMetadata(obj), pattern=pattern, level=level)
ind <- GetIndices_WholeLine(mask)

}
