% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filter_qMetadata.R
\name{qMetadatWholeMatrix}
\alias{qMetadatWholeMatrix}
\title{Search lines which respects request on one or more conditions.}
\usage{
qMetadatWholeMatrix(object, cmd, pattern, percent, th, operator)
}
\arguments{
\item{percent}{A boolean to indicate whether the threshold represent an absolute value (percent = FALSE) or
a percentage (percent=TRUE).}

\item{th}{A floating number which is in the interval [0, 1]}

\item{mask}{xxx}

\item{op}{String for operator to use. List of operators is available with SymFilteringOperators().}
}
\value{
NA
}
\description{
This function looks for the lines that respect the request in either all conditions
or at least one condition.
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(10)]
level <- obj@experimentData@other$typeOfData
pattern <- 'missing'
mask <- match.qMetadata(metadata=Get_qMetadata(obj), pattern=pattern, level=level)
percent <- FALSE
th <- 3
op <- '>='
ind <- GetIndices_WholeMatrix(mask, op, percent, th)

}
