% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{QuantileCentering}
\alias{QuantileCentering}
\title{Normalisation QuantileCentering}
\usage{
QuantileCentering(
  qData,
  conds = NULL,
  type = NULL,
  subset.norm = NULL,
  quantile = 0.15
)
}
\arguments{
\item{qData}{A numeric matrix.}

\item{type}{"overall" (shift all the sample distributions at once) or
"within conditions" (shift the sample distributions within each condition at a time).}
}
\value{
A normalized numeric matrix
}
\description{
Normalisation QuantileCentering
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[['original']])
conds <- colData(Exp1_R25_pept)[['Condition']]
normalized <- QuantileCentering(qData, conds, type="overall")
}
\author{
Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy
}
