% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_query_metacell.R
\name{mod_query_metacell_ui}
\alias{mod_query_metacell_ui}
\alias{mod_query_metacell_server}
\title{xxxx}
\usage{
mod_query_metacell_ui(id)

mod_query_metacell_server(
  id,
  obj,
  list_tags = reactive({     NULL }),
  keep_vs_remove = reactive({     NULL }),
  filters = reactive({     NULL }),
  val_vs_percent = reactive({     NULL }),
  operator = reactive({     NULL }),
  reset = reactive({     NULL })
)
}
\arguments{
\item{id}{xxx}

\item{obj}{xxx}

\item{list_tags}{xxx}

\item{keep_vs_remove}{xxx}

\item{filters}{xxx}

\item{val_vs_percent}{xxx}

\item{operator}{xxx}

\item{reset}{xxx}
}
\description{
xxx
}
