% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_query_metacell.R
\name{query_qMetadata}
\alias{query_qMetadata}
\alias{mod_query_qMetadata_ui}
\title{xxxx}
\usage{
mod_query_qMetadata_ui(id)
}
\arguments{
\item{id}{xxx}
}
\description{
xxx
}
\examples{
if (interactive()){
data(ft_na)
 ui <- mod_query_qMetadata_ui('query')

 server <- function(input, output, session) {
 ll.tags <- c('None' = 'None', qMetadata.def(typeDataset(ft_na[[1]]))$node)
  mod_query_qMetadata_server('query', 
                             obj = reactive({ft[[1]]}),
                             list_tags = reactive({ll.tags}),
                             keep_vs_remove = reactive({setNames(nm = c("delete", "keep"))}),
                             filters = reactive({c("None" = "None",
                             "Whole Line" = "WholeLine",
                             "Whole matrix" = "WholeMatrix",
                             "For every condition" = "AllCond",
                             "At least one condition" = "AtLeastOneCond")}),
                             val_vs_percent = reactive({setNames(nm=c('Count', 'Percentage'))}),
                             operator = reactive({setNames(nm = SymFilteringOperators())})
                             )
                             }
 
 shinyApp(ui=ui, server=server)
}
}
