% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering_MV.R
\name{removeAdditionalCol}
\alias{removeAdditionalCol}
\title{Filter lines in the matrix of intensities w.r.t. some criteria}
\usage{
removeAdditionalCol(obj, newCol_name)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment} containing
quantitative data.}

\item{newCol_name}{}
}
\value{
An vector of indices that correspond to the lines to keep.
}
\description{
Returns the \code{SummarizedExperiment} object with a extra column in \code{rowData()}.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept[[2]]
removeAdditionalCol(obj=object, newCol="plop")
}
\author{
Enora Fremy
}
