% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReplaceSpecialChars}
\alias{ReplaceSpecialChars}
\title{Standardize names}
\usage{
ReplaceSpecialChars(x)
}
\arguments{
\item{x}{A \code{character()} to be processed}
}
\value{
A \code{character()} of the same length as 'x' with modified
names.
}
\description{
Replace ".", ' ', '-' in \code{character()} by '_' to be compliant
with functions of \code{Shinyjs}, \code{Shiny}
}
\examples{

ReplaceSpecialChars(c("foo.1", "foo-2", "foo 3"))

}
\author{
Samuel Wieczorek
}
