% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\name{search.qMetadata.tags}
\alias{search.qMetadata.tags}
\title{Search pattern in qMetadata vocabulary}
\usage{
search.qMetadata.tags(pattern, level, depth = "1")
}
\arguments{
\item{pattern}{The string to search.}

\item{level}{The available levels are : names()}

\item{depth}{xxx}
}
\value{
NA
}
\description{
Gives all the tags of the metadata vocabulary containing the pattern 
(parent and all its children).
}
\examples{
search.qMetadata.tags('missing POV', 'peptide')
search.qMetadata.tags('quanti', 'peptide')

}
\author{
Samuel Wieczorek
}
