% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\name{Set_POV_MEC_tags}
\alias{Set_POV_MEC_tags}
\title{Sets the MEC tag in the qMetadata}
\usage{
Set_POV_MEC_tags(conds, df, level)
}
\arguments{
\item{conds}{xxx.}

\item{df}{xxx}

\item{level}{xxx}
}
\value{
An instance of class \code{MSnSet}.
}
\description{
This function is based on the qMetadata dataframe to look for either missing
values (used to update an initial dataset) or imputed values (used when
post processing protein qMetadata after aggregation)
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
object <- Exp1_R25_pept
conds <- colData(object)$Condition
df <- assay(Exp1_R25_pept, 2)
level <- GetTypeDataset(object, 2)
df <- Set_POV_MEC_tags(object, 1, level)

}
\author{
Samuel Wieczorek
}
