% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2_features.R
\name{setMEC}
\alias{setMEC}
\title{Sets the MEC tag in the OriginOfValues}
\usage{
setMEC(origin, qData, conds = NULL)
}
\arguments{
\item{origin}{xxxxx}

\item{qData}{xxxxxx}

\item{conds}{xxxxxxx}
}
\value{
An instance of class \code{DataFrame} which is an update of the parameter names origin
}
\description{
Sets the MEC tag in the origin parameter
}
\examples{
\dontrun{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept
}

}
\author{
Samuel Wieczorek
}
