% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\docType{data}
\name{SetTypeDataset}
\alias{SetTypeDataset}
\alias{SetTypeDataset,SummarizedExperiment-method}
\alias{SetTypeDataset,QFeatures-method}
\title{Get the type of dataset}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
SetTypeDataset

\S4method{SetTypeDataset}{SummarizedExperiment}(object, type, ...)

\S4method{SetTypeDataset}{QFeatures}(object, i, type, ...)
}
\arguments{
\item{object}{An object of class 'QFeatures'}

\item{type}{xxx}

\item{i}{xxx}
}
\value{
NA

NA
}
\description{
Get the type of dataset
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
SetTypeDataset(Exp1_R25_pept[[2]])
}
\author{
Samuel Wieczorek
}
\keyword{datasets}
