% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{standardiseMeanIntensities}
\alias{standardiseMeanIntensities}
\title{Average and standardize protein/peptide abundances for each condition studied}
\usage{
standardiseMeanIntensities(obj, sTab)
}
\arguments{
\item{obj}{SummarizedExperiment object containing all the data}

\item{sTab}{xxx}
}
\value{
a dataframe in wide format providing (in the case of 3 or more
conditions) the standardized means of intensities for each protein/peptide
in each condition. If there are less than 3 conditions, standardization
cannot be performed and only the means of intensities are returned.
}
\description{
Calculate the average of the abundances for each protein in
each condition for an ExpressionSet or MSnSet. Needs to have the array
expression data ordered in the same way as the phenotype data (columns of
the array data in the same order than the condition column in the phenotype
data).
}
\examples{
\dontrun{
library(QFeatures)
utils::data(Exp1_R25_prot, package='DAPARdata2')
obj <- Exp1_R25_prot
sTab <- colData(obj)
standards <- standardiseMeanIntensities(obj[[2]], sTab)
}

}
\author{
Hélène Borges
}
