% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{submatadj}
\alias{submatadj}
\title{Substract part of a adjacency matrix}
\usage{
submatadj(X, onlyShared = FALSE, onlySpec = FALSE)
}
\arguments{
\item{X}{xxx}

\item{onlyShared}{xxx}

\item{onlySpec}{xxx}
}
\value{
A matrix
}
\description{
Substract part of a adjacency matrix
}
\examples{
obj <- readRDS('~/GitHub/DaparToolshedData/data/Exp2_R100_pept.rda')
X <-rowData(obj[[2]])$adjacencyMatrix
X <- X[1:6, 1:6]
X[2,3] <- 2
X[3,5] <- 2
X[6, 1:2] <- 3
X[5, 4] <- 3
X

}
\author{
Samuel Wieczorek
}
