% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{SumByColumns}
\alias{SumByColumns}
\title{Normalisation SumByColumns}
\usage{
SumByColumns(qData, conds = NULL, type = NULL, subset.norm = NULL)
}
\arguments{
\item{qData}{A numeric matrix.}

\item{type}{"overall" (shift all the sample distributions at once) or
"within conditions" (shift the sample distributions within each condition at a time).}
}
\value{
A normalized numeric matrix
}
\description{
Normalisation SumByColumns
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[['original']])
conds <- colData(Exp1_R25_pept)[["Condition"]]
normalized <- SumByColumns(qData, conds, type="overall")
}
\author{
Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy
}
