% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filter_qMetadata.R
\name{SymFilteringOperators}
\alias{SymFilteringOperators}
\title{Operators for complex queries}
\usage{
SymFilteringOperators()
}
\value{
NA
}
\description{
Operators for complex queries
}
