% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{test.design}
\alias{test.design}
\title{Check if xxxxxx}
\usage{
test.design(tab)
}
\arguments{
\item{tab}{A DataFrame which correspond to the samples definition. This DataFrame contains three columns: 
"Bio.Rep" , "Tech.Rep" and "Analyt.Rep"}
}
\value{
A list of two items:
* xxxxxx
* xxxxxxx
}
\description{
Check if xxxxxx
}
\examples{
library(Features)
utils::data(Exp1_R25_pept, package='DAPARdata2')
test.design(colData(Exp1_R25_pept)[,1:3])

}
\author{
Thomas Burger, Samuel Wieczorek
}
