% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filterAdjmat.R
\name{topn}
\alias{topn}
\title{xxxxx}
\usage{
topn(object, ...)
}
\arguments{
\item{qData}{xxx}

\item{X}{xxx}

\item{fun}{xxx}

\item{n}{xxx}
}
\description{
xxx
}
\details{
This function builds an intermediate matrix with scores for each peptide
based on 'fun' parameter. Once this matrix is built, one select the 'n' peptides
which have the higher score

- rowMedians xxx
- rowMeans xxx
- rowSums xxx
}
\examples{

}
