% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots_tracking.R
\name{tracking}
\alias{tracking}
\alias{mod_tracking_ui}
\alias{mod_tracking_server}
\title{Tracking of entities within plots}
\usage{
mod_tracking_ui(id)

mod_tracking_server(id, object)
}
\arguments{
\item{id}{xxx}

\item{object}{A instance of the class \code{SummarizedExperiment}}

\item{reset}{xxx}
}
\value{
NA

A \code{list()} of integers
}
\description{
This shiny module offers a UI to select a subset of a dataset
and superimpose quantitative values of this selection on the complete plot
Three modes of selection are implemented:
\itemize{
\item 'Protein list': xxx,
\item 'Random': xxx,
\item 'Specific column': xxx
}
}
\examples{
if (interactive()){
library(QFeatures)
library(DaparToolshed)
data(ft)
ui <- mod_tracking_ui('track')

server <- function(input, output, session) {
 mod_tracking_server(id = 'track',
                     object = reactive({ft[[1]]})
                     )
 }
shinyApp(ui, server)
}
}
