% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_func.R
\name{updateAdjacencyMatrix}
\alias{updateAdjacencyMatrix}
\title{xxx}
\usage{
updateAdjacencyMatrix(X, mode = "all", ...)
}
\arguments{
\item{X}{xxx}

\item{mode}{xxx}

\item{...}{xxx}

\item{se}{xxx}
}
\description{
xxx
}
\details{
Mode can be
- all xxxx
- onlySpec xxx
- onlyShared xxx
- topn xxx
}
\examples{
feat2 <- readRDS('~/GitHub/DaparToolshedData/data/Exp2_R100_pept.rda')
feat2 <- feat2[1:10,]
X <- makeAdjacencyMatrix(rowData(feat2[['original_log']])[,'Protein_group_IDs'])
rownames(X) <- rownames(feat2[['original_log']])
updateAdjacencyMatrix(feat2[['original_log']], X, mode = 'all')

}
