% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\docType{data}
\name{UpdateMetacell}
\alias{UpdateMetacell}
\title{Update metacell after imputation}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
UpdateMetacell
}
\value{
NA
}
\description{
Update the metacell information of missing values that were imputed
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:10,]
obj[[2]] <- UpdateMetacell(obj[[2]], na.type = 'missing')

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
