% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qMetadata.R
\docType{data}
\name{UpdateqMetadata}
\alias{UpdateqMetadata}
\alias{UpdateqMetadata,SummarizedExperiment-method}
\title{Update quantitative metadata after imputation}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
UpdateqMetadata

\S4method{UpdateqMetadata}{SummarizedExperiment}(object, from, to, ...)
}
\arguments{
\item{object}{An object of class `SummarizedExperiment`}

\item{na.type}{xxx}
}
\value{
NA

NA
}
\description{
Update the quantitative metadata information of missing values that were imputed
}
\examples{

obj <- Exp1_R25_pept[seq_len(10),]
obj[[2]] <- UpdateqMetadata(obj[[2]], 'missing', 'imputed')

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
