% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_violin.R
\name{violinPlotD}
\alias{violinPlotD}
\title{Builds a violinplot from a numeric matrix}
\usage{
violinPlotD(qData, fData, legend = NULL, palette = NULL, subset.view = NULL)
}
\arguments{
\item{qData}{Numeric matrix}

\item{fData}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}

\item{subset.view}{A vector of index indicating rows to highlight}
}
\value{
A violinplot
}
\description{
ViolinPlot for quantitative proteomics data
}
\examples{
library(vioplot)
library(DAPAR2)
library(Features)
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[['original_log']])
fData <- rowData(Exp1_R25_pept[['original_log']])
legend <- colData(Exp1_R25_pept)[["Condition"]]
violinPlotD(qData, fData, legend=legend)
}
\seealso{
\code{\link{densityPlotD}}
}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy
}
