% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{wrapper.normalizeD}
\alias{wrapper.normalizeD}
\title{Normalisation}
\usage{
wrapper.normalizeD(
  obj,
  method,
  conds = NULL,
  type = NULL,
  scaling = FALSE,
  quantile = 0.15,
  span = 0.7,
  subset.norm = NULL
)
}
\arguments{
\item{obj}{A \code{SummerizedExperiment} object}

\item{method}{"GlobalQuantileAlignment" (for normalizations of important magnitude),
"SumByColumns", "QuantileCentering", "Mean Centering", "LOESS" and "vsn".}

\item{conds}{xxx}

\item{type}{"overall" (shift all the sample distributions at once) or "within conditions" (shift the sample
distributions within each condition at a time).}

\item{scaling}{A boolean that indicates if the variance of the data have to
be forced to unit (variance reduction) or not.}

\item{quantile}{A float that corresponds to the quantile used to align the
data.}

\item{span}{Parameter for LOESS method}

\item{subset.norm}{A vector of index indicating rows to be used for normalization}
}
\value{
An instance of class \code{SummerizedExperiment} where the quantitative
data in the \code{array()} tab has been normalized.
}
\description{
Provides several methods to normalize quantitative data from a numeric matrix.
They are organized in six main families : GlobalQuantileAlignement, sumByColumns, QuantileCentering, MeanCentering, LOESS, vsn
For the first family, there is no type.
For the five other families, two type categories are available : 
"Overall" which means that the value for each protein (ie row) is computed over all the samples ;
"within conditions" which means that the value for each protein (ie row) is computed condition by condition.
Normalization can be based only on a subset of protein.
Proteins with NA in this subset are ignored.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[['original']]
conds <- colData(Exp1_R25_pept)[["Condition"]]
r<-wrapper.normalizeD(obj, "GlobalQuantileAlignment")
r<-wrapper.normalizeD(obj, "SumByColumns", conds, type="within conditions")
r<-wrapper.normalizeD(obj, "QuantileCentering", conds, type="within conditions")
r<-wrapper.normalizeD(obj, "MeanCentering", conds, type="overall")
r<-wrapper.normalizeD(obj, "vsn", conds, type="within conditions")
r<-wrapper.normalizeD(obj, "LOESS", conds, type="within conditions")
}
\author{
Samuel Wieczorek, Thomas Burger, Helene Borges, Anais Courtier, Enora Fremy
}
