% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_pca.R
\docType{data}
\name{wrapper.pca}
\alias{wrapper.pca}
\title{Compute the PCA}
\format{
An object of class \code{function} of length 1.
}
\usage{
wrapper.pca
}
\arguments{
\item{qData}{numeric matrix}

\item{condition}{xxx}

\item{var.scaling}{The dimensions to plot}

\item{ncp}{Number of dimensions kept in the results}
}
\value{
A list including eigenvalues of obj
}
\description{
Compute the PCA
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[[2]])
condition <- colData(Exp1_R25_pept)[["Condition"]]
res.pca <- wrapper.pca(qData, condition)

plotPCA_Var(res.pca)

plotPCA_Ind(res.pca)

plotPCA_Eigen_hc(res.pca)


}
\author{
Samuel Wieczorek, Enora Fremy
}
\keyword{datasets}
