% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_features.R
\name{writeFeaturesToExcel}
\alias{writeFeaturesToExcel}
\title{This function exports a \code{MSnSet} object to a Excel file.}
\usage{
writeFeaturesToExcel(obj, filename)
}
\arguments{
\item{obj}{An object of class \code{Features}.}

\item{filename}{A character string for the name of the Excel file.}
}
\value{
A Excel file (.xlsx)
}
\description{
This function exports a \code{Features} data object to a Excel file.
Each of the three data.frames in the \code{MSnSet} object (ie experimental data,
phenoData and metaData are respectively integrated into separate sheets in
the Excel file). The colored cells in the experimental data correspond to the 
original missing valueswhich have been imputed.
}
\examples{
\donttest{
Sys.setenv("R_ZIPCMD"= Sys.which("zip"))
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000]
writeFeaturesToExcel(obj, "foo")
}

}
\author{
Samuel Wieczorek
}
