% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_qfeatures.R
\name{writeQFeaturesToExcel}
\alias{writeQFeaturesToExcel}
\title{This function exports a \code{QFeatures} object to a Excel file.}
\usage{
writeQFeaturesToExcel(obj, filename)
}
\arguments{
\item{obj}{An object of class \code{QFeatures}.}

\item{filename}{A character string for the name of the Excel file.}
}
\value{
A Excel file (.xlsx)
}
\description{
This function exports a \code{QFeatures} data object to a Excel file.
Each of the three data.frames in the \code{QFeatures} object (ie expression data,
colData and rowData are respectively integrated into separate sheets in
the Excel file). The colored cells in the experimental data correspond to the 
original missing values which have been imputed.
}
\examples{
\donttest{
Sys.setenv("R_ZIPCMD"= Sys.which("zip"))
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[seq_len(1000)]
writeQFeaturesToExcel(obj, "foo")
}

}
\author{
Samuel Wieczorek
}
