% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggIter}
\alias{aggIter}
\title{Method to aggregate peptides into proteins with the iterative approach.}
\usage{
aggIter(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)
}
\arguments{
\item{qPepData}{A matrix of intensities of peptides.}

\item{X}{An adjacency matrix}

\item{conditions}{xxx}

\item{init.method}{The method used to initialize the iterative algotirhm. Default is 'Sum'.}

\item{method}{The method used for xxx. Default value is 'Mean'.}

\item{n}{xxxx}
}
\value{
A matrix of protein intensities.
}
\description{
Method to aggregate peptides into proteins with the iterative approach
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000]
obj <- SetAdjMat(obj, 2)
X <- GetAdjMat(obj, 2)$all
conds <- colData(obj)$Condition
aggIter(assay(obj,2), X, conds)

}
\author{
Samuel Wieczorek
}
