% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{aggIterParallel}
\alias{aggIterParallel}
\title{Method to aggregate peptides into proteins with the iterative approach with use of parallelism.}
\usage{
aggIterParallel(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)
}
\arguments{
\item{qPepData}{A matrix of intensities of peptides.}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{conditions}{xxxx}

\item{init.method}{The method used to initialize the iterative algotirhm. Default is 'Sum'.}

\item{method}{The method used for xxx. Default value is 'Mean'.}

\item{n}{xxxx}
}
\value{
A matrix of intensities of proteins
}
\description{
Method to aggregate peptides into proteins with the iterative approach with use of parallelism.
}
\author{
Samuel Wieczorek
}
