% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggIterParallel}
\alias{aggIterParallel}
\title{Method to aggregate peptides into proteins with the iterative approach with use of parallelism.}
\usage{
aggIterParallel(
  qPepData,
  X,
  conditions = NULL,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)
}
\arguments{
\item{qPepData}{A matrix of intensities of peptides.}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{conditions}{xxxx}

\item{init.method}{The method used to initialize the iterative algotirhm. Default is 'Sum'.}

\item{method}{The method used for xxx. Default value is 'Mean'.}

\item{n}{xxxx}
}
\value{
A matrix of intensities of proteins
}
\description{
Method to aggregate peptides into proteins with the iterative approach with use of parallelism.
}
\examples{
library(doParallel)
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]])$all)
conditions <- SummarizedExperiment::colData(obj)$Condition
aggIterParallel(assay(obj,2), X, conditions)

}
\author{
Samuel Wieczorek
}
