% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggMetadata_parallel_sam}
\alias{aggMetadata_parallel_sam}
\title{This function aggregates features of metadata of peptides into a dataframe that is to be included
in the rowdata of the new protein dataset.}
\usage{
aggMetadata_parallel_sam(pepMetadata, names, X, simplify = TRUE)
}
\arguments{
\item{pepMetadata}{A data.frame of meta data of peptides. It is the rowData 
of the QFeatures object.}

\item{names}{A vector of column names of the rowdata of peptides.}

\item{X}{The adjacency matrix used to agregate the peptides data.}

\item{simplify}{This parameter is either a boolean or a vector of boolean. The TRUE value means that, for a given protein,
duplicates metadata are reduced into one. The FALSE value means that, for a given protein, duplicates metadata are not
reduced: one keep the real original values.
If the parameter is a vector, its size must be equal to the size of the parameter names. Each item in 'simplify' is the indication
for the item of the same indice in the vector 'names' to apply the reduction or not.}
}
\value{
A data frame
}
\description{
This function aggregates features of metadata of peptides into a dataframe that is to be included
in the rowdata of the new protein dataset. It is the same function as aggMetadata_sam but with parallelism.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]])$all)
ft <- aggMetadata_parallel_sam(rowData(obj[[2]]), c('Sequence', 'Proteins'), X)

}
\author{
Samuel Wieczorek
}
