% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggTopn}
\alias{aggTopn}
\title{Compute the intensity of proteins as the sum of the intensities of their n best peptides.}
\usage{
aggTopn(qPepData, X, method = "Mean", n = 10)
}
\arguments{
\item{qPepData}{A matrix of intensities of peptides}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}

\item{method}{Default is 'Mean'}

\item{n}{The maximum number of peptides used to aggregate a protein.}
}
\value{
A matrix of intensities of proteins
}
\description{
This function computes the intensity of proteins as the sum of the 
intensities of their n best peptides.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]])$all)
aggTopn(assay(obj,2), X, n=3)

}
\author{
Alexia Dorffer, Samuel Wieczorek
}
