% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_boxplot.R
\name{boxPlotD_HC}
\alias{boxPlotD_HC}
\title{Builds a boxplot from a dataframe using the library \code{highcharter}}
\usage{
boxPlotD_HC(qData, fData, legend = NULL, palette = NULL, subset.view = NULL)
}
\arguments{
\item{qData}{Numeric matrix}

\item{fData}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}

\item{subset.view}{A vector of index indicating rows to highlight}
}
\value{
A boxplot
}
\description{
Boxplot for quantitative proteomics data using the library \code{highcharter}
}
\examples{
library(Features)
library(DAPAR2)
library(highcharter)
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[['original_log']])
fData <- rowData(Exp1_R25_pept[['original_log']])
legend <- colData(Exp1_R25_pept)[["Sample.name"]]
boxPlotD_HC(qData, fData, legend, subset.view=1:10)
}
\seealso{
\code{\link{densityPlotD_HC}}
}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy
}
