% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_boxplot.R
\name{boxPlotD_HC}
\alias{boxPlotD_HC}
\title{Builds a boxplot from a dataframe using the library \code{highcharter}}
\usage{
boxPlotD_HC(
  qData,
  conds,
  keyId = NULL,
  legend = NULL,
  palette = NULL,
  subset.view = NULL
)
}
\arguments{
\item{qData}{Numeric matrix}

\item{conds}{xxx}

\item{keyId}{xxxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}

\item{subset.view}{A vector of index indicating rows to highlight}
}
\value{
A boxplot
}
\description{
Boxplot for quantitative proteomics data using the library \code{highcharter}
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[[2]])
conds <- colData(Exp1_R25_pept)[["Condition"]]
key <- rowData(Exp1_R25_pept[[2]])[[metadata(Exp1_R25_pept)$keyId]]
boxPlotD_HC(qData, conds, keyId = key, conds, subset.view=1:10)

}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy
}
