% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectedComps.R
\name{buildGraph}
\alias{buildGraph}
\title{Build graph definition}
\usage{
buildGraph(The.CC, X)
}
\arguments{
\item{The.CC}{A cc (a list)}

\item{X}{An adjacency matrix as given by the function BuildListAdjacencyMatrices}
}
\value{
A list of two itens: nodes and edges
}
\description{
Builds a graph as defined by its nodes and edges.
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData")) 
X <- GetAdjMat(Exp1_R25_pept[[2]])$all
ll <- get.pep.prot.cc(X)
g <- buildGraph(ll[[1]], X)

}
\author{
Thomas Burger, Samuel Wieczorek
}
