% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{buildGraph}
\alias{buildGraph}
\title{Build graph definition}
\usage{
buildGraph(The.CC, X)
}
\arguments{
\item{The.CC}{A cc (a list)}

\item{X}{An adjacency matrix as given by the function BuildListAdjacencyMatrices}
}
\value{
A list of two itens: nodes and edges
}
\description{
Builds a graph as defined by its nodes and edges.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2') 
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- GetAdjMat(obj[[2]], 'all')
ll <- get.pep.prot.cc(X)
g <- buildGraph(ll[[1]], X)

}
\author{
Thomas Burger, Samuel Wieczorek
}
