% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_compare_normalization_plots.R
\name{compareNormalizationD_HC}
\alias{compareNormalizationD_HC}
\title{Builds a plot from a numeric matrix.}
\usage{
compareNormalizationD_HC(
  qDataBefore,
  qDataAfter,
  condsForLegend = NULL,
  indData2Show = NULL,
  palette = NULL
)
}
\arguments{
\item{qDataBefore}{A numeric matrix that contains quantitative data before 
normalization.}

\item{qDataAfter}{A numeric matrix that contains quantitative data after 
normalization.}

\item{condsForLegend}{A vector of the conditions (one condition 
per sample).}

\item{indData2Show}{A vector of the indices of the columns to show in 
the plot. The indices are those of indices of 
the columns int the data.frame qDataBefore.}

\item{palette}{xxx}
}
\value{
A plot
}
\description{
Plot to compare the quantitative proteomics data before and after 
normalization using the library \code{highcharter}
}
\examples{
library(highcharter)
library(DAPAR)
# source('~/TELETRAVAIL/github_DAPARforFeatures/DAPAR2/R/normalize.R')
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
conds <- colData(obj)[["Condition"]]
objBefore <- obj[['original_log']]
qDataBefore <- assay(objBefore)
objAfter <- wrapper.normalizeD(objBefore,"QuantileCentering", conds, "within conditions")
qDataAfter <- assay(objAfter)
compareNormalizationD_HC(qDataBefore, qDataAfter, condsForLegend=conds)
}
\author{
Samuel Wieczorek, Enora Fremy
}
