% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_compare_normalization_plots.R
\name{compareNormalizationD_HC}
\alias{compareNormalizationD_HC}
\title{Builds a plot from a numeric matrix.}
\usage{
compareNormalizationD_HC(
  qDataBefore,
  qDataAfter,
  conds = NULL,
  palette = NULL,
  subset.view = NULL,
  n = NULL,
  type = "scatter"
)
}
\arguments{
\item{qDataBefore}{A numeric matrix that contains quantitative data before normalization.}

\item{qDataAfter}{A numeric matrix that contains quantitative data after normalization.}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}

\item{subset.view}{xxx}

\item{n}{xxx}

\item{type}{scatter or line}
}
\value{
A plot
}
\description{
Builds a plot from a numeric matrix.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
conds <- colData(obj)[["Condition"]]
obj <- normalizeD(obj, 2, name='norm', method='SumByColumns', conds=conds, type='overall')
compareNormalizationD_HC(assay(obj, 2), assay(obj, 3), conds=conds, n=100)

pal <- ExtendPalette(2, "Dark2")
compareNormalizationD_HC(assay(obj, 2), assay(obj, 3), conds=conds, n=100, palette=pal)

}
\author{
Samuel Wieczorek, Enora Fremy
}
