% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_compare_normalization_plots.R
\name{compareNormalizationD_HC}
\alias{compareNormalizationD_HC}
\title{Builds a plot from a numeric matrix.}
\usage{
compareNormalizationD_HC(
  qDataBefore,
  qDataAfter,
  conds,
  keyId = NULL,
  palette = NULL,
  subset.view = NULL,
  n = 100,
  type = "scatter"
)
}
\arguments{
\item{qDataBefore}{A numeric matrix containing quantitative data before normalization.}

\item{qDataAfter}{A numeric matrix containing quantitative data after normalization.}

\item{conds}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}

\item{subset.view}{xxx}

\item{n}{xxx}

\item{type}{scatter or line}
}
\value{
A plot
}
\description{
This plot compares the quantitative proteomics data before and after 
normalization
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
conds <- colData(obj)$Condition
id <- metadata(obj)$keyId
obj <- normalizeD(obj, 2, name='norm', method='SumByColumns', conds=conds, type='overall')
palette <- ExtendPalette(2, "Dark2")
qBefore <- assay(obj, 2)
qAfter <- assay(obj, 3)
compareNormalizationD_HC(qBefore, qAfter, conds=conds, n=100)
compareNormalizationD_HC(qBefore, qAfter, conds=conds, keyId = id, n=100, palette=palette)

}
\author{
Samuel Wieczorek, Enora Fremy
}
