% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\docType{data}
\name{ComplexFilterFeatures}
\alias{ComplexFilterFeatures}
\title{Complex filter features}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
ComplexFilterFeatures
}
\description{
Filter an assay with a complex query
}
\author{
Samuel Wieczorek
}
\keyword{datasets}
