% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{ComputeConnexComposants}
\alias{ComputeConnexComposants}
\title{Computes the connex components of a graph, represented by an adjacency matrix.}
\usage{
ComputeConnexComposants(X.list = NULL)
}
\arguments{
\item{X.list}{A list of adjacency matrices as provided by the function BuildListAdjacencyMatrices.}
}
\value{
A list of items containing the connex components derived from the matrices given in arguments
}
\description{
This function computes the connex components of a graph, represented by an adjacency matrix.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original_log']][1:1000,])[,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original_log']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All', 'Shared', 'Spec'))
cc <- ComputeConnexComposants(X.list)

}
\author{
Samuel Wieczorek
}
