% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_corr_matrix.R, R/mod_plots_corr_matrix.R
\name{corrMatrixPlot}
\alias{corrMatrixPlot}
\alias{corrmatrix_plot}
\alias{mod_corrmatrix_plot_ui}
\alias{mod_corrmatrix_plot_server}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
corrMatrixPlot(qData, rate = 0.5, showValues = FALSE)

mod_corrmatrix_plot_ui(id)

mod_corrmatrix_plot_server(
  id,
  qData,
  rate = reactive({     0.5 }),
  showValues = reactive({     FALSE })
)
}
\arguments{
\item{qData}{xxx}

\item{rate}{xxx. Default value is 0.9}

\item{showValues}{Default is FALSE.}

\item{id}{xxx}
}
\value{
A colored correlation matrix
}
\description{
A shiny Module.
}
\examples{
library(QFeatures)
data(ft)
corrMatrixPlot(assay(ft, 1))

if (interactive()){
library(QFeatures)
library(DaparToolshed)
data(ft)
ui <- mod_corrmatrix_plot_ui('plot')

server <- function(input, output, session) {
 mod_corrmatrix_plot_server('plot',
                            qData = reactive({assay(ft[[1]])})
                            )
 }
shinyApp(ui=ui, server=server)
}
}
\author{
Samuel Wieczorek, Enora Fremy
}
