% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_corr_matrix.R
\name{corrMatrixD_HC}
\alias{corrMatrixD_HC}
\title{Displays a correlation matrix of the quantitative data of a
numeric matrix.}
\usage{
corrMatrixD_HC(obj.se, names = NULL, rate = 0.5, showValues = TRUE)
}
\arguments{
\item{obj.se}{An object of class 'SummarizedExperiment'}

\item{names}{A vector of strings which will be used as legend. The length
of 'names' must be equal to the number of samples in the dataset.}

\item{rate}{The rate parameter to control the exponential law for 
the gradient of colors}

\item{showValues}{xxx}
}
\value{
A colored correlation matrix
}
\description{
Displays a correlation matrix of the quantitative data of a
numeric matrix.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
corrMatrixD_HC(Exp1_R25_pept[[2]])

}
\author{
Samuel Wieczorek, Enora Fremy
}
