% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{create_ft_example}
\alias{create_ft_example}
\title{Creates an object of class [QFeatures] as example}
\usage{
create_ft_example()
}
\value{
An example object of class [QFeatures]
}
\description{
xxxxx
}
\author{
Samuel Wieczorek
}
