% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-data.R
\name{create_ft_example}
\alias{create_ft_example}
\title{Creates an object of class \link{QFeatures} as example}
\usage{
create_ft_example(with.na = FALSE)
}
\value{
An example object of class \link{QFeatures}
}
\description{
xxxxx
}
\examples{
create_ft_example()
create_ft_example(TRUE)

}
\author{
Samuel Wieczorek
}
