% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_density_plot.R
\name{densityPlotD_HC}
\alias{densityPlotD_HC}
\title{Builds a densityplot from a dataframe}
\usage{
densityPlotD_HC(qData, conds, legend = NULL, palette = NULL)
}
\arguments{
\item{qData}{numeric matrix}

\item{conds}{DataFrame}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{A vector of HEX Code colors (one color per condition).}
}
\value{
A density plot
}
\description{
Densityplot of quantitative proteomics data over samples.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
qData <- assay(Exp1_R25_pept[[2]])
conds <- colData(Exp1_R25_pept)[["Condition"]]
densityPlotD_HC(qData, conds

legend <- colData(Exp1_R25_pept)[["Sample.name"]]
pal <- ExtendPalette(2, 'Dark2')
densityPlotD_HC(qData, conds, legend, palette=pal)

}
\author{
Samuel Wieczorek, Enora Fremy
}
