% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_all_plots.R, R/mod_ds_corrmatrix.R,
%   R/mod_ds_density.R, R/mod_ds_explorer.R, R/mod_ds_heatmap.R,
%   R/mod_ds_intensity.R, R/mod_ds_mv.R, R/mod_plot_mvImputation.R,
%   R/mod_ds_pca.R, R/mod_ds_variance.R, R/plot_ds_boxplot.R,
%   R/plot_ds_corrmatrix.R, R/plot_ds_variance.R, R/plot_ds_density.R,
%   R/plot_ds_heatmap.R, R/plot_ds_pca.R, R/plot_ds_violin.R, R/plot_ds_mv.R
\name{descriptive-statistics}
\alias{descriptive-statistics}
\alias{listPlotModules}
\alias{mod_all_plots_ui}
\alias{mod_all_plots_server}
\alias{mod_ds_corrmatrix_ui}
\alias{mod_ds_corrmatrix_server}
\alias{mod_ds_density_ui}
\alias{mod_ds_density_server}
\alias{mod_ds_explorer_ui}
\alias{mod_ds_explorer_server}
\alias{mod_ds_heatmap_ui}
\alias{mod_ds_heatmap_server}
\alias{mod_ds_intensity_ui}
\alias{mod_ds_intensity_server}
\alias{mod_ds_mv_ui}
\alias{mod_ds_mv_server}
\alias{mod_plot_mvImputation_ui}
\alias{mod_plot_mvImputation_server}
\alias{mod_ds_pca_ui}
\alias{mod_ds_pca_server}
\alias{mod_ds_variance_ui}
\alias{mod_ds_variance_server}
\alias{boxPlotD}
\alias{corrMatrixPlot}
\alias{CVDist}
\alias{densityPlot}
\alias{heatmapD}
\alias{mv.heatmap}
\alias{wrapper_pca}
\alias{violinPlotD}
\alias{mvPerLinesHisto}
\alias{mvPerLinesHistoPerCondition}
\alias{mvHisto}
\title{Bar plot of missing values per lines using highcharter.}
\usage{
listPlotModules()

mod_all_plots_ui(id)

mod_all_plots_server(id, object)

mod_ds_corrmatrix_ui(id)

mod_ds_corrmatrix_server(
  id,
  object,
  rate = reactive({     0.5 }),
  showValues = reactive({     FALSE })
)

mod_ds_density_ui(id)

mod_ds_density_server(id, object, conds, pal.name = reactive({
    NULL
}))

mod_ds_explorer_ui(id)

mod_ds_explorer_server(id, object, which.assay, digits = reactive({
    3
}))

mod_ds_heatmap_ui(id)

mod_ds_heatmap_server(id, object, conds, width = 900)

mod_ds_intensity_ui(id)

mod_ds_intensity_server(id, object, exp.design, ...)

mod_ds_mv_ui(id)

mod_ds_mv_server(id, object, conds, pal.name = reactive({
    NULL
}))

mod_plot_mvImputation_ui(id)

mod_plot_mvImputation_server(
  id,
  object = reactive({     NULL }),
  conds = reactive({     NULL }),
  title = reactive({     NULL }),
  pal.name = reactive({     NULL })
)

mod_ds_pca_ui(id)

mod_ds_pca_server(id, object, conds)

mod_ds_variance_ui(id)

mod_ds_variance_server(id, object, conds, pal.name = NULL)

boxPlotD(object, exp.design, subset = NULL, pal.name = NULL)

corrMatrixPlot(object, rate = 0.5, showValues = FALSE)

CVDist(object, conds = NULL, pal.name = NULL)

densityPlot(qData, conds, pal.name = NULL)

heatmapD(qData, conds, distfun = "euclidean", hclustfun = "complete")

mv.heatmap(
  x,
  col = grDevices::heat.colors(100),
  srtCol = NULL,
  labCol = NULL,
  labRow = NULL,
  key = TRUE,
  key.title = NULL,
  main = NULL,
  ylab = NULL
)

wrapper_pca(object, conds, var.scaling = TRUE, ncp = NULL)

violinPlotD(object, exp.design, subset = NULL, pal.name)

mvPerLinesHisto(qData)

mvPerLinesHistoPerCondition(qData, conds, pal.name)

mvHisto(qData, conds, showValues = FALSE, pal.name = NULL)
}
\arguments{
\item{id}{shiny id}

\item{object}{An instance of the class \code{SummarizedExperiment}}

\item{rate}{The rate parameter to control the exponential law for
the gradient of colors}

\item{showValues}{xxx}

\item{conds}{xxx}

\item{pal.name}{xxx}

\item{which.assay}{A}

\item{digits}{xxx}

\item{width}{xxx}

\item{exp.design}{xxx}

\item{...}{xxx}

\item{title}{A \code{character(1)} which is}

\item{subset}{A vector of index indicating rows to highlight}

\item{qData}{A \code{matrix} or \code{array} quantitative values}

\item{distfun}{A \code{character(1)} defining the distance used by the clustering
algorithm to compute. Default value is 'euclidean'. See \code{help(heatmap.2)}}

\item{hclustfun}{the clustering algorithm used to build the dendrogram.
Default value is 'complete'. See \code{help(heatmap.2)}}

\item{x}{A \code{matrix} or \code{array} containing the quantitative data.}

\item{col}{Colors used for the image. Defaults to heat colors (heat.colors).}

\item{srtCol}{Angle of column conds, in degrees from horizontal}

\item{labCol}{Character vectors with column conds to use.}

\item{labRow}{Character vectors with row conds to use.}

\item{key}{Logical indicating whether a color-key should be shown.}

\item{key.title}{Main title of the color key. If set to NA no title will
be plotted.}

\item{main}{Main title; default to none.}

\item{ylab}{y-axis title; default to none.}

\item{var.scaling}{The dimensions to plot}

\item{ncp}{A \code{integer(1)} which represents the umber of dimensions kept in the results.}

\item{obj}{xxx}

\item{base_palette}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}
}
\value{
A plot

NA

NA

NA

A boxplot

A colored correlation matrix

A violinplot
}
\description{
This method plots a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins).
}
\details{
\itemize{
\item distribution of the missing values per line,
\item a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins) and per conditions,
\item Histogram of missing values.
\item Variance : Builds a densityplot of the CV of entities in numeric matrix.
The CV is calculated for each condition present in the dataset
(see the slot \code{'Condition'} in the \code{colData()} DataFrame)
\item Heatmap:
}

The function \code{\link[=heatmapD]{heatmapD()}}

The function [] is inspired from the function 'heatmap.2'
that displays a numeric matrix. For more information, please refer to the help
of the heatmap.2 function.
}
\section{Missing values}{


#' - distribution of the missing values per line,
\itemize{
\item a bar plot which represents the distribution of the
number of missing values (NA) per lines (ie proteins) and per conditions,
\item Histogram of missing values.
}
}

\examples{
library(QFeatures)
library(DaparToolshed)
data(ft)
data(ft_na)

qData <- assay(ft, 1)
conds <- design(ft)$Condition
pal <- 'Dark2'

data(ft)
qData <- assay(ft, 1)
conds <- design(ft)$Condition
densityPlot(qData, conds)

legend <- design(ft)$Sample.name
densityPlot(qData, conds, pal.name = 'Dark2')

CVDist(qData, conds)

#----------------------------------------------
# Plots a heatmap for generic quantitative data
#----------------------------------------------

heatmapD(assay(ft, 1), conds)

#----------------------------------------------
# Plots a heatmap for missing values visualization
#----------------------------------------------


#' #----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvPerLinesHisto(qData)

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------

mvHisto(qData, conds, pal.name = pal)

#----------------------------------------
# Plots a histogram of missing values
#----------------------------------------



library(QFeatures)
data(ft)
mv.heatmap(assay(ft, 1))

#----------------------------------------
# Launch a single shiny module
# The list of all plots module is avaiable
# with the function [listPlotModules()].
# In the example, replace 'FOO' by the name of the module
# and add the necessaray parameters
#----------------------------------------

if(interactive()){
 data(ft)
 ui <- FOO_ui('plot')

 server <- function(input, output, session) {
  conds <- design(ft)$Condition
 
  FOO_server('plot',
             # Add appropriate parameters
             )
  }
 
 shinyApp(ui=ui, server=server)
}


#----------------------------------------
# Launch a the shiny module for all plots
# Here, an example with the density plot
#----------------------------------------

if(interactive()){
 data(ft)
 ui <- mod_all_plots_ui('plot')

 server <- function(input, output, session) {
  mod_all_plots_server('plot',
                       object = reactive({ft})
                       )
  }
 
 shinyApp(ui=ui, server=server)
}
library(QFeatures)
data(ft)
corrMatrixPlot(assay(ft, 1))

}
\author{
Samuel Wieczorek, Enora Fremy

Samuel Wieczorek, Anais Courtier, Enora Fremy

Samuel Wieczorek, Enora Fremy
}
