% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_CV_dist.R,
%   R/mod_plots_legend_colored_exprs.R, R/mod_plots_se_explorer.R,
%   R/mod_plots_var_dist.R
\name{CVDistD_HC}
\alias{CVDistD_HC}
\alias{mod_plots_legend_colored_exprs_ui}
\alias{mod_plots_legend_colored_exprs_server}
\alias{mod_plots_se_explorer_ui}
\alias{mod_plots_se_explorer_server}
\alias{mod_plots_var_dist_ui}
\alias{mod_plots_var_dist_server}
\title{Distribution of CV of entities}
\usage{
CVDistD_HC(qData, conds = NULL, palette = NULL)

mod_plots_legend_colored_exprs_ui(id)

mod_plots_legend_colored_exprs_server(id)

mod_plots_se_explorer_ui(id)

mod_plots_se_explorer_server(id, obj, originOfValues = NULL, colData = NULL)

mod_plots_var_dist_ui(id)

mod_plots_var_dist_server(id, obj, conds, base_palette = NULL)
}
\arguments{
\item{qData}{A numeric matrix of quantitative data.}

\item{conds}{A \code{character()} representings the condition for
each sample of the \link{QFeatures} object.}

\item{palette}{A vector of HEX Code colors (one color per condition).}

\item{id}{shiny id}

\item{obj}{xxx}

\item{base_palette}{xxx}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\value{
A density plot

NA

NA

NA

NA
}
\description{
Builds a densityplot of the CV of entities in numeric matrix.
The CV is calculated for each condition present in the dataset
(see the slot \code{'Condition'} in the \code{colData()} DataFrame)

A shiny Module.

A shiny Module.

A shiny Module.
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
qData <- assay(Exp1_R25_pept[[2]])[seq_len(10),]
conds <- colData(Exp1_R25_pept)[["Condition"]]
CVDistD_HC(qData, conds)


}
\author{
Samuel Wieczorek, Enora Fremy

Samuel Wieczorek
}
\keyword{internal}
