% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\docType{data}
\name{diff_analysis_sam}
\alias{diff_analysis_sam}
\title{Check the validity of the experimental design}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
diff_analysis_sam
}
\description{
This manual page describes the computation of statistical test using \link{QFeatures} objects. In the following
functions, if \code{object} is of class \code{QFeatures}, and optional assay
index or name \code{i} can be specified to define the assay (by name of
index) on which to operate.

The following functions are currently available:
\itemize{
\item \code{compute.t.test(xxxxx)} xxxxx.
\item \code{compute.group.t.test(xxxxx)} xxxxx.
\item \code{limma.complete.test(object, sampleTab)} uses the package Limma
}
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
object <- Exp1_R25_pept[seq_len(1000),]
object <- addAssay(object, QFeatures::filterNA(object[[2]],  pNA = 0), name='filtered')
sTab <- colData(object)
gttest.se <- t_test_sam(object[[3]], sTab, FUN = compute.t.test)
object <- addAssay(object, gttest.se, name='t-test')
comp <- '25fmol_vs_10fmol'
da.se <- diff_analysis_sam(object[['t-test']], comp)

object <- diff_analysis_sam(object, 't-test', name='diffAna', comp)

}
\keyword{datasets}
