% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectedComps.R
\name{display.CC.visNet}
\alias{display.CC.visNet}
\title{Display a connected component.}
\usage{
display.CC.visNet(g, obj = NULL, prot.tooltip = NULL, pept.tooltip = NULL)
}
\arguments{
\item{g}{A connex component (a list)}

\item{obj}{xxx}

\item{prot.tooltip}{A vector of strings which are the text for the tooltips of the proteins.}

\item{pept.tooltip}{A vector of strings which are the text for the tooltips of the peptides.}
}
\value{
A plot
}
\description{
Displays a connected component as a graph with the package 'visNetwork'.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2') 
X <- GetAdjMat(Exp1_R25_pept[[2]])$all
ll <- get.pep.prot.cc(X)
g <- buildGraph(ll[[1]], X)
display.CC.visNet(g)

}
\author{
Thomas Burger, Samuel Wieczorek
}
