% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\docType{data}
\name{ComplexFilterFeatures,QFeatures-method}
\alias{ComplexFilterFeatures,QFeatures-method}
\alias{filterFeaturesSam}
\alias{filterFeaturesSam,SummarizedExperiment-method}
\alias{filterFeaturesSam,QFeatures-method}
\title{Check the validity of the experimental design}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
\S4method{ComplexFilterFeatures}{QFeatures}(object, i, name = "filterAssay", ...)

filterFeaturesSam

\S4method{filterFeaturesSam}{SummarizedExperiment}(object, filter, ...)

\S4method{filterFeaturesSam}{QFeatures}(object, i, name = "filterAssay", filter, ...)
}
\arguments{
\item{object}{An object of class `QFeatures`.}

\item{i}{A numeric vector or a character vector giving the index or the 
name, respectively, of the assay(s) to be processed.}

\item{name}{A `character(1)` naming the new assay name. Defaults
are `ttestAssay`.}

\item{...}{Additional parameters passed to inner functions.}

\item{filter}{xxx}
}
\description{
This manual page describes the The filtering functions of DAPAR have not been all moved to DaparToolshed as we now use the QFeatures package
which provides some filtering functions, especially on features that are present in the rowData
of the datasets.
 The filtering functions on numerical values are deleted because the same functions exist in QFeatures.
 For the missing values filtering on conditions, we do not use DAPAR functions anymore. Instead, we use
 the numerical filtering functions in QFeatures. To do so, it is necessary to build some rowdata for the
 SummarizedExperiment (not necessary stored in the object) which count the number of missing values w.r.t.
 the type of filtering: whole matrix, at least one value per condition, etc...
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_prot, package='DAPARdata2')
object <- Exp1_R25_prot
filter_600 <- VariableFilter(field='Sequence_length', value=as.numeric('600'), condition='>')
object <- filterFeaturesSam(object, i=2, filter=reverse_600)




}
\keyword{datasets}
