% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{find_MEC_matrix}
\alias{find_MEC_matrix}
\title{Finds the LAPALA into a \code{MSnSet} object}
\usage{
find_MEC_matrix(x, conds)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment}.}
}
\value{
A data.frame that contains the indexes of LAPALA
}
\description{
Finds the LAPALA into a \code{MSnSet} object
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000]
lapala <- find_MEC_matrix(assay(obj[[2]]), colData(obj)$Condition)

}
\author{
Samuel Wieczorek
}
