% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectedComps.R
\name{get.pep.prot.cc}
\alias{get.pep.prot.cc}
\title{Build the list of connected components of an adjacency matrix.}
\usage{
get.pep.prot.cc(X)
}
\arguments{
\item{X}{An adjacency matrix}
}
\value{
A list of Connected Components (CC)
}
\description{
Method to build the list of connected components of an adjacency matrix
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2') 
X <- GetAdjMat(Exp1_R25_pept[[2]])$all
ll <- get.pep.prot.cc(X)

}
\author{
Thomas Burger, Samuel Wieczorek
}
