% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\docType{data}
\name{Get_qMetadata}
\alias{Get_qMetadata}
\title{xxxx}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
Get_qMetadata
}
\description{
xxxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:10,]
mc <- Get_qMetadata(obj[[2]])

}
\keyword{datasets}
