% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{GetAdjMat}
\alias{GetAdjMat}
\alias{GetAdjMat,SummarizedExperiment-method}
\alias{GetAdjMat,QFeatures-method}
\title{Returns the contains of the slot processing  of an object of 
class \code{MSnSet}}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
GetAdjMat

\S4method{GetAdjMat}{SummarizedExperiment}(object, ...)

\S4method{GetAdjMat}{QFeatures}(object, i, ...)
}
\arguments{
\item{object}{xxx}

\item{...}{xxx}

\item{i}{xxx}
}
\description{
xxx
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
ll.X <- GetAdjMat(Exp1_R25_pept, 2)

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
