% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{GetAdjMat}
\alias{GetAdjMat}
\title{Returns the contains of the slot processing  of an object of 
class \code{MSnSet}}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
GetAdjMat
}
\description{
Returns the contains of the slot processing  of an object of 
class \code{MSnSet}
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
ll.X <- GetAdjMat(Exp1_R25_pept, 2)

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
