% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_func.R
\name{GetDetailedNbPeptides}
\alias{GetDetailedNbPeptides}
\title{Computes the detailed number of peptides for each protein}
\usage{
GetDetailedNbPeptides(X)
}
\arguments{
\item{X}{An adjacency matrices}
}
\value{
A data.frame containing the following vectors:
* nTotal: The number of peptides for each protein,
* nShared: the number of shared peptides for each protein,
* nSPec: the number of specific peptides for each protein.

Each row of these three vectors represent a protein. Thus, the length of these vectors is equal 
to the number of proteins.
}
\description{
Method to compute the detailed number of peptides for each protein
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000),]
obj <- addListAdjacencyMatrices(obj, 2)
X <- GetAdjMat(obj[[2]])$all
n <- GetDetailedNbPeptides(X)

}
\author{
Samuel Wieczorek
}
