% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{GetDetailedNbPeptides}
\alias{GetDetailedNbPeptides}
\title{Computes the detailed number of peptides for each protein}
\usage{
GetDetailedNbPeptides(X)
}
\arguments{
\item{X}{An adjacency matrices}
}
\value{
A data.frame containing the following vectors:
* nTotal: The number of peptides for each protein,
* nShared: the number of shared peptides for each protein,
* nSPec: the number of specific peptides for each protein.

Each row of these three vectors represent a protein. Thus, the length of these vectors is equal 
to the number of proteins.
}
\description{
Method to compute the detailed number of peptides for each protein
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
X <- GetAdjMat(Exp1_R25_pept, 2)$all
n <- GetDetailedNbPeptides(X)

}
\author{
Samuel Wieczorek
}
