% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{GetDetailedNbPeptidesUsed}
\alias{GetDetailedNbPeptidesUsed}
\title{Computes the detailed number of peptides used for aggregating each protein w.r.t NA values in 
peptide quantitative dataset. Even if a peptide is part of a protein, if its value is NA, it do not be used
for aggreation.}
\usage{
GetDetailedNbPeptidesUsed(X, qPepData)
}
\arguments{
\item{X}{An adjacency matrix.}

\item{qPepData}{A data.frame of quantitative data}
}
\value{
A list of three items:
* nAll: the number of pall eptides used for aggregation,
* nShared: the number of shared peptides used for aggregation,
* nSpec: the total number of pspecific eptides used for aggregation.
}
\description{
Computes the detailed number of peptides used for aggregating each protein w.r.t NA values in 
peptide quantitative dataset. Even if a peptide is part of a protein, if its value is NA, it do not be used
for aggreation.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
qPepData <- assay(Exp1_R25_pept,2)
X <- GetAdjMat(Exp1_R25_pept, 2)$all
n <- GetDetailedNbPeptidesUsed(X, qPepData)
 
}
\author{
Samuel Wieczorek
}
