% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{GetNbPeptidesUsed}
\alias{GetNbPeptidesUsed}
\title{Computes the number of peptides used for aggregating each protein}
\usage{
GetNbPeptidesUsed(qPepData, X)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
A data.frame
}
\description{
Method to compute the number of quantified peptides used for aggregating each protein
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000),]
qPepData <- assay(obj,2)
obj <- addListAdjacencyMatrices(obj, 2)
X <- as.matrix(GetAdjMat(obj[[2]])$all)
n <- GetNbPeptidesUsed(qPepData, X)

}
\author{
Samuel Wieczorek
}
