% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\docType{data}
\name{GetTypeDataset}
\alias{GetTypeDataset}
\title{Get the type of dataset}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
GetTypeDataset
}
\arguments{
\item{object}{An object of class 'SummarizedExperiment'}
}
\description{
Get the type of dataset
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
GetTypeDataset(Exp1_R25_pept[[2]])

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
