% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_func.R
\name{GraphPepProt_hc}
\alias{GraphPepProt_hc}
\title{Function to create a histogram that shows the repartition of
peptides w.r.t. the proteins}
\usage{
GraphPepProt_hc(X, type = "all")
}
\arguments{
\item{X}{An adjacency matrix.}

\item{type}{A string which is the type of matrix (used to build the plot title). Default value is 'all'.}
}
\value{
A histogram
}
\description{
Method to plot the disrtibution (histogram) of peptides w.r.t the proteins with proteins and peptides 
in an adjacency matrix
}
\examples{
obj <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
X <- rowData(obj[[2]])$adjacencyMatrix
GraphPepProt_hc(X)

}
\author{
Alexia Dorffer, Samuel Wieczorek
}
