% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_heatmap.R
\name{heatmapD}
\alias{heatmapD}
\title{This function is a wrapper to \code{\link{heatmap.2}} that displays
a numeric matrix}
\usage{
heatmapD(qData, distance = "euclidean", cluster = "complete", dendro = FALSE)
}
\arguments{
\item{qData}{A dataframe of numeric values}

\item{distance}{The distance used by the clustering algorithm to compute
the dendrogram. See \code{help(heatmap.2)}}

\item{cluster}{the clustering algorithm used to build the dendrogram.
See \code{help(heatmap.2)}}

\item{dendro}{A boolean to indicate if the dendrogram has to be displayed}
}
\value{
A heatmap
}
\description{
Heatmap of the quantitative proteomic data of a \code{data.frame} object
}
\examples{
\dontrun{
utils::data(Exp1_R25_pept, package='DAPARdata2')
ft <- Exp1_R25_pept
qData <- assay(ft[['original_log']])[1:1000,]
heatmapD(qData)
}

}
\author{
Samuel Wieczorek
}
