% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffAnalysis.R
\name{histPValue_HC}
\alias{histPValue_HC}
\title{Plots a histogram of p-values}
\usage{
histPValue_HC(pval_ll, bins = 80, pi0 = 1)
}
\arguments{
\item{pval_ll}{A data frame which contains the p-values for the set of comparisons to show.}

\item{bins}{xxx}

\item{pi0}{xxx}
}
\value{
A plot
}
\description{
Plots a histogram of p-values
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept[1:1000,]
object <- addAssay(object, QFeatures::filterNA(object[[2]],  pNA = 0), name='filtered')
sTab <- colData(object)
se <- t_test_sam(object[[3]], sTab, FUN = compute.t.test)
pval <- metadata(se)$HypothesisTest_cols[['25fmol_vs_10fmol_pval']]
histPValue_HC(pval)

}
\author{
Samuel Wieczorek
}
