% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_test.R
\docType{data}
\name{HypothesisTest}
\alias{HypothesisTest}
\alias{HypothesisTest,SummarizedExperiment-method}
\alias{HypothesisTest,QFeatures-method}
\title{Record the adjacency matrices}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
HypothesisTest

\S4method{HypothesisTest}{SummarizedExperiment}(object, test, sTab, ...)

\S4method{HypothesisTest}{QFeatures}(object, i, name = "HypothesisTestAssay", test = NULL, ...)
}
\arguments{
\item{object}{xxx}

\item{test}{xxx}

\item{sTab}{xxx}

\item{...}{xxx}

\item{i}{xxx}

\item{name}{xxx}
}
\description{
Record the adjacency matrices
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
Exp1_R25_pept <- addAssay(Exp1_R25_pept, QFeatures::filterNA(Exp1_R25_pept[[2]],  pNA = 0), name='filtered')

Exp1_R25_pept <- HypothesisTest(object = Exp1_R25_pept, i = 3, test = 't_test', contrast='OnevsOne', type='Student')

Exp1_R25_pept <- HypothesisTest(object = Exp1_R25_pept, i=3, test = 'limma', comp.type = "OnevsOne")

Exp1_R25_pept <- HypothesisTest(object = Exp1_R25_pept, i=3, test = '1wayAnova', with_post_hoc = TRUE, post_hoc_test = 'Dunnett')

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
