% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_qfeatures.R
\name{writeQFeaturesToExcel}
\alias{writeQFeaturesToExcel}
\alias{readExcel}
\alias{listSheets}
\title{This function exports a \code{QFeatures} object to a Excel file.}
\usage{
writeQFeaturesToExcel(obj, filename)

readExcel(file, extension, sheet)

listSheets(file)
}
\arguments{
\item{obj}{An object of class \code{QFeatures}.}

\item{filename}{A character string for the name of the Excel file.}

\item{file}{The name of the Excel file.}

\item{extension}{The extension of the file}

\item{sheet}{The name of the sheet}
}
\value{
A Excel file (.xlsx)

A data.frame

A vector

A list of sheets name
}
\description{
This function exports a \code{QFeatures} data object to a Excel file.
Each of the three data.frames in the \code{QFeatures} object (ie expression data,
colData and rowData are respectively integrated into separate sheets in
the Excel file). The colored cells in the experimental data correspond to the 
original missing values which have been imputed.
}
\examples{
\donttest{
Sys.setenv("R_ZIPCMD"= Sys.which("zip"))
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000)]
writeQFeaturesToExcel(obj, "foo")
}

}
\author{
Samuel Wieczorek

Samuel Wieczorek

Samuel Wieczorek
}
