% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2_qfeatures.R
\name{createQFeatures}
\alias{createQFeatures}
\title{Creates an object of class \code{QFeatures} from text file.}
\usage{
createQFeatures(
  data = NULL,
  file = NULL,
  sample,
  indQData,
  keyId = "AutoID",
  indexForMetacell = NULL,
  force.na = TRUE,
  typeDataset,
  parentProtId = NULL,
  analysis = "foo",
  processes = NULL,
  typePipeline = NULL,
  software = NULL,
  name = "original"
)
}
\arguments{
\item{data}{The name of a tab-separated file that contains the data.}

\item{file}{A \code{character(1)}. The name of a file xxx}

\item{sample}{A dataframe describing the samples (in lines).}

\item{indQData}{A vector of string where each element is the name
of a column in designTable that have to be integrated in
the \code{rowData()} table of the \code{QFeatures} object.}

\item{keyId}{A \code{character(1)} or \code{numeric(1)} which is the indice of the
column containing the ID of entities (peptides or proteins)}

\item{indexForMetacell}{xxxxxxxxxxx}

\item{force.na}{A \code{boolean} that indicates if the '0' and 'NaN' values of
quantitative values  must be replaced by 'NA' (Default is FALSE)}

\item{typeDataset}{A string that indicates whether the dataset is about}

\item{parentProtId}{A \code{character(1)} For peptide entities, a string which
is the name of a column in rowData. It contains the id of parent proteins
and is used to generate adjacency matrix and process to aggregation.}

\item{analysis}{A \code{character(1)} which is the name of the MS study.}

\item{processes}{A vector of A \code{character()} which contains the name of
processes which has already been run on the data. Default is 'original'.}

\item{typePipeline}{A \code{character(1)} The type of pipeline used with this
dataset. The list of predefined pipelines in DaparToolshed can be obtained
with the function \code{pipelines()}. Default value is NULL}

\item{software}{A \code{character(1)}}

\item{name}{A \code{character(1)} which is the name of the assay in the
QFeatures object. Default is 'original'}
}
\value{
An instance of class \code{QFeatures}.
}
\description{
Creates an object of class \code{QFeatures} from a
single tabulated-like file for quantitative and meta-data and a dataframe
for the samples description.
}
\examples{
inst/extdata/examples/ex_createQFeatures.R

}
\author{
Samuel Wieczorek
}
