% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{impute_pa}
\alias{impute_pa}
\title{Imputation of peptides having no values in a biological condition.}
\usage{
impute_pa(x, conds, q.min = 0.025)
}
\arguments{
\item{x}{xxxxx}

\item{conds}{xxxx}

\item{q.min}{Same as the function \code{impute.pa} in the package \code{imp4p}}
}
\value{
The xxxx.

NA
}
\description{
Imputation of peptides having no values in a biological condition.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept[seq_len(1000),]
object <- addAssay(object, QFeatures::filterNA(object[[2]],  pNA = 0.2), name='filtered')
imp <- impute_pa(assay(object[['filtered']]), colData(object)$Condition)

}
\author{
Samuel Wieczorek
}
