% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.iter}
\alias{inner.aggregate.iter}
\title{Method to aggregate peptides into proteins with the iterative approach.}
\usage{
inner.aggregate.iter(
  qPepData,
  X,
  init.method = "Sum",
  method = "Mean",
  n = NULL
)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}

\item{init.method}{The method used to initialize the iterative algotirhm. Default is 'Sum'.}

\item{method}{The method used for xxx. Default value is 'Mean'.}

\item{n}{An integer which is xxx}
}
\value{
A  matrix containing the quantitative aggregated data for proteins.
}
\description{
Method to aggregate peptides into proteins with the iterative approach.
}
\author{
Samuel Wieczorek, Thomas Burger
}
