% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_func.R
\name{inner.aggregate.iter}
\alias{inner.aggregate.iter}
\title{Method to aggregate peptides into proteins with the iterative approach.}
\usage{
inner.aggregate.iter(
  qData,
  X,
  init.method = "colSumsMat",
  iter.method = "Mean",
  n = NULL
)
}
\arguments{
\item{X}{An adjacency matrix}

\item{init.method}{The method used to initialize the iterative algotirhm. Default is 'Sum'.}

\item{n}{An integer which is xxx}

\item{qPepData}{A data.frame of quantitative data}

\item{method}{The method used for xxx. Default value is 'Mean'.}
}
\value{
A  matrix containing the quantitative aggregated data for proteins.
}
\description{
Method to aggregate peptides into proteins with the iterative approach.
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000),]
obj <- addListAdjacencyMatrices(obj, 2)
X <- GetAdjMat(obj[[2]])$all
qPepData <- assay(obj[[2]])
inner.aggregate.iter(qPepData, X)

}
\author{
Samuel Wieczorek, Thomas Burger
}
