% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.aggregate.topn}
\alias{inner.aggregate.topn}
\title{Method to aggregate peptides into proteins with the top n approach.}
\usage{
inner.aggregate.topn(qPepData, X, method = "Mean", n = 10)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}

\item{method}{Method used to aggregate (see function xxx)}

\item{n}{An integer which is the number of top peptides used for aggregation.}
}
\value{
A matrix containing the quantitative aggregated data for proteins.
}
\description{
Method to aggregate peptides into proteins with the top n approach.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original']][1:1000,])[1:1000,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
inner.aggregate.topn(assay(Exp1_R25_pept[['original_log']][1:1000,]), X.list$all)

}
\author{
Samuel Wieczorek
}
