% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{inner.aggregate.topn}
\alias{inner.aggregate.topn}
\title{Method to aggregate peptides into proteins with the top n approach.}
\usage{
inner.aggregate.topn(qPepData, X, method = "Mean", n = 10)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}

\item{method}{Method used to aggregate (see function xxx)}

\item{n}{An integer which is the number of top peptides used for aggregation.}
}
\value{
A matrix containing the quantitative aggregated data for proteins.
}
\description{
Method to aggregate peptides into proteins with the top n approach.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- GetAdjMat(obj[[2]])$all
inner.aggregate.topn(assay(obj[[2]]), X, n=3)

}
\author{
Samuel Wieczorek
}
