% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{inner.mean}
\alias{inner.mean}
\title{Aggregate peptides into proteins with the mean of the quantitative data per conditions.}
\usage{
inner.mean(qPepData, X)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
A matrix containing the quantitative aggregated data for proteins.
}
\description{
Method to aggregate peptides into proteins with the mean of the quantitative data per conditions.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000,]
obj <- addListAdjacencyMatrices(obj, 2)
X <- GetAdjMat(obj[[2]])$all
inner.mean(assay(obj[[2]]), X)

}
\author{
Samuel Wieczorek
}
