% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{inner.sum}
\alias{inner.sum}
\title{aggregate peptides into proteins with the sum of the quantitative data per conditions.}
\usage{
inner.sum(qPepData, X)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data}

\item{X}{An adjacency matrix}
}
\value{
A matrix containing the quantitative aggregated data for proteins.
}
\description{
Method to aggregate peptides into proteins with the sum of the quantitative data per conditions.
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original']][1:1000,])[1:1000,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
n <- inner.sum(assay(Exp1_R25_pept[['original_log']])[1:1000,], X.list$all)
}
\author{
Samuel Wieczorek
}
