% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_boxplot.R, R/plots_violin.R,
%   R/mod_plots_intensity.R
\name{boxPlotD}
\alias{boxPlotD}
\alias{violinPlotD}
\alias{intensity_plots}
\alias{mod_intensity_plots_ui}
\alias{mod_intensity_plots_server}
\title{Intensity family plots}
\usage{
boxPlotD(object, exp.design, subset = NULL, pal.name = NULL)

violinPlotD(object, exp.design, subset = NULL, pal.name)

mod_intensity_plots_ui(id)

mod_intensity_plots_server(id, object, exp.design, ...)
}
\arguments{
\item{object}{xxx}

\item{exp.design}{xxx}

\item{subset}{A vector of index indicating rows to highlight}

\item{pal.name}{xxx}

\item{id}{xxx}

\item{...}{xxx}
}
\value{
A boxplot

A violinplot

NA

NA
}
\description{
These functions are plotting different views of quantitative data.
Two plots are available:
\itemize{
\item \code{\link[=boxplotD]{boxplotD()}} xxxx,
\item \code{\link[=violinPlotD]{violinPlotD()}} xxxx.
}
}
\examples{
library(QFeatures)
data(ft)

#----------------------------------------------
# View violin plot with and without subset view
#----------------------------------------------

violinPlotD(ft[[1]], design(ft))
violinPlotD(ft[[1]], design(ft), subset = c(3,4,6), pal.name = 'Dark2')

#----------------------------------------
# xxxx
#----------------------------------------

boxPlotD(ft[[1]], design(ft))
boxPlotD(ft[[1]], design(ft), subset = c(2, 3), pal.name = 'Dark2')

# ----------------------------------------------------
# A shiny module which can track entities on the plots
# ----------------------------------------------------

if (interactive()){
library(QFeatures)
library(shiny)
library(shinyjs)
library(highcharter)
library(DaparToolshed)

data(ft)


ui <- tagList(
  mod_intensity_plots_ui('iplots'),
  uiOutput('show')
)

server <- function(input, output, session) {
  rv <- reactiveValues(
    tmp = NULL
  )
  
  mod_intensity_plots_server(id = 'iplots',
                             object = reactive({ft[[1]]}),
                             exp.design = reactive({design(ft)}),
                             pal.name = 'Dark2'
                             )
  
}
shinyApp(ui = ui, server = server)
}

}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy

Samuel Wieczorek, Anais Courtier, Enora Fremy
}
