% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaAnalysis.R
\name{make.design.3}
\alias{make.design.3}
\title{Builds the design matrix for designs of level 3}
\usage{
make.design.3(sTab)
}
\arguments{
\item{sTab}{The data.frame which correspond to the definitino of the replicates}
}
\value{
A design matrix
}
\description{
Builds the design matrix for designs of level 3
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
sTab <-cbind(colData(Exp1_R25_pept), Tech.Rep=1:6)
make.design.3(sTab)

}
\author{
Thomas Burger, Quentin Giai-Gianetto, Samuel Wieczorek
}
