% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Agregation.R
\name{mod_agregation}
\alias{mod_agregation}
\alias{mod_Agregation_ui}
\title{Module agregation}
\usage{
mod_Agregation_ui(id)
}
\arguments{
\item{id}{}
}
\description{
xxxxxx
}
\examples{
if(interactive()){
library(magellan)
 run_workflow('Agregation', verbose = TRUE)


}
}
