% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_all_plots.R
\name{mod_all_plots_server}
\alias{mod_all_plots_server}
\title{all_plots UI Function}
\usage{
mod_all_plots_server(id, dataIn)
}
\arguments{
\item{id}{xxxx}

\item{dataIn}{xxxx}
}
\description{
A shiny Module.
}
