% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert.R
\name{Convert_ui}
\alias{Convert_ui}
\alias{Convert_server}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
Convert_ui(id)

Convert_server(
  id,
  dataIn = NULL,
  steps.enabled = reactive({
     NULL
 }),
  remoteReset = reactive({
     FALSE
 }),
  steps.status = reactive({
     NULL
 }),
  current.pos = reactive({
     1
 }),
  verbose = FALSE
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{steps.enabled}{xxx}

\item{remoteReset}{xxx}
}
\value{
NA

NA
}
\description{
A shiny Module.
}
\keyword{internal}
