% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert.R
\name{mod_convert}
\alias{mod_convert}
\alias{Convert_conf}
\alias{Convert_ui}
\alias{Convert_server}
\title{xxx}
\usage{
Convert_conf()

Convert_ui(id)

Convert_server(
  id,
  dataIn = NULL,
  steps.enabled = reactive({
     NULL
 }),
  remoteReset = reactive({
     FALSE
 }),
  steps.status = reactive({
     NULL
 }),
  current.pos = reactive({
     1
 }),
  verbose = FALSE
)
}
\arguments{
\item{id}{xxx}

\item{dataIn}{xxx}

\item{steps.enabled}{xxx}

\item{remoteReset}{xxx}
}
\value{
NA

NA
}
\description{
xxx

A shiny Module.
}
\examples{


ui <- fluidPage(
  nav_ui('Convert')
  
)


server <- function(input, output){
  
  rv <- reactiveValues(
    dataIn = NULL,
    dataOut = NULL
  )
  
  observe({
    rv$dataOut <- nav_server(id = 'Convert',
                             dataIn = reactive({data.frame()}))
  })
  
  observeEvent(rv$dataOut, {
    print('toto')
    browser()
  })
}


shinyApp(ui, server)
}
\author{
Samuel Wieczorek
}
\keyword{internal}
