% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_convert.R
\name{mod_Convert_ui}
\alias{mod_Convert_ui}
\title{mod_choose_pipeline_ui and mod_choose_pipeline_server}
\usage{
mod_Convert_ui(id)
}
\arguments{
\item{id}{shiny id}
}
\description{
A shiny Module.
}
\keyword{internal}
