% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_Filtering.R
\name{mod_filtering}
\alias{mod_filtering}
\title{Module Filtering}
\description{
xxxxx
}
\section{Step 'Quanti metadata filtering'}{


xxxxxxx
}

\section{Step 'Variable filtering'}{


xxxxx
}

\section{Step 'Save'}{


xxxxxx
}

\examples{
if(interactive()){
 run_workflow('Filtering', verbose = TRUE)
}
}
\seealso{
The user manual of the package \code{Magellan}.
}
