% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering_MV.R
\name{MVrowsTagToOne}
\alias{MVrowsTagToOne}
\title{Filter lines in the matrix of intensities w.r.t. some criteria}
\usage{
MVrowsTagToOne(
  obj = NULL,
  sampleTab = NULL,
  type = NULL,
  th = 0,
  percent = TRUE,
  newColName = "newCol"
)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment}}

\item{sampleTab}{\code{colData()} of obj}

\item{type}{Method used to choose the lines to delete.
Values are : "None", "EmptyLines", "WholeMatrix", "AllCond", "AtLeastOneCond"}

\item{th}{Either a numeric between 0 and 1, or a integer between 0 and maximum number of samples.}

\item{percent}{TRUE by default. When FALSE, use the number of samples}

\item{newColName}{Name of the new column containing the filtration information}
}
\value{
The object of class \code{SummarizedExperiment} with extra column in rowData
indicating 1 for the lines to remove, else 0.
}
\description{
Returns the \code{SummarizedExperiment} object with a extra column in \code{rowData()}.
The extra column indicates feature(s) to delete in 1 and 0 otherwise.
The user chooses the threshold, either the percentage of NAS per row (default)
or the number of samples containing NAs per row allowed. Then, 
the filter tags the lines that do not respect this condition.
The condition may be on the whole line or condition by condition.

The different methods are :
- "WholeMatrix": given a threshold \code{th}, only the lines that contain
  at least \code{th} values are kept.
- "AllCond": given a threshold \code{th}, only the lines which contain
  at least \code{th} values for each of the conditions are kept.
- "AtLeastOneCond": given a threshold \code{th}, only the lines that contain
  at least \code{th} values, and for at least one condition, are kept.
}
\examples{
library(Features)
utils::data(Exp1_R25_pept, package='DAPARdata2')
sampleTab <- colData(Exp1_R25_pept)
obj <- Exp1_R25_pept[[2]][100:120,]
obj <- MVrowsTagToOne(obj, sampleTab, "None", 3, percent=F, "None")
obj <- MVrowsTagToOne(obj, sampleTab, "EmptyLines", 3, percent=F, "EmptyLines")
obj <- MVrowsTagToOne(obj, sampleTab, "WholeMatrix", 3, percent=F, "WholeMatrix")
obj <- MVrowsTagToOne(obj, sampleTab, "AllCond", 3, percent=F, "AllCond")
res <- MVrowsTagToOne(obj, sampleTab, "AtLeastOneCond", 3, percent=F, "AtLeastOneCond")
as.data.frame(rowData(res)[,72:76])

}
\author{
Enora Fremy
}
