% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{MVrowsTagToOne}
\alias{MVrowsTagToOne}
\title{Filter lines in the matrix of intensities w.r.t. some criteria}
\usage{
MVrowsTagToOne(object, type, th = 0, percent = TRUE)
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{type}{Method used to choose the lines to delete.
Values are : "None", "EmptyLines", "WholeMatrix", "AllCond", "AtLeastOneCond"}

\item{th}{Either a numeric between 0 and 1 where only the lines which contain
at least \code{th}% of non NA values are kept.
Or a integer between 0 and maximum number of samples for 'WholeMatrix' and 
between 0 and maximum number of replicate for "AllCond" and "AtLeastOneCond", where
only the lines which contain at least \code{th} values are kept.}

\item{percent}{TRUE by default. When FALSE, use the number of samples}
}
\value{
The object of class \code{SummarizedExperiment} with extra column in rowData
indicating 1 for the lines to remove, else 0.
}
\description{
Returns the \code{SummarizedExperiment} object with a extra column in \code{rowData()}.
The extra column indicates feature(s) to delete in 1 and 0 otherwise.
The user chooses the threshold, either the percentage of NAS per row (default)
or the number of samples containing NAs per row allowed. Then, 
the filter tags the lines that do not respect this condition.
The condition may be on the whole line or condition by condition.

The different methods are :
- "WholeMatrix": given a threshold \code{th}, only the lines that contain
  at least \code{th} values are kept.
- "AllCond": given a threshold \code{th}, only the lines which contain
  at least \code{th} values for each of the conditions are kept.
- "AtLeastOneCond": given a threshold \code{th}, only the lines that contain
  at least \code{th} values, and for at least one condition, are kept.
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept
res <- MVrowsTagToOne(object, type = "AtLeastOneCond", th=0.7, percent=TRUE)

}
\author{
Enora Fremy, Samuel Wieczorek
}
