% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nEmptyLines}
\alias{nEmptyLines}
\title{Number of empty lines in the data}
\usage{
nEmptyLines(df)
}
\arguments{
\item{df}{A \code{data.frame}.}
}
\value{
A \code{integer(1)}
}
\description{
This function counts the number of empty lines (all elements
are equal to NA).
}
\examples{
data(ft, package='DaparToolshed')
nEmptyLines(assay(ft, 1))

}
\author{
Samuel Wieczorek
}
