% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ds_pca.R
\name{plotPCA_Eigen}
\alias{plotPCA_Eigen}
\title{#' @param res.pca Result of FactoMineR::PCA
#' @param chosen.axes The dimensions to plot
#'
#' @author Samuel Wieczorek, Enora Fremy
#'
#' @importFrom factoextra fviz_pca_var
#'
#' @export
#'
#' @rdname pca_plots
#'
plotPCA_Var <- function(res.pca = NULL,
chosen.axes = c(1, 2)){
if (is.null(res.pca))
return(NULL)

fviz_pca_var(res.pca,
axes = chosen.axes,
col.var = "cos2",
gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
repel = TRUE
)
}
#' @param res.pca Result of FactoMineR::PCA
#' @param axes The dimensions to plot
#' @param geom xxx
#'
#' @author Samuel Wieczorek, Enora Fremy
#'
#' @importFrom factoextra fviz_pca_ind
#'
#' @export
#'
#' @rdname pca_plots
#'
plotPCA_Ind <- function(res.pca, chosen.axes = c(1,2)){
if (is.null(res.pca))
return(NULL)

fviz_pca_ind(res.pca,
axes = chosen.axes,
geom  ="point")

}}
\usage{
plotPCA_Eigen(res.pca)
}
\arguments{
\item{res.pca}{Result of FactoMineR::PCA}
}
\description{
#' @param res.pca Result of FactoMineR::PCA
#' @param chosen.axes The dimensions to plot
#'
#' @author Samuel Wieczorek, Enora Fremy
#'
#' @importFrom factoextra fviz_pca_var
#'
#' @export
#'
#' @rdname pca_plots
#'
plotPCA_Var <- function(res.pca = NULL,
chosen.axes = c(1, 2)){
if (is.null(res.pca))
return(NULL)

fviz_pca_var(res.pca,
axes = chosen.axes,
col.var = "cos2",
gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07"),
repel = TRUE
)
}
#' @param res.pca Result of FactoMineR::PCA
#' @param axes The dimensions to plot
#' @param geom xxx
#'
#' @author Samuel Wieczorek, Enora Fremy
#'
#' @importFrom factoextra fviz_pca_ind
#'
#' @export
#'
#' @rdname pca_plots
#'
plotPCA_Ind <- function(res.pca, chosen.axes = c(1,2)){
if (is.null(res.pca))
return(NULL)

fviz_pca_ind(res.pca,
axes = chosen.axes,
geom  ="point")

}
}
\author{
Samuel Wieczorek, Enora Fremy
}
