% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectedComps.R
\name{plotJitter_hc}
\alias{plotJitter_hc}
\title{Display a a jitter plot for CC}
\usage{
plotJitter_hc(df, clickFunction = NULL)
}
\arguments{
\item{df}{A dataframe containing xxx}

\item{clickFunction}{A JavaScript code to show an information when a point is clicked}
}
\value{
A plot
}
\description{
Display a jitter plot for connex component with the package highcharter
}
\examples{
df <- data.frame(x=1:10, y=1:10)
plotJitter_hc(df)

}
\author{
Thomas Burger, Samuel Wieczorek
}
