% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pep_prot_cc.R
\name{plotJitterCC}
\alias{plotJitterCC}
\title{Jitter plot of CC}
\usage{
plotJitterCC(list.of.cc)
}
\arguments{
\item{list.of.cc}{List of cc such as returned by the function get.pep.prot.cc}
}
\value{
A plot
}
\description{
Jitter plot of CC
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2') 
plist <- rowData(Exp1_R25_pept[['original_log']][1:1000,])[,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original_log']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
ll <- get.pep.prot.cc(X.list$all)
plotJitterCC(ll)

}
\author{
Thomas Burger
}
