% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectedComps.R
\name{plotJitterCC}
\alias{plotJitterCC}
\title{Jitter plot of CC}
\usage{
plotJitterCC(list.of.cc)
}
\arguments{
\item{list.of.cc}{List of cc such as returned by the function get.pep.prot.cc}
}
\value{
A plot
}
\description{
Jitter plot of CC
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2') 
X <- GetAdjMat(Exp1_R25_pept[[2]])$all
ll <- get.pep.prot.cc(X)
plotJitterCC(ll)

}
\author{
Thomas Burger
}
