% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_pca.R
\name{plotPCA_Var}
\alias{plotPCA_Var}
\title{Plots variables of PCA}
\usage{
plotPCA_Var(res.pca = NULL, chosen.axes = c(1, 2))
}
\arguments{
\item{res.pca}{Result of FactoMineR::PCA}

\item{chosen.axes}{The dimensions to plot}
}
\value{
A plot
}
\description{
Plots variables of PCA
}
\author{
Samuel Wieczorek, Enora Fremy
}
