% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{proportionConRev_HC}
\alias{proportionConRev_HC}
\title{Barplot of proportion of contaminants and reverse}
\usage{
proportionConRev_HC(lDataset, nBoth = 0, nCont = 0, nRev = 0)
}
\arguments{
\item{lDataset}{The total length (number of rows) of the dataset}

\item{nBoth}{The number of both contaminants and reverse identified in the dataset.}

\item{nCont}{The number of contaminants identified in the dataset.}

\item{nRev}{The number of reverse entities identified in the dataset.}
}
\value{
A barplot
}
\description{
getPourcentageOfMV <- function(obj)
Voir nNA(object, i) de la classe QFeatures
}
\examples{
proportionConRev_HC(100, 10, 20)

}
\author{
Samuel Wieczorek
}
