% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ProstarVersions}
\alias{ProstarVersions}
\title{Versions of installed packages of Prostar suite}
\usage{
ProstarVersions()
}
\value{
A list
}
\description{
Return the versions of Prostar, DaparToolshed and DaparData
}
\examples{
ProstarVersions()

}
\author{
Samuel Wieczorek
}
