% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ProstarVersions}
\alias{ProstarVersions}
\title{Version number of Prostar suite}
\usage{
ProstarVersions()
}
\value{
A \code{list(3)}
}
\description{
This function gives the version number of the packages of the
Prostar suite and which propose data processing. This information
can be useful if the user wants to publish its works or to
rerun a data processing pipeline tin a given set of conditions.
The packages which are concerned are \code{Prostar}, \code{DaparToolshed}
and \code{DaparToolshedData}
}
\examples{
ProstarVersions()

}
\author{
Samuel Wieczorek
}
