% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_qfeatures.R
\name{QFeatures-excel}
\alias{QFeatures-excel}
\alias{write2excel,QFeatures-method}
\alias{write2excel,SummarizedExperiment-method}
\title{This function exports a \code{QFeatures} object to a Excel file.}
\usage{
\S4method{write2excel}{QFeatures}(object, i = NULL, filename = "newFile", ...)

\S4method{write2excel}{SummarizedExperiment}(object, filename, exp.design, ...)
}
\arguments{
\item{filename}{A character string for the name of the Excel file.}

\item{obj}{An object of class \code{QFeatures}.}
}
\value{
A Excel file (.xlsx)
}
\description{
This function exports a \code{QFeatures} data object to a Excel file.
Each of the three data.frames in the \code{QFeatures} object (ie expression data,
colData and rowData are respectively integrated into separate sheets in
the Excel file). The colored cells in the experimental data correspond to the
original missing values which have been imputed.
}
\author{
Samuel Wieczorek
}
