% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_qfeatures.R
\name{QFeatures-excel}
\alias{QFeatures-excel}
\alias{write2excel,QFeatures-method}
\alias{write2excel,SummarizedExperiment-method}
\alias{.write2excel}
\alias{addColors}
\title{Exports a \code{QFeatures} object to a Excel file.}
\usage{
\S4method{write2excel}{QFeatures}(object, i = NULL, filename = "newFile", ...)

\S4method{write2excel}{SummarizedExperiment}(object, filename, exp.design, ...)

.write2excel(object, filename, exp.design)

addColors(wb, n, tags, colors)
}
\arguments{
\item{object}{xxx}

\item{i}{xxx}

\item{filename}{xxx}

\item{...}{xxx}

\item{exp.design}{xxx}

\item{wb}{A workbook}

\item{n}{A \code{integer(1)} which is the number of sheet in the workbook.}

\item{tags}{xxx}

\item{colors}{A \code{character()} which contains the HEX code for colors.
The size of this vector must be the same as the number of tags.}
}
\value{
A Excel file.
}
\description{
This function exports an instance of the class \code{QFeatures} to a Excel file.
The resulting file is composed of four sheets:
\itemize{
\item \verb{quantitative data} which contains the content of \code{assay()} object with a
color code for each cell w.r.t. to cell quantitative metadata.
\item \code{metadata} which is the content of \code{rowData()} with only one-dimensionnal
data (i.e. the adjacencyMatrix and the qMetacell slots are not part of
the sheet),
\item \verb{exp. design} which is the content of \code{colData()}. Each condition in the
table is colored with a different color,
\item \verb{quantitative metadata} which is the content of \code{qMetacell()}. There is a
color code for the different tags.
}

xxx
}
\examples{
data(ft, package='DaparToolshed')

#---------------------------------------
# Export the whole dataset
#---------------------------------------

write2excel(ft, filename = "foo")
unlink('foo.xls')
write2excel(ft, 1, "foo")
unlink('foo.xls')

}
\author{
Samuel Wieczorek
}
