% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filter_qMetadata.R
\name{qMetadataFilteringScope}
\alias{qMetadataFilteringScope}
\alias{SymFilteringOperators}
\alias{qMetadatWholeMatrix}
\alias{qMetadataWholeLine}
\alias{qMetadataOnConditions}
\title{Quantitative cell metadata scopes for filtering}
\usage{
qMetadataFilteringScope()

SymFilteringOperators()

qMetadatWholeMatrix(object, cmd, pattern, percent, th, operator)

qMetadataWholeLine(object, cmd, pattern)

qMetadataOnConditions(object, cmd, mode, pattern, conds, percent, operator, th)
}
\arguments{
\item{percent}{xxx}

\item{th}{The theshold to apply}

\item{conds}{xxx}

\item{mask}{xxx}

\item{type}{Available values are:
* 'AllCond' (the query is valid in all the conditions),
* 'AtLeatOneCond' (the query is valid in at leat one condition.}

\item{op}{String for operator to use. List of operators is available with SymFilteringOperators().}
}
\value{
NA

NA

NA

NA

NA

#' @examples
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(10)]
level <- obj@experimentData@other$typeOfData
pattern <- 'missing'
mask <- match.qMetadata(metadata=qMetadata(obj), pattern=pattern, level=level)
type <- 'AllCond'
conds <- Biobase::pData(obj)$Condition
op <- '>='
th <- 2
percent <- FALSE
ind <- GetIndices_BasedOnConditions(mask, type, conds, percent, op, th)
}
\description{
This function looks for the lines that respect the request in either all conditions
or at least one condition.

This function looks for the lines where each element respect the query.

This function looks for the lines that respect the request in either all conditions
or at least one condition.
}
\examples{
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(10)]
level <- obj@experimentData@other$typeOfData
pattern <- 'missing'
mask <- match.qMetadata(metadata=Get_qMetadata(obj), pattern=pattern, level=level)
percent <- FALSE
th <- 3
op <- '>='
ind <- GetIndices_WholeMatrix(mask, op, percent, th)

Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[20:30]
level <- obj@experimentData@other$typeOfData
pattern <- 'missing POV'
mask <- match.qMetadata(metadata=GetqMetadata(obj), pattern=pattern, level=level)
ind <- GetIndices_WholeLine(mask)

}
