% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DaparToolshed_filter_qMetadata.R
\name{qMetadataFilteringScope}
\alias{qMetadataFilteringScope}
\alias{SymFilteringOperators}
\alias{qMetadataWholeMatrix}
\alias{qMetadataWholeLine}
\alias{qMetadataOnConditions}
\title{Quantitative cell metadata scopes for filtering}
\usage{
qMetadataFilteringScope()

SymFilteringOperators()

qMetadataWholeMatrix(object, cmd, pattern, percent, th, operator)

qMetadataWholeLine(object, cmd, pattern)

qMetadataOnConditions(object, cmd, mode, pattern, conds, percent, operator, th)
}
\arguments{
\item{object}{xxx}

\item{cmd}{Available values are:}

\item{pattern}{ss}

\item{percent}{xxx}

\item{th}{The theshold to apply}

\item{operator}{String for operator to use. List of operators is available with SymFilteringOperators().}

\item{mode}{xxx}

\item{conds}{xxx}
}
\value{
NA

NA

NA

NA

NA
}
\description{
Quantitative cell metadata scopes for filtering

Operators for complex queries
}
