% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering_MV.R
\name{removeAdditionalCol}
\alias{removeAdditionalCol}
\title{Restore the rowData() before MVindicesToZero}
\usage{
removeAdditionalCol(obj, newColName)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment}}

\item{newColName}{Column to remove}
}
\value{
The \code{SummarizedExperiment} object without the rowData newColName column
}
\description{
Returns the \code{SummarizedExperiment} object without the \code{rowData()} extra column
from MVindicesToZero
}
\examples{
library(DAPAR2)
library(Features)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept[[2]]
sampleTab <- colData(Exp1_R25_pept)
obj<-MVindicesToZero(obj=object, sampleTab, newColName="LinesKept", type="AllCond", th=2)
removeAdditionalCol(obj,newColName="LinesKept" )
}
\author{
Enora Fremy
}
