% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering_MV.R
\name{removeAdditionalCol}
\alias{removeAdditionalCol}
\title{Restore the rowData() before MVrowsTagToOne}
\usage{
removeAdditionalCol(obj, colToRemove)
}
\arguments{
\item{obj}{An object of class \code{SummarizedExperiment}}

\item{colToRemove}{Column to remove}
}
\value{
The \code{SummarizedExperiment} object without the rowData colToRemove column
}
\description{
Returns the \code{SummarizedExperiment} object without the \code{rowData()} extra column
from MVrowsTagToOne
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept[[2]]
sampleTab <- colData(Exp1_R25_pept)
obj<-MVrowsTagToOne(obj=object, sampleTab, newColName="LinesKept", type="AllCond", th=2)
obj<-removeAdditionalCol(obj,colToRemove="None" )

}
\author{
Enora Fremy
}
