% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filtering.R
\name{removeAdditionalCol}
\alias{removeAdditionalCol}
\title{Restore the rowData() before MVrowsTagToOne}
\usage{
removeAdditionalCol(object, colToRemove = NULL)
}
\arguments{
\item{object}{An object of class \code{QFeatures}}

\item{colToRemove}{Column to remove}
}
\value{
The \code{SummarizedExperiment} object without the rowData colToRemove column
}
\description{
Restore the rowData() before MVrowsTagToOne
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
object <- Exp1_R25_pept[[2]]
sampleTab <- colData(Exp1_R25_pept)
obj <- MVrowsTagToOne(obj=object, i=2, sampleTab, newColName="LinesKept", type="AllCond", th=2)
obj <- removeAdditionalCol(obj, i=2, colToRemove="LinesKept" )

}
\author{
Enora Fremy, Samuel Wieczorek
}
