% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_2_qfeatures.R
\name{ReplaceSpecialChars}
\alias{ReplaceSpecialChars}
\title{Standardize names}
\usage{
ReplaceSpecialChars(x)
}
\arguments{
\item{x}{A vector of strings to be processed}
}
\description{
Replace some characters in names by 'underscore'
}
\author{
Samuel Wieczorek
}
