% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingValuesImputation_ProteinLevel.R
\name{restore_MEC_matrix}
\alias{restore_MEC_matrix}
\title{Put back LAPALA into  a \code{MSnSet} object}
\usage{
restore_MEC_matrix(x, conds, MECIndex)
}
\arguments{
\item{MECIndex}{A data.frame that contains index of MEC (see findMECBlock) .}

\item{obj}{An object of class \code{Su}.}
}
\value{
The object \code{obj} where LAPALA have been reintroduced
}
\description{
This method is used to put back the LAPALA that have been identified previously
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000]
lapala <- find_MEC(assay(obj[[2]]), colData(obj)$Condition)
assay(obj[[2]]) <- impute_det_quant(assay(obj[[2]]))
assay(obj[[2]]) <- restore_MEC_matrix(assay(obj[[2]]), colData(obj)$Condition, lapala)

}
\author{
Samuel Wieczorek
}
