% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation_func.R
\name{rowdata_stats_Aggregation_sam}
\alias{rowdata_stats_Aggregation_sam}
\title{Finalizes the aggregation process}
\usage{
rowdata_stats_Aggregation_sam(qPepData, X)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data not logged of peptides}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}
}
\value{
A protein object of class \code{SummarizedExperiment}
}
\description{
Finalizes the aggregation process
}
\examples{
library(QFeatures)
Exp1_R25_pept <- readRDS(system.file("data", 'Exp1_R25_pept.rda', package="DaparToolshedData"))
obj <- Exp1_R25_pept[seq_len(1000),]
obj <- addListAdjacencyMatrices(obj, 2)
X <- GetAdjMat(obj[[2]])$all
rowdata_stats_Aggregation_sam(assay(obj,2), X)

}
\author{
Samuel Wieczorek
}
