% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregation.R
\name{rowdata_stats_Aggregation_sam}
\alias{rowdata_stats_Aggregation_sam}
\title{Finalizes the aggregation process}
\usage{
rowdata_stats_Aggregation_sam(qPepData, X)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data not logged of peptides}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}
}
\value{
A protein object of class \code{SummarizedExperiment}
}
\description{
Method to finalize the aggregation process
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
plist <- rowData(Exp1_R25_pept[['original_log']][1:1000,])[,metadata(Exp1_R25_pept)$parentProtId]
names <- names(Exp1_R25_pept[['original_log']])[1:1000]
X.list <- BuildListAdjacencyMatrices(plist, names, type=c('All'))
rowdata_stats_Aggregation_sam(assay(Exp1_R25_pept,2)[1:1000,], X.list$all)

}
\author{
Samuel Wieczorek
}
