% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{rowdata_stats_Aggregation_sam}
\alias{rowdata_stats_Aggregation_sam}
\title{Finalizes the aggregation process}
\usage{
rowdata_stats_Aggregation_sam(qPepData, X)
}
\arguments{
\item{qPepData}{A data.frame of quantitative data not logged of peptides}

\item{X}{An adjacency matrix in which lines and columns correspond 
respectively to peptides and proteins.}
}
\value{
A protein object of class \code{SummarizedExperiment}
}
\description{
Method to finalize the aggregation process
}
\examples{
library(QFeatures)
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:1000]
obj <- SetAdjMat(obj, 2)
X <- GetAdjMat(obj, 2)$all
rowdata_stats_Aggregation_sam(assay(obj,2), X)

}
\author{
Samuel Wieczorek
}
