% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inOutFiles.R
\name{setMEC}
\alias{setMEC}
\title{Sets the MEC tag in the OriginOfValues}
\usage{
setMEC(origin, qData, conds)
}
\arguments{
\item{origin}{xxxxx}

\item{qData}{xxxxxx}

\item{conds}{xxxxxxx}
}
\value{
An instance of class \code{DataFrame} which is an update of the parameter names origin
}
\description{
Sets the MEC tag in the origin parameter
}
\author{
Samuel Wieczorek
}
