% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{standardise}
\alias{standardise}
\title{xxx}
\usage{
standardise(data)
}
\arguments{
\item{data}{xxx}
}
\description{
This code is an adaptation of the standardise function in the package Mfuzz
it works directly on a matrix or data.frame of intensities
}
\author{
Hélène Borges
}
