% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{submatadj}
\alias{submatadj}
\title{Substract part of a adjacency matrix}
\usage{
submatadj(X, onlyShared = FALSE, onlySpec = FALSE)
}
\arguments{
\item{X}{xxx}

\item{onlyShared}{xxx}

\item{onlySpec}{xxx}
}
\value{
A matrix
}
\description{
Substract part of a adjacency matrix
}
\author{
Samuel Wieczorek
}
