% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tracker.R
\name{tracking}
\alias{tracking}
\alias{mod_tracker_ui}
\alias{mod_tracker_server}
\title{Tracking of entities within plots}
\usage{
mod_tracker_ui(id)

mod_tracker_server(id, object)
}
\arguments{
\item{id}{xxx}

\item{object}{A instance of the class \code{SummarizedExperiment}}
}
\value{
NA

A \code{list()} of integers
}
\description{
This shiny module offers a UI to select a subset of a dataset and
superimpose quantitative values of this selection on the complete plot
Three modes of selection are implemented:
\itemize{
\item 'Protein list': xxx,
\item 'Random': xxx,
\item 'Specific column': xxx
}
}
\examples{
if (interactive()) {
    ui <- tagList(
        mod_tracker_ui("track"),
        uiOutput("show")
    )

    server <- function(input, output, session) {
        rv <- reactiveValues(
            tmp = NULL
        )

        rv$tmp <- mod_tracker_server(
            id = "track",
            object = reactive({
                ft[[1]]
            })
        )

        output$show <- renderUI({
            p(paste0(rv$tmp(), collapse = " "))
        })
    }
    shinyApp(ui = ui, server = server)
}
}
