% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metacell.R
\docType{data}
\name{UpdateMetacell}
\alias{UpdateMetacell}
\alias{UpdateMetacell,SummarizedExperiment-method}
\title{Update metacell after imputation}
\format{
An object of class \code{standardGeneric} of length 1.
}
\usage{
UpdateMetacell

\S4method{UpdateMetacell}{SummarizedExperiment}(object, na.type, ...)
}
\arguments{
\item{object}{An object of class `SummarizedExperiment`}

\item{na.type}{xxx}

\item{...}{xxx}
}
\description{
Update the metacell information of missing values that were imputed
}
\examples{
utils::data(Exp1_R25_pept, package='DAPARdata2')
obj <- Exp1_R25_pept[1:10,]
obj[[2]] <- UpdateMetacell(obj[[2]], na.type = 'missing')

}
\author{
Samuel Wieczorek
}
\keyword{datasets}
