% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_violin.R
\name{violinPlotD}
\alias{violinPlotD}
\title{Builds a violinplot from a numeric matrix}
\usage{
<<<<<<< HEAD
violinPlotD(
  qData,
  conds,
  keyId,
=======
<<<<<<< HEAD
violinPlotD(
  qData,
  keyId,
  conds,
>>>>>>> 95fc38b9d2eb90ae8981dffdf89f4a7112dfaa2d
  legend = NULL,
  palette = NULL,
  subset.view = NULL
)
<<<<<<< HEAD
=======
=======
violinPlotD(qData, keyId, legend, palette = NULL, subset.view = NULL)
>>>>>>> 1fd9d66c2d3c79eb31f1e6bc797ddef71b6a0fae
>>>>>>> 95fc38b9d2eb90ae8981dffdf89f4a7112dfaa2d
}
\arguments{
\item{qData}{Numeric matrix}

\item{conds}{xxx}

\item{keyId}{xxx}

\item{legend}{A vector of the conditions (one condition per sample).}

\item{palette}{xxx}

\item{subset.view}{A vector of index indicating rows to highlight}
}
\value{
A violinplot
}
\description{
The parameter 'conds' is used to generate the color palette.
}
\examples{
library(SummarizedExperiment)
library(vioplot)
library(Features)
utils::data(Exp1_R25_prot, package='DAPARdata2')
qData <- assay(Exp1_R25_prot[[2]])
key <- rowData(Exp1_R25_prot[[2]])[[ metadata(Exp1_R25_prot)[['keyId']] ]]
conds <- colData(Exp1_R25_prot)[["Condition"]]
violinPlotD(qData, conds, key, subset.view = 1:5)

}
\author{
Samuel Wieczorek, Anais Courtier, Enora Fremy
}
